Constrained.API

Documentation

class LogicRequires s t dom rng ⇒ Logic s t dom rng | s → t where

Logical operations are one that support reasoning about how a function symbol relates to logical properties, that we call Specification's

Minimal complete definition

propagate

Methods

info ∷ t s dom rng → String

propagate ∷ Context s t dom rng hole → Specification rng → Specification hole

rewriteRules ∷ (TypeList dom, Typeable dom, HasSpec rng, All HasSpec dom) ⇒ t s dom rng → List Term dom → Evidence (AppRequires s t dom rng) → Maybe (Term rng)

mapTypeSpec ∷ ∀ a b. (dom ~ '[a], rng ~ b, HasSpec a, HasSpec b) ⇒ t s '[a] b → TypeSpec a → Specification b

saturate ∷ t s dom Bool → List Term dom → [Pred]

class Syntax t ⇒ Semantics (t ∷ Symbol → [Type] → Type → Type) where

Semantic operations are ones that give the function symbol, meaning as a function. I.e. how to apply the function to a list of arguments and return a value.

Methods

semantics ∷ ∀ s d r. t s d r → FunTy d r

class Syntax (t ∷ FSType) where

Syntactic operations are ones that have to do with the structure and appearence of the type.

Minimal complete definition

Nothing

Methods

isInFix ∷ ∀ s dom rng. t s dom rng → Bool

prettyWit ∷ ∀ s dom rng ann. (All HasSpec dom, HasSpec rng) ⇒ t s dom rng → List Term dom → Int → Maybe (Doc ann)

class (HasSpec a, NumLike a, Logic "addFn" IntW '[a, a] a) ⇒ Foldy a where

Minimal complete definition

genList, genSizedList, noNegativeValues

Methods

genList ∷ MonadGenError m ⇒ Specification a → Specification a → GenT m [a]

theAddFn ∷ IntW "addFn" '[a, a] a

theZero ∷ a

genSizedList ∷ MonadGenError m ⇒ Specification Integer → Specification a → Specification a → GenT m [a]

noNegativeValues ∷ Bool

data BaseW (sym ∷ Symbol) (dom ∷ [Type]) (rng ∷ Type) where

Constructors

InjLeftW ∷ ∀ a b. BaseW "leftFn" '[a] (Sum a b)

InjRightW ∷ ∀ a b. BaseW "rightFn" '[b] (Sum a b)

ProdW ∷ ∀ a b. BaseW "prod_" '[a, b] (Prod a b)

ProdFstW ∷ ∀ a b. BaseW "prodFst_" '[Prod a b] a

ProdSndW ∷ ∀ a b. BaseW "prodSnd_" '[Prod a b] b

EqualW ∷ ∀ a. Eq a ⇒ BaseW "==." '[a, a] Bool

ToGenericW ∷ GenericRequires a ⇒ BaseW "toGenericFn" '[a] (SimpleRep a)

FromGenericW ∷ GenericRequires a ⇒ BaseW "fromGenericFn" '[SimpleRep a] a

ElemW ∷ ∀ a. Eq a ⇒ BaseW "elem_" '[a, [a]] Bool

data BoolW (sym ∷ Symbol) (dom ∷ [Type]) (rng ∷ Type) where

Operations on Bool

Constructors

NotW ∷ BoolW "not_" '[Bool] Bool

OrW ∷ BoolW "or_" '[Bool, Bool] Bool

data NumOrdW (sym ∷ Symbol) (dom ∷ [Type]) (rng ∷ Type) where

Constructors

LessOrEqualW ∷ OrdLike a ⇒ NumOrdW "<=." '[a, a] Bool

LessW ∷ OrdLike a ⇒ NumOrdW "<." '[a, a] Bool

GreaterOrEqualW ∷ OrdLike a ⇒ NumOrdW ">=." '[a, a] Bool

GreaterW ∷ OrdLike a ⇒ NumOrdW ">." '[a, a] Bool

data IntW (s ∷ Symbol) (as ∷ [Type]) b where

Constructors

AddW ∷ NumLike a ⇒ IntW "addFn" '[a, a] a

NegateW ∷ NumLike a ⇒ IntW "negateFn" '[a] a

data SizeW (s ∷ Symbol) (dom ∷ [Type]) rng ∷ Type where

Constructors

SizeOfW ∷ ∀ n. Sized n ⇒ SizeW "sizeOf_" '[n] Integer

data FunW (sym ∷ Symbol) (dom ∷ [Type]) (rng ∷ Type) where

Constructors

IdW ∷ ∀ a. FunW "id_" '[a] a

ComposeW ∷ ∀ b s1 s2 t1 t2 a r. (Logic s1 t1 '[b] r, Logic s2 t2 '[a] b, HasSpec b) ⇒ t1 s1 '[b] r → t2 s2 '[a] b → FunW "composeFn" '[a] r

FlipW ∷ ∀ sym t a b r. Logic sym t '[a, b] r ⇒ t sym '[a, b] r → FunW "flip_" '[b, a] r

data ListW (s ∷ Symbol) (args ∷ [Type]) (res ∷ Type) where

Constructors

FoldMapW ∷ ∀ a b. (Foldy b, HasSpec a) ⇒ Fun '[a] b → ListW "foldMap_" '[[a]] b

SingletonListW ∷ ListW "singeltonList_" '[a] [a]

AppendW ∷ (Typeable a, Show a) ⇒ ListW "append_" '[[a], [a]] [a]

data MapW (sym ∷ Symbol) (dom ∷ [Type]) (rng ∷ Type) where

Constructors

DomW ∷ ∀ k v. Ord k ⇒ MapW "dom_" '[Map k v] (Set k)

RngW ∷ ∀ k v. Ord k ⇒ MapW "rng_" '[Map k v] [v]

LookupW ∷ ∀ k v. Ord k ⇒ MapW "lookup_" '[k, Map k v] (Maybe v)

data NumSpec n

Constructors

NumSpecInterval (Maybe n) (Maybe n)

class MaybeBounded a where

Minimal complete definition

Nothing

Methods

lowerBound ∷ Maybe a

default lowerBound ∷ Bounded a ⇒ Maybe a

upperBound ∷ Maybe a

default upperBound ∷ Bounded a ⇒ Maybe a

data NonEmpty a

Constructors

a :| [a]

data Specification a where

A `Specification a` denotes a set of as

Constructors

ExplainSpec ∷ [String] → Specification a → Specification a
Explain a Specification

MemberSpec
Elements of a known set

Fields

∷ NonEmpty a

It must be an element of this OrdSet (List). Try hard not to put duplicates in the List.

→ Specification a

ErrorSpec ∷ NonEmpty String → Specification a
The empty set

SuspendedSpec
The set described by some predicates over the bound variable.

Fields

∷ HasSpec a

⇒ Var a

This variable ranges over values denoted by the spec

→ Pred

And the variable is subject to these constraints

→ Specification a

TypeSpec
A type-specific spec

Fields

∷ HasSpec a

⇒ TypeSpec a

→ [a]

It can't be any of the elements of this set

→ Specification a

TrueSpec ∷ Specification a
Anything

data Term a where

Constructors

App ∷ ∀ sym t dom rng. AppRequires sym t dom rng ⇒ t sym dom rng → List Term dom → Term rng

Lit ∷ Literal a ⇒ a → Term a

V ∷ HasSpec a ⇒ Var a → Term a

data Fun dom rng where

Constructors

Fun ∷ ∀ s t dom rng. Logic s t dom rng ⇒ t s dom rng → Fun dom rng

name ∷ String → Term a → Term a

named ∷ String → Term a → Term a

Give a Term a nameHint, if its a Var, and doesn't already have one, otherwise return the Term unchanged.

data Pred where

Constructors

ElemPred ∷ ∀ a. HasSpec a ⇒ Bool → Term a → NonEmpty a → Pred

Monitor ∷ ((∀ a. Term a → a) → Property → Property) → Pred

And ∷ [Pred] → Pred

Exists

Fields

∷ ((∀ b. Term b → b) → GE a)

Constructive recovery function for checking existential quantification

→ Binder a

→ Pred

Subst ∷ HasSpec a ⇒ Var a → Term a → Pred → Pred

Let ∷ Term a → Binder a → Pred

Assert ∷ Term Bool → Pred

Reifies

Fields

∷ (HasSpec a, HasSpec b)

⇒ Term b

This depends on the a term

→ Term a

→ (a → b)

Recover a useable value from the a term.

→ Pred

DependsOn ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Pred

ForAll ∷ (Forallable t a, HasSpec t, HasSpec a) ⇒ Term t → Binder a → Pred

Case

Fields

∷ HasSpec (SumOver as)

⇒ Term (SumOver as)

→ List (Weighted Binder) as

Each branch of the type is bound with only one variable because as are types. Constructors with multiple arguments are encoded with ProdOver (c.f. Univ).

→ Pred

When ∷ HasSpec Bool ⇒ Term Bool → Pred → Pred

GenHint ∷ HasGenHint a ⇒ Hint a → Term a → Pred

TruePred ∷ Pred

FalsePred ∷ NonEmpty String → Pred

Explain ∷ NonEmpty String → Pred → Pred

type Literal a = (Typeable a, Eq a, Show a)

Properties needed by objects to appear in the system, if they have no semantic or logical requirements. Mostly used for Lit terms, which are self evaluating But such things also occurr in Contexts.

class (Typeable a, Eq a, Show a, Show (TypeSpec a), Typeable (TypeSpec a)) ⇒ HasSpec a where

Minimal complete definition

Nothing

Associated Types

type TypeSpec a

The `TypeSpec a` is the type-specific `Specification a`.

type TypeSpec a = TypeSpec (SimpleRep a)

type Prerequisites a ∷ Constraint

Prerequisites for the instance that are sometimes necessary when working with e.g. Specifications or functions in the universe.

type Prerequisites a = ()

Methods

emptySpec ∷ TypeSpec a

default emptySpec ∷ (HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a)) ⇒ TypeSpec a

combineSpec ∷ TypeSpec a → TypeSpec a → Specification a

default combineSpec ∷ (HasSimpleRep a, HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a)) ⇒ TypeSpec a → TypeSpec a → Specification a

genFromTypeSpec ∷ (HasCallStack, MonadGenError m) ⇒ TypeSpec a → GenT m a

Generate a value that satisfies the Specification. The key property for this generator is soundness: ∀ a ∈ genFromTypeSpec spec. a conformsTo spec

default genFromTypeSpec ∷ (HasSimpleRep a, HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a)) ⇒ (HasCallStack, MonadGenError m) ⇒ TypeSpec a → GenT m a

conformsTo ∷ HasCallStack ⇒ a → TypeSpec a → Bool

Check conformance to the spec.

default conformsTo ∷ (HasSimpleRep a, HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a)) ⇒ HasCallStack ⇒ a → TypeSpec a → Bool

shrinkWithTypeSpec ∷ TypeSpec a → a → [a]

Shrink an a with the aide of a Specification

default shrinkWithTypeSpec ∷ (HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a) ⇒ TypeSpec a → a → [a]

toPreds ∷ Term a → TypeSpec a → Pred

Convert a spec to predicates: The key property here is: ∀ a. a conformsTo spec == a conformsTo constrained (t -> toPreds t spec)

default toPreds ∷ (HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a) ⇒ Term a → TypeSpec a → Pred

cardinalTypeSpec ∷ TypeSpec a → Specification Integer

Compute an upper and lower bound on the number of solutions genFromTypeSpec might return

default cardinalTypeSpec ∷ (HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a)) ⇒ TypeSpec a → Specification Integer

cardinalTrueSpec ∷ Specification Integer

A bound on the number of solutions `genFromTypeSpec TrueSpec` can produce. For a type with finite elements, we can get a much more accurate answer than TrueSpec

typeSpecHasError ∷ TypeSpec a → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec a → BinaryShow

monadConformsTo ∷ a → TypeSpec a → Writer [String] Bool

typeSpecOpt ∷ TypeSpec a → [a] → Specification a

For some types (especially finite ones) there may be much better ways to construct a Specification than the default method of just adding a large bad list to TypSpec. This function allows each HasSpec instance to decide.

guardTypeSpec ∷ [String] → TypeSpec a → Specification a

This can be used to detect self inconsistencies in a (TypeSpec t) Note this is similar to typeSpecHasError, and the default value for typeSpecHasError is written in terms of guardTypeSpec Both typeSpecHasError and guardTypeSpec can be set individually.

prerequisites ∷ Evidence (Prerequisites a)

Materialize the Prerequisites dictionary. It should not be necessary to implement this function manually.

default prerequisites ∷ Prerequisites a ⇒ Evidence (Prerequisites a)

class Typeable (SimpleRep a) ⇒ HasSimpleRep a where

Minimal complete definition

Nothing

Associated Types

type SimpleRep a

type SimpleRep a = SOP (TheSop a)

type TheSop a ∷ [Type]

type TheSop a = SOPOf (Rep a)

Methods

toSimpleRep ∷ a → SimpleRep a

default toSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ a → SimpleRep a

fromSimpleRep ∷ SimpleRep a → a

default fromSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ SimpleRep a → a

class (Ord a, HasSpec a) ⇒ OrdLike a where

Minimal complete definition

Nothing

Methods

leqSpec ∷ a → Specification a

default leqSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

ltSpec ∷ a → Specification a

default ltSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

geqSpec ∷ a → Specification a

default geqSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

gtSpec ∷ a → Specification a

default gtSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

conformsToSpecE ∷ ∀ a. HasSpec a ⇒ a → Specification a → NonEmpty String → Maybe (NonEmpty String)

conformsToSpec with explanation. Nothing if (conformsToSpec a spec), but (Just explanations) if not(conformsToSpec a spec).

conformsToSpec ∷ HasSpec a ⇒ a → Specification a → Bool

satisfies ∷ ∀ a. HasSpec a ⇒ Term a → Specification a → Pred

genFromSpecT ∷ ∀ a m. (HasCallStack, HasSpec a, MonadGenError m) ⇒ Specification a → GenT m a

Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

genFromSpec ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Specification a → Gen a

A version of genFromSpecT that simply errors if the generator fails

debugSpec ∷ ∀ a. HasSpec a ⇒ Specification a → IO ()

A version of genFromSpecT that runs in the IO monad. Good for debugging.

genFromSpecWithSeed ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Int → Int → Specification a → a

A version of genFromSpecT that takes a seed and a size and gives you a result

simplifySpec ∷ HasSpec a ⇒ Specification a → Specification a

cardinality ∷ ∀ a. (Number Integer, HasSpec a) ⇒ Specification a → Specification Integer

Put some (admittedly loose bounds) on the number of solutions that genFromTypeSpec might return. For lots of types, there is no way to be very accurate. Here we lift the HasSpec methods cardinalTrueSpec and cardinalTypeSpec from (TypeSpec Integer) to (Specification Integer)

ifElse ∷ (IsPred p, IsPred q) ⇒ Term Bool → p → q → Pred

If the `Specification Bool` doesn't constrain the boolean you will get a TrueSpec out.

whenTrue ∷ ∀ p. IsPred p ⇒ Term Bool → p → Pred

simplifyTerm ∷ ∀ a. Term a → Term a

Simplify a Term, if the Term is an App, apply the rewrite rules chosen by the (Logic sym t bs a) instance attached to the function witness f

constrained ∷ ∀ a p. (IsPred p, HasSpec a) ⇒ (Term a → p) → Specification a

assertExplain ∷ IsPred p ⇒ NonEmpty String → p → Pred

assert ∷ IsPred p ⇒ p → Pred

forAll ∷ (Forallable t a, HasSpec t, HasSpec a, IsPred p) ⇒ Term t → (Term a → p) → Pred

exists ∷ ∀ a p. (HasSpec a, IsPred p) ⇒ ((∀ b. Term b → b) → GE a) → (Term a → p) → Pred

unsafeExists ∷ ∀ a p. (HasSpec a, IsPred p) ⇒ (Term a → p) → Pred

letBind ∷ (HasSpec a, IsPred p) ⇒ Term a → (Term a → p) → Pred

reify ∷ (HasSpec a, HasSpec b, IsPred p) ⇒ Term a → (a → b) → (Term b → p) → Pred

assertReified ∷ (HasSpec Bool, HasSpec a) ⇒ Term a → (a → Bool) → Pred

requires (HasSpec Bool)

explanation ∷ NonEmpty String → Pred → Pred

Wrap an Explain around a Pred, unless there is a simpler form.

monitor ∷ ((∀ a. Term a → a) → Property → Property) → Pred

Add QuickCheck monitoring (e.g. collect or counterexample) to a predicate. To use the monitoring in a property call monitorSpec on the Specification containing the monitoring and a value generated from the specification.

reifies ∷ (HasSpec a, HasSpec b) ⇒ Term b → Term a → (a → b) → Pred

dependsOn ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Pred

lit ∷ HasSpec a ⇒ a → Term a

genHint ∷ ∀ t. HasGenHint t ⇒ Hint t → Term t → Pred

giveHint ∷ HasGenHint a ⇒ Hint a → Specification a

(<.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(<=.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(>=.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(>.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(==.) ∷ HasSpec a ⇒ Term a → Term a → Term Bool infix 4

(/=.) ∷ HasSpec a ⇒ Term a → Term a → Term Bool infix 4

not_ ∷ Term Bool → Term Bool

or_ ∷ Term Bool → Term Bool → Term Bool

(||.) ∷ Term Bool → Term Bool → Term Bool infixr 2

toGeneric_ ∷ ∀ a. GenericRequires a ⇒ Term a → Term (SimpleRep a)

fromGeneric_ ∷ ∀ a. (GenericRequires a, AppRequires "fromGenericFn" BaseW '[SimpleRep a] a) ⇒ Term (SimpleRep a) → Term a

(+.) ∷ NumLike a ⇒ Term a → Term a → Term a infix 4

(-.) ∷ Numeric n ⇒ Term n → Term n → Term n infix 4

negate_ ∷ NumLike a ⇒ Term a → Term a

addFn ∷ ∀ a. NumLike a ⇒ Term a → Term a → Term a

negateFn ∷ ∀ a. NumLike a ⇒ Term a → Term a

type Numeric a = (HasSpec a, Ord a, Num a, TypeSpec a ~ NumSpec a, MaybeBounded a)

pair_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term a → Term b → Term (a, b)

fst_ ∷ (HasSpec x, HasSpec y) ⇒ Term (x, y) → Term x

snd_ ∷ (HasSpec x, HasSpec y) ⇒ Term (x, y) → Term y

prodSnd_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term b

prodFst_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term a

prod_ ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Term (Prod a b)

type IsNormalType a = (Cases a ~ '[a], Args a ~ '[a], IsProd a, CountCases a ~ 1)

leftFn ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term a → Term (Sum a b)

rightFn ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term b → Term (Sum a b)

left_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term a → Term (Either a b)

right_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term b → Term (Either a b)

cJust_ ∷ (HasSpec a, IsNormalType a) ⇒ Term a → Term (Maybe a)

cNothing_ ∷ (HasSpec a, IsNormalType a) ⇒ Term (Maybe a)

caseOn ∷ ∀ a. (HasSpec a, HasSpec (SimpleRep a), HasSimpleRep a, TypeSpec a ~ TypeSpec (SimpleRep a), SimpleRep a ~ SumOver (Cases (SimpleRep a)), TypeList (Cases (SimpleRep a))) ⇒ Term a → FunTy (MapList (Weighted Binder) (Cases (SimpleRep a))) Pred

branch ∷ ∀ p a. (HasSpec a, All HasSpec (Args a), IsPred p, IsProd a) ⇒ FunTy (MapList Term (Args a)) p → Weighted Binder a

branchW ∷ ∀ p a. (HasSpec a, All HasSpec (Args a), IsPred p, IsProd a) ⇒ Int → FunTy (MapList Term (Args a)) p → Weighted Binder a

forAll' ∷ ∀ t a p. (Forallable t a, Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec a ~ TypeSpec (SimpleRep a), HasSpec t, HasSpec (SimpleRep a), HasSimpleRep a, All HasSpec (Args (SimpleRep a)), IsPred p, IsProd (SimpleRep a), HasSpec a) ⇒ Term t → FunTy (MapList Term (Args (SimpleRep a))) p → Pred

Like forAll but pattern matches on the `Term a`

constrained' ∷ ∀ a p. (Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec a ~ TypeSpec (SimpleRep a), HasSpec (SimpleRep a), HasSimpleRep a, All HasSpec (Args (SimpleRep a)), IsProd (SimpleRep a), HasSpec a, IsPred p) ⇒ FunTy (MapList Term (Args (SimpleRep a))) p → Specification a

Like constrained but pattern matches on the bound `Term a`

reify' ∷ ∀ a b p. (Cases (SimpleRep b) ~ '[SimpleRep b], TypeSpec b ~ TypeSpec (SimpleRep b), HasSpec (SimpleRep b), HasSimpleRep b, All HasSpec (Args (SimpleRep b)), IsProd (SimpleRep b), HasSpec a, HasSpec b, IsPred p) ⇒ Term a → (a → b) → FunTy (MapList Term (Args (SimpleRep b))) p → Pred

Like reify but pattern matches on the bound `Term b`

con ∷ ∀ c a r. (SimpleRep a ~ SOP (TheSop a), TypeSpec a ~ TypeSpec (SOP (TheSop a)), TypeList (ConstrOf c (TheSop a)), HasSpec a, HasSimpleRep a, r ~ FunTy (MapList Term (ConstrOf c (TheSop a))) (Term a), ResultType r ~ Term a, SOPTerm c (TheSop a), ConstrTerm (ConstrOf c (TheSop a))) ⇒ r

onCon ∷ ∀ c a p. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, HasSpec a, HasSpec (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All HasSpec (Cases (SOP (TheSop a))), HasSpec (ProdOver (ConstrOf c (TheSop a))), IsPred p, Args (ProdOver (ConstrOf c (TheSop a))) ~ ConstrOf c (TheSop a), All HasSpec (ConstrOf c (TheSop a)), IsProd (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term a → FunTy (MapList Term (ConstrOf c (TheSop a))) p → Pred

isCon ∷ ∀ c a. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, HasSpec a, HasSpec (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All HasSpec (Cases (SOP (TheSop a))), HasSpec (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term a → Pred

sel ∷ ∀ n a c as. (SimpleRep a ~ ProdOver as, TheSop a ~ '[c ::: as], TypeSpec a ~ TypeSpec (ProdOver as), Select n as, HasSpec a, HasSpec (ProdOver as), HasSimpleRep a) ⇒ Term a → Term (At n as)

match ∷ ∀ p a. (HasSpec a, IsProductType a, IsPred p) ⇒ Term a → FunTy (MapList Term (ProductAsList a)) p → Pred

onJust ∷ ∀ a p. (HasSpec a, IsNormalType a, IsPred p) ⇒ Term (Maybe a) → (Term a → p) → Pred

isJust ∷ ∀ a. (HasSpec a, IsNormalType a) ⇒ Term (Maybe a) → Pred

chooseSpec ∷ HasSpec a ⇒ (Int, Specification a) → (Int, Specification a) → Specification a

ChooseSpec is one of the ways we can Or two Specs together This works for any kind of type that has a HasSpec instance. If your type is a Sum type. One can use CaseOn which is much easier.

equalSpec ∷ a → Specification a

notEqualSpec ∷ ∀ a. HasSpec a ⇒ a → Specification a

notMemberSpec ∷ ∀ a f. (HasSpec a, Foldable f) ⇒ f a → Specification a

id_ ∷ ∀ a. HasSpec a ⇒ Term a → Term a

flip_ ∷ ∀ (t ∷ FSType) (sym ∷ Symbol) a b r. (HasSpec b, HasSpec a, HasSpec r, ∀ sym1 t1. Logic sym1 t1 '[a, b] r) ⇒ t sym '[a, b] r → Term b → Term a → Term r

compose_ ∷ ∀ b s1 s2 t1 t2 a r. (AppRequires s1 t1 '[b] r, AppRequires s2 t2 '[a] b) ⇒ t1 s1 '[b] r → t2 s2 '[a] b → Term a → Term r

foldMap_ ∷ ∀ a b. (Sized [a], Foldy b, HasSpec a) ⇒ (Term a → Term b) → Term [a] → Term b

sum_ ∷ Foldy a ⇒ Term [a] → Term a

elem_ ∷ (Sized [a], HasSpec a) ⇒ Term a → Term [a] → Term Bool infix 4

singletonList_ ∷ (Sized [a], HasSpec a) ⇒ Term a → Term [a]

append_ ∷ (Sized [a], HasSpec a) ⇒ Term [a] → Term [a] → Term [a]

(++.) ∷ HasSpec a ⇒ Term [a] → Term [a] → Term [a] infixr 5

sizeOf ∷ Sized t ⇒ t → Integer

sizeOf_ ∷ (HasSpec t, Sized t) ⇒ Term t → Term Integer

null_ ∷ (HasSpec a, Sized a) ⇒ Term a → Term Bool

hasSize ∷ (HasSpec t, Sized t) ⇒ SizeSpec → Specification t

rangeSize ∷ Integer → Integer → SizeSpec

length_ ∷ HasSpec a ⇒ Term [a] → Term Integer

genFromSizeSpec ∷ MonadGenError m ⇒ Specification Integer → GenT m Integer

Because Sizes should always be >= 0, We provide this alternate generator that can be used to replace (genFromSpecT @Integer), to ensure this important property

between ∷ (HasSpec a, TypeSpec a ~ NumSpec a) ⇒ a → a → Specification a

maxSpec ∷ Specification Integer → Specification Integer

The widest interval whose largest element is admitted by the original spec

data SetW (s ∷ Symbol) (d ∷ [Type]) (r ∷ Type) where

Constructors

SingletonW ∷ Ord a ⇒ SetW "singleton_" '[a] (Set a)

UnionW ∷ (Literal a, Ord a) ⇒ SetW "union_" '[Set a, Set a] (Set a)

SubsetW ∷ (Literal a, Ord a) ⇒ SetW "subset_" '[Set a, Set a] Bool

MemberW ∷ (Literal a, Ord a) ⇒ SetW "member_" '[a, Set a] Bool

DisjointW ∷ (Literal a, Ord a) ⇒ SetW "disjoint_" '[Set a, Set a] Bool

FromListW ∷ ∀ a. (HasSpec a, Ord a) ⇒ SetW "fromList_" '[[a]] (Set a)

data SetSpec a

Constructors

SetSpec (Set a) (Specification a) (Specification Integer)

singleton_ ∷ (Ord a, HasSpec a) ⇒ Term a → Term (Set a)

member_ ∷ (Ord a, HasSpec a) ⇒ Term a → Term (Set a) → Term Bool

union_ ∷ (Ord a, HasSpec a) ⇒ Term (Set a) → Term (Set a) → Term (Set a)

subset_ ∷ (Ord a, HasSpec a) ⇒ Term (Set a) → Term (Set a) → Term Bool

disjoint_ ∷ (Ord a, HasSpec a) ⇒ Term (Set a) → Term (Set a) → Term Bool

fromList_ ∷ ∀ a. (Ord a, HasSpec a) ⇒ Term [a] → Term (Set a)

pattern Equal ∷ ∀ b. () ⇒ ∀ a. (b ~ Bool, Eq a, HasSpec a) ⇒ Term a → Term a → Term b

pattern Elem ∷ ∀ b. () ⇒ ∀ a. (b ~ Bool, Eq a, HasSpec a) ⇒ Term a → Term [a] → Term b

pattern ToGeneric ∷ ∀ rng. () ⇒ ∀ a. (rng ~ SimpleRep a, GenericRequires a, HasSpec a, AppRequires "toGenericFn" BaseW '[a] rng) ⇒ Term a → Term rng

pattern FromGeneric ∷ ∀ rng. () ⇒ ∀ a. (rng ~ a, GenericRequires a, HasSpec a, AppRequires "fromGenericFn" BaseW '[SimpleRep a] rng) ⇒ Term (SimpleRep a) → Term rng

pattern InjLeft ∷ ∀ c. () ⇒ ∀ a b. (c ~ Sum a b, AppRequires "leftFn" BaseW '[a] c) ⇒ Term a → Term c

pattern InjRight ∷ ∀ c. () ⇒ ∀ a b. (c ~ Sum a b, AppRequires "rightFn" BaseW '[b] c) ⇒ Term b → Term c

pattern ProdFst ∷ ∀ c. () ⇒ ∀ a b. (c ~ a, Typeable (Prod a b), HasSpec a, AppRequires "prodFst_" BaseW '[Prod a b] a) ⇒ Term (Prod a b) → Term c

pattern ProdSnd ∷ ∀ c. () ⇒ ∀ a b. (c ~ b, Typeable (Prod a b), HasSpec b, AppRequires "prodSnd_" BaseW '[Prod a b] b) ⇒ Term (Prod a b) → Term c

pattern Product ∷ ∀ c. () ⇒ ∀ a b. (c ~ Prod a b, AppRequires "prod_" BaseW '[a, b] c) ⇒ Term a → Term b → Term c

printPlan ∷ HasSpec a ⇒ Specification a → IO ()

class (Num a, HasSpec a) ⇒ NumLike a

data PairSpec a b

Constructors

Cartesian (Specification a) (Specification b)

data MapSpec k v

Constructors

MapSpec

Fields

mapSpecHint ∷ Maybe Integer

mapSpecMustKeys ∷ Set k

mapSpecMustValues ∷ [v]

mapSpecSize ∷ Specification Integer

mapSpecElem ∷ Specification (k, v)

mapSpecFold ∷ FoldSpec v

dom_ ∷ (HasSpec (Map k v), HasSpec v, HasSpec k, Ord k, IsNormalType k, IsNormalType v) ⇒ Term (Map k v) → Term (Set k)

rng_ ∷ (HasSpec k, HasSpec v, Ord k, IsNormalType k, IsNormalType v) ⇒ Term (Map k v) → Term [v]

lookup_ ∷ (HasSpec k, HasSpec v, Ord k, IsNormalType k, IsNormalType v) ⇒ Term k → Term (Map k v) → Term (Maybe v)

fstSpec ∷ ∀ k v. (HasSpec k, HasSpec v) ⇒ Specification (k, v) → Specification k

sndSpec ∷ ∀ k v. (HasSpec k, HasSpec v) ⇒ Specification (k, v) → Specification v

var ∷ QuasiQuoter

data Prod a b

Constructors

Prod

Fields

prodFst ∷ a

prodSnd ∷ b