Constrained.API.Extend

Abstract syntax

data SpecificationD deps a where

A `Specification a` denotes a set of as

Constructors

ExplainSpec ∷ [String] → SpecificationD deps a → SpecificationD deps a
Explain a Specification

MemberSpec
Elements of a known set

Fields

∷ NonEmpty a

It must be an element of this OrdSet (List). Try hard not to put duplicates in the List.

→ SpecificationD deps a

ErrorSpec ∷ NonEmpty String → SpecificationD deps a
The empty set

SuspendedSpec
The set described by some predicates over the bound variable.

Fields

∷ HasSpecD deps a

⇒ Var a

This variable ranges over values denoted by the spec

→ PredD deps

And the variable is subject to these constraints

→ SpecificationD deps a

TypeSpecD
A type-specific spec

Fields

∷ HasSpecD deps a

⇒ TypeSpecD deps a

→ [a]

It can't be any of the elements of this set

→ SpecificationD deps a

TrueSpec ∷ SpecificationD deps a
Anything

pattern TypeSpec ∷ () ⇒ HasSpec a ⇒ TypeSpec a → [a] → Specification a

data PredD deps where

Constructors

ElemPred ∷ (HasSpecD deps a, Show a) ⇒ Bool → TermD deps a → NonEmpty a → PredD deps

Monitor ∷ ((∀ a. TermD deps a → a) → Property → Property) → PredD deps

And ∷ [PredD deps] → PredD deps

Exists

Fields

∷ ((∀ b. TermD deps b → b) → GE a)

Constructive recovery function for checking existential quantification

→ BinderD deps a

→ PredD deps

Subst ∷ (HasSpecD deps a, Show a) ⇒ Var a → TermD deps a → PredD deps → PredD deps

Let ∷ TermD deps a → BinderD deps a → PredD deps

Assert ∷ TermD deps Bool → PredD deps

Reifies

Fields

∷ (HasSpecD deps a, HasSpecD deps b, Show a, Show b)

⇒ TermD deps b

This depends on the a term

→ TermD deps a

→ (a → b)

Recover a useable value from the a term.

→ PredD deps

DependsOn ∷ (HasSpecD deps a, HasSpecD deps b, Show a, Show b) ⇒ TermD deps a → TermD deps b → PredD deps

ForAll ∷ (ForallableD deps t e, HasSpecD deps t, HasSpecD deps e, Show t, Show e) ⇒ TermD deps t → BinderD deps e → PredD deps

Case

Fields

∷ (HasSpecD deps (SumOver as), Show (SumOver as))

⇒ TermD deps (SumOver as)

→ List (Weighted (BinderD deps)) as

Each branch of the type is bound with only one variable because as are types. Constructors with multiple arguments are encoded with ProdOver (c.f. Univ).

→ PredD deps

When ∷ TermD deps Bool → PredD deps → PredD deps

GenHint ∷ (HasGenHintD deps a, Show a, Show (HintD deps a)) ⇒ HintD deps a → TermD deps a → PredD deps

TruePred ∷ PredD deps

FalsePred ∷ NonEmpty String → PredD deps

Explain ∷ NonEmpty String → PredD deps → PredD deps

data TermD deps a where

Constructors

App ∷ AppRequiresD deps t dom rng ⇒ t dom rng → List (TermD deps) dom → TermD deps rng

Lit ∷ (Typeable a, Eq a, Show a) ⇒ a → TermD deps a

V ∷ (HasSpecD deps a, Typeable a) ⇒ Var a → TermD deps a

data BinderD deps a where

Constructors

(:->) ∷ (HasSpecD deps a, Show a) ⇒ Var a → PredD deps → BinderD deps a

Implementing new functions

appTerm ∷ ∀ t ds r. AppRequires t ds r ⇒ t ds r → FunTy (MapList Term ds) (Term r)

class Semantics (t ∷ [Type] → Type → Type) where

Semantic operations are ones that give the function symbol, meaning as a function. I.e. how to apply the function to a list of arguments and return a value.

Methods

semantics ∷ t d r → FunTy d r

class Syntax (t ∷ [Type] → Type → Type) where

Syntactic operations are ones that have to do with the structure and appearence of the type. See DependencyInjection to better understand deps - it's a pointer to postpone having to define HasSpec and friends here.

Minimal complete definition

Nothing

Methods

isInfix ∷ t dom rng → Bool

prettySymbol ∷ ∀ deps dom rng ann. t dom rng → List (TermD deps) dom → Int → Maybe (Doc ann)

The Logic instance

class (Typeable t, Semantics t, Syntax t) ⇒ Logic t where

A First-order typed logic has 4 components 1. Terms (Variables (x), Constants (5), and Applications (F x 5) Applications, apply a function symbol to a list of arguments: (FunctionSymbol term1 .. termN) 2. Predicates (Ordered, Odd, ...) 3. Connectives (And, Or, Not, =>, ...) 4. Quantifiers (Forall, Exists)

The Syntax, Semantics, and Logic classes implement new function symbols in the first order logic. Note that a function symbol is first order data, that uniquely identifies a higher order function. The three classes supply varying levels of functionality, relating to the Syntax, Semantics, and Logical operations of the function symbol.

Logical operations are one that support reasoning about how a function symbol relates to logical properties, that we call Specification's

Minimal complete definition

propagate | propagateTypeSpec, propagateMemberSpec

Methods

propagateTypeSpec ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → TypeSpec b → [b] → Specification a

propagateMemberSpec ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → NonEmpty b → Specification a

propagate ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → Specification b → Specification a

rewriteRules ∷ (TypeList dom, Typeable dom, HasSpec rng, All HasSpec dom) ⇒ t dom rng → List Term dom → Evidence (AppRequires t dom rng) → Maybe (Term rng)

mapTypeSpec ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ t '[a] b → TypeSpec a → Specification b

saturate ∷ t dom Bool → List Term dom → [Pred]

data HOLE a b where

This is used together with ListCtx to form just the arguments to `f vs Ctx vs'` - replacing Ctx with HOLE, to get a `ListCtx Value as (HOLE a)` which then can be used as an input to propagate.

Constructors

HOLE ∷ HOLE a a

pattern Unary ∷ HOLE a' a → ListCtx f '[a] (HOLE a')

A Convenient pattern for singleton contexts

pattern (:<:) ∷ (Typeable b, Show b) ⇒ HOLE c a → b → ListCtx Value '[a, b] (HOLE c)

Convenient patterns for binary contexts (the arrow :<: points towards the hole)

pattern (:>:) ∷ (Typeable a, Show a) ⇒ a → HOLE c b → ListCtx Value '[a, b] (HOLE c)

Convenient patterns for binary contexts (the arrow :>: points towards the hole)

Built-in HasSpecs

data PairSpec a b

Constructors

Cartesian (Specification a) (Specification b)

data MapSpec k v

Constructors

MapSpec

Fields

mapSpecHint ∷ Maybe Integer

mapSpecMustKeys ∷ Set k

mapSpecMustValues ∷ [v]

mapSpecSize ∷ Specification Integer

mapSpecElem ∷ Specification (k, v)

mapSpecFold ∷ FoldSpec v

data SetSpec a

Constructors

SetSpec (Set a) (Specification a) (Specification Integer)

data NumSpec n

Constructors

NumSpecInterval (Maybe n) (Maybe n)

data TreeSpec a

Constructors

TreeSpec

Fields

roseTreeAvgLength ∷ Maybe Integer

roseTreeMaxSize ∷ Maybe Integer

roseTreeRootSpec ∷ Specification a

roseTreeCtxSpec ∷ Specification (a, [Tree a])

Re-export of API

module Constrained.API