Constrained.AbstractSyntax

Documentation

data TermD deps a where

Constructors

App ∷ AppRequiresD deps t dom rng ⇒ t dom rng → List (TermD deps) dom → TermD deps rng

Lit ∷ (Typeable a, Eq a, Show a) ⇒ a → TermD deps a

V ∷ (HasSpecD deps a, Typeable a) ⇒ Var a → TermD deps a

type AppRequiresD deps (t ∷ [Type] → Type → Type) dom rng = (LogicD deps t, Syntax t, Semantics t, TypeList dom, Eq (t dom rng), Show (t dom rng), Typeable t, All Typeable dom, Typeable dom, Typeable rng, All (HasSpecD deps) dom, All Show dom, HasSpecD deps rng, Show rng)

runTermE ∷ ∀ a deps. Env → TermD deps a → Either (NonEmpty String) a

runTerm ∷ MonadGenError m ⇒ Env → TermD deps a → m a

fastInequality ∷ TermD deps a → TermD deps b → Bool

Sound but not complete inequality on terms

class Syntax (t ∷ [Type] → Type → Type) where

Syntactic operations are ones that have to do with the structure and appearence of the type. See DependencyInjection to better understand deps - it's a pointer to postpone having to define HasSpec and friends here.

Minimal complete definition

Nothing

Methods

isInfix ∷ t dom rng → Bool

prettySymbol ∷ ∀ deps dom rng ann. t dom rng → List (TermD deps) dom → Int → Maybe (Doc ann)

data PredD deps where

Constructors

ElemPred ∷ (HasSpecD deps a, Show a) ⇒ Bool → TermD deps a → NonEmpty a → PredD deps

Monitor ∷ ((∀ a. TermD deps a → a) → Property → Property) → PredD deps

And ∷ [PredD deps] → PredD deps

Exists

Fields

∷ ((∀ b. TermD deps b → b) → GE a)

Constructive recovery function for checking existential quantification

→ BinderD deps a

→ PredD deps

Subst ∷ (HasSpecD deps a, Show a) ⇒ Var a → TermD deps a → PredD deps → PredD deps

Let ∷ TermD deps a → BinderD deps a → PredD deps

Assert ∷ TermD deps Bool → PredD deps

Reifies

Fields

∷ (HasSpecD deps a, HasSpecD deps b, Show a, Show b)

⇒ TermD deps b

This depends on the a term

→ TermD deps a

→ (a → b)

Recover a useable value from the a term.

→ PredD deps

DependsOn ∷ (HasSpecD deps a, HasSpecD deps b, Show a, Show b) ⇒ TermD deps a → TermD deps b → PredD deps

ForAll ∷ (ForallableD deps t e, HasSpecD deps t, HasSpecD deps e, Show t, Show e) ⇒ TermD deps t → BinderD deps e → PredD deps

Case

Fields

∷ (HasSpecD deps (SumOver as), Show (SumOver as))

⇒ TermD deps (SumOver as)

→ List (Weighted (BinderD deps)) as

Each branch of the type is bound with only one variable because as are types. Constructors with multiple arguments are encoded with ProdOver (c.f. Univ).

→ PredD deps

When ∷ TermD deps Bool → PredD deps → PredD deps

GenHint ∷ (HasGenHintD deps a, Show a, Show (HintD deps a)) ⇒ HintD deps a → TermD deps a → PredD deps

TruePred ∷ PredD deps

FalsePred ∷ NonEmpty String → PredD deps

Explain ∷ NonEmpty String → PredD deps → PredD deps

data BinderD deps a where

Constructors

(:->) ∷ (HasSpecD deps a, Show a) ⇒ Var a → PredD deps → BinderD deps a

data Weighted f a

Constructors

Weighted

Fields

weight ∷ Maybe Int

thing ∷ f a

mapWeighted ∷ (f a → g b) → Weighted f a → Weighted g b

traverseWeighted ∷ Applicative m ⇒ (f a → m (g a)) → Weighted f a → m (Weighted g a)

data SpecificationD deps a where

A `Specification a` denotes a set of as

Constructors

ExplainSpec ∷ [String] → SpecificationD deps a → SpecificationD deps a
Explain a Specification

MemberSpec
Elements of a known set

Fields

∷ NonEmpty a

It must be an element of this OrdSet (List). Try hard not to put duplicates in the List.

→ SpecificationD deps a

ErrorSpec ∷ NonEmpty String → SpecificationD deps a
The empty set

SuspendedSpec
The set described by some predicates over the bound variable.

Fields

∷ HasSpecD deps a

⇒ Var a

This variable ranges over values denoted by the spec

→ PredD deps

And the variable is subject to these constraints

→ SpecificationD deps a

TypeSpecD
A type-specific spec

Fields

∷ HasSpecD deps a

⇒ TypeSpecD deps a

→ [a]

It can't be any of the elements of this set

→ SpecificationD deps a

TrueSpec ∷ SpecificationD deps a
Anything