Constrained.Base

Description

This module contains the most basic parts the implementation. Essentially everything to define Specification, HasSpec, HasSimpleRep, Term, Pred, and the Syntax, Semantics, and Logic class. It also has a few HasSpec, HasSimpleRep, and Logic instances for basic types needed to define the default types and methods of HasSpec. It also supplies Eq, Pretty, and Show instances on the syntax (Term, Pred, Binder etc.) because many functions require these instances. It exports functions that define the user interface to the domain embedded language (constrained, forall, exists etc.). And, by design, nothing more.

Documentation

class Syntax (t ∷ [Type] → Type → Type) where

A First-order typed logic has 4 components 1. Terms (Variables (x), Constants (5), and Applications (F x 5) Applications, apply a function symbol to a list of arguments: (FunctionSymbol term1 .. termN) 2. Predicates (Ordered, Odd, ...) 3. Connectives (And, Or, Not, =>, ...) 4. Quantifiers (Forall, Exists)

The Syntax, Semantics, and Logic classes implement new function symbols in the first order logic. Note that a function symbol is first order data, that uniquely identifies a higher order function. The three classes supply varying levels of functionality, relating to the Syntax, Semantics, and Logical operations of the function symbol.

Syntactic operations are ones that have to do with the structure and appearence of the type.

Minimal complete definition

Nothing

Methods

inFix ∷ ∀ dom rng. t dom rng → Bool

prettyWit ∷ ∀ dom rng ann. (All HasSpec dom, HasSpec rng) ⇒ t dom rng → List Term dom → Int → Maybe (Doc ann)

class Syntax t ⇒ Semantics (t ∷ [Type] → Type → Type) where

Semantic operations are ones that give the function symbol, meaning as a function. I.e. how to apply the function to a list of arguments and return a value.

Methods

semantics ∷ ∀ d r. t d r → FunTy d r

type LogicRequires t = (Typeable t, Syntax t, Semantics t)

class LogicRequires t ⇒ Logic t where

Logical operations are one that support reasoning about how a function symbol relates to logical properties, that we call Specification's

Minimal complete definition

propagate | propagateTypeSpec, propagateMemberSpec

Methods

propagateTypeSpec ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → TypeSpec b → [b] → Specification a

propagateMemberSpec ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → NonEmpty b → Specification a

propagate ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → Specification b → Specification a

rewriteRules ∷ (TypeList dom, Typeable dom, HasSpec rng, All HasSpec dom) ⇒ t dom rng → List Term dom → Evidence (AppRequires t dom rng) → Maybe (Term rng)

mapTypeSpec ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ t '[a] b → TypeSpec a → Specification b

saturate ∷ t dom Bool → List Term dom → [Pred]

propagateSpec ∷ ∀ v a. HasSpec v ⇒ Specification a → Ctx v a → Specification v

ctxHasSpec ∷ Ctx v a → Evidence (HasSpec a)

data Ctx v a where

Contexts for Terms, basically a term with a _single_ HOLE instead of a variable. This is used to traverse the defining constraints for a variable and turn them into a spec. Each subterm `f vs Ctx vs'` for lists of values vs and vs` gets given to the propagateSpecFun for f as `(f vs HOLE vs')`.

Constructors

CtxHOLE ∷ HasSpec v ⇒ Ctx v v
A single hole of type v. Note ctxHOLE is a nullary constructor, where the a type index is the same as the v type index.

CtxApp ∷ (AppRequires fn as b, HasSpec b, TypeList as, Typeable as, All HasSpec as, Logic fn) ⇒ fn as b → ListCtx Value as (Ctx v) → Ctx v b
The application `f vs Ctx vs'`

data HOLE a b where

This is used together with ListCtx to form just the arguments to `f vs Ctx vs'` - replacing Ctx with HOLE, to get a `ListCtx Value as (HOLE a)` which then can be used as an input to propagate.

Constructors

HOLE ∷ HOLE a a

toCtx ∷ ∀ m v a. (Typeable v, Show v, MonadGenError m, HasCallStack) ⇒ Var v → Term a → m (Ctx v a)

toCtxList ∷ ∀ m v as. (Show v, Typeable v, MonadGenError m, HasCallStack) ⇒ Var v → List Term as → m (ListCtx Value as (Ctx v))

pattern Unary ∷ HOLE a' a → ListCtx f '[a] (HOLE a')

A Convenient pattern for singleton contexts

pattern (:<:) ∷ (Typeable b, Show b) ⇒ HOLE c a → b → ListCtx Value '[a, b] (HOLE c)

Convenient patterns for binary contexts (the arrow :<: points towards the hole)

pattern (:>:) ∷ (Typeable a, Show a) ⇒ a → HOLE c b → ListCtx Value '[a, b] (HOLE c)

Convenient patterns for binary contexts (the arrow :>: points towards the hole)

flipCtx ∷ (Typeable a, Show a, Typeable b, Show b) ⇒ ListCtx Value '[a, b] (HOLE c) → ListCtx Value '[b, a] (HOLE c)

fromListCtx ∷ All HasSpec as ⇒ ListCtx Value as (HOLE a) → Term a → List Term as

From a ListCtx, build a (List Term as), to which the function symbol can be applied.

sameFunSym ∷ ∀ (t1 ∷ [Type] → Type → Type) d1 r1 (t2 ∷ [Type] → Type → Type) d2 r2. (AppRequires t1 d1 r1, AppRequires t2 d2 r2) ⇒ t1 d1 r1 → t2 d2 r2 → Maybe (t1 d1 r1, t1 :~: t2, d1 :~: d2, r1 :~: r2)

getWitness ∷ ∀ t t' d r. (AppRequires t d r, Typeable t') ⇒ t d r → Maybe (t' d r)

Here we only care about the Type t and the Symbol s the dom, and the rng can be anything. Useful for defining patterns.

data Specification a where

A `Specification a` denotes a set of as

Constructors

ExplainSpec ∷ [String] → Specification a → Specification a
Explain a Specification

MemberSpec
Elements of a known set

Fields

∷ NonEmpty a

It must be an element of this OrdSet (List). Try hard not to put duplicates in the List.

→ Specification a

ErrorSpec ∷ NonEmpty String → Specification a
The empty set

SuspendedSpec
The set described by some predicates over the bound variable.

Fields

∷ HasSpec a

⇒ Var a

This variable ranges over values denoted by the spec

→ Pred

And the variable is subject to these constraints

→ Specification a

TypeSpec
A type-specific spec

Fields

∷ HasSpec a

⇒ TypeSpec a

→ [a]

It can't be any of the elements of this set

→ Specification a

TrueSpec ∷ Specification a
Anything

typeSpec ∷ HasSpec a ⇒ TypeSpec a → Specification a

type GenericallyInstantiated a = (AssertComputes (SimpleRep a) (((Text "Trying to use a generic instantiation of " :<>: ShowType a) :<>: Text ", likely in a HasSpec instance.") :$$: Text "However, the type has no definition of SimpleRep, likely because of a missing instance of HasSimpleRep."), HasSimpleRep a, HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a))

type TypeSpecEqShow a = (AssertComputes (TypeSpec a) ((Text "Can't compute " :<>: ShowType (TypeSpec a)) :$$: Text "Either because of a missing definition of TypeSpec or a missing instance of HasSimpleRep."), Show (TypeSpec a), Typeable (TypeSpec a))

class (Typeable a, Eq a, Show a, TypeSpecEqShow a) ⇒ HasSpec a where

Minimal complete definition

Nothing

Associated Types

type TypeSpec a

The `TypeSpec a` is the type-specific `Specification a`.

type TypeSpec a = TypeSpec (SimpleRep a)

type Prerequisites a ∷ Constraint

Prerequisites for the instance that are sometimes necessary when working with e.g. Specifications or functions in the universe.

type Prerequisites a = ()

Methods

emptySpec ∷ TypeSpec a

default emptySpec ∷ GenericallyInstantiated a ⇒ TypeSpec a

combineSpec ∷ TypeSpec a → TypeSpec a → Specification a

default combineSpec ∷ GenericallyInstantiated a ⇒ TypeSpec a → TypeSpec a → Specification a

genFromTypeSpec ∷ (HasCallStack, MonadGenError m) ⇒ TypeSpec a → GenT m a

Generate a value that satisfies the Specification. The key property for this generator is soundness: ∀ a ∈ genFromTypeSpec spec. a conformsTo spec

default genFromTypeSpec ∷ (GenericallyInstantiated a, HasCallStack, MonadGenError m) ⇒ TypeSpec a → GenT m a

conformsTo ∷ HasCallStack ⇒ a → TypeSpec a → Bool

Check conformance to the spec.

default conformsTo ∷ (GenericallyInstantiated a, HasCallStack) ⇒ a → TypeSpec a → Bool

shrinkWithTypeSpec ∷ TypeSpec a → a → [a]

Shrink an a with the aide of a Specification

default shrinkWithTypeSpec ∷ GenericallyInstantiated a ⇒ TypeSpec a → a → [a]

toPreds ∷ Term a → TypeSpec a → Pred

Convert a spec to predicates: The key property here is: ∀ a. a conformsTo spec == a conformsTo constrained (t -> toPreds t spec)

default toPreds ∷ GenericallyInstantiated a ⇒ Term a → TypeSpec a → Pred

cardinalTypeSpec ∷ TypeSpec a → Specification Integer

Compute an upper and lower bound on the number of solutions genFromTypeSpec might return

default cardinalTypeSpec ∷ GenericallyInstantiated a ⇒ TypeSpec a → Specification Integer

cardinalTrueSpec ∷ Specification Integer

A bound on the number of solutions `genFromTypeSpec TrueSpec` can produce. For a type with finite elements, we can get a much more accurate answer than TrueSpec

typeSpecHasError ∷ TypeSpec a → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec a → BinaryShow

monadConformsTo ∷ a → TypeSpec a → Writer [String] Bool

typeSpecOpt ∷ TypeSpec a → [a] → Specification a

For some types (especially finite ones) there may be much better ways to construct a Specification than the default method of just adding a large bad list to TypSpec. This function allows each HasSpec instance to decide.

guardTypeSpec ∷ [String] → TypeSpec a → Specification a

This can be used to detect self inconsistencies in a (TypeSpec t) Note this is similar to typeSpecHasError, and the default value for typeSpecHasError is written in terms of guardTypeSpec Both typeSpecHasError and guardTypeSpec can be set individually.

prerequisites ∷ Evidence (Prerequisites a)

Materialize the Prerequisites dictionary. It should not be necessary to implement this function manually.

default prerequisites ∷ Prerequisites a ⇒ Evidence (Prerequisites a)

type GenericRequires a = (HasSpec a, HasSimpleRep a, HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a))

data BaseW (dom ∷ [Type]) (rng ∷ Type) where

Constructors

ToGenericW ∷ GenericRequires a ⇒ BaseW '[a] (SimpleRep a)

FromGenericW ∷ GenericRequires a ⇒ BaseW '[SimpleRep a] a

toGeneric_ ∷ ∀ a. GenericRequires a ⇒ Term a → Term (SimpleRep a)

fromGeneric_ ∷ ∀ a. (GenericRequires a, AppRequires BaseW '[SimpleRep a] a) ⇒ Term (SimpleRep a) → Term a

fromSimpleRepSpec ∷ ∀ a. (HasSpec a, HasSimpleRep a, TypeSpec a ~ TypeSpec (SimpleRep a)) ⇒ Specification (SimpleRep a) → Specification a

toSimpleRepSpec ∷ ∀ a. (HasSpec (SimpleRep a), HasSimpleRep a, TypeSpec a ~ TypeSpec (SimpleRep a)) ⇒ Specification a → Specification (SimpleRep a)

data BinaryShow where

Constructors

BinaryShow ∷ ∀ a. String → [Doc a] → BinaryShow

NonBinary ∷ BinaryShow

type AppRequires t dom rng = (Logic t, TypeList dom, Eq (t dom rng), Show (t dom rng), Typeable dom, Typeable rng, All HasSpec dom, HasSpec rng)

What constraints does the Term constructor App require? (Logic sym t dom rng) supplies the Logic of propagating contexts (All HasSpec dom) the argument types are part of the system (HasSpec rng) the return type is part of the system.

data Term a where

Constructors

App ∷ ∀ t dom rng. AppRequires t dom rng ⇒ t dom rng → List Term dom → Term rng

Lit ∷ (Typeable a, Eq a, Show a) ⇒ a → Term a

V ∷ HasSpec a ⇒ Var a → Term a

sameTerms ∷ All HasSpec as ⇒ List Term as → List Term as → Bool

appSym ∷ ∀ t as b. AppRequires t as b ⇒ t as b → List Term as → Term b

Recall function symbols are objects that you can use to build applications They carry information about both its semantic and logical properties. Usually the Haskel name ends in '_', for example consider: not_, subset_ ,lookup_, toGeneric_ Infix function symbols names end in ., for example: ==. , <=. E.g appTerm ToGenericW :: Term a -> Term(SimpleRep a) (appTerm ToGenericW (lit True)) builds the Term (toGeneric_ True) Note the witness (ToGenericW ) must have a Logic instance like: instance Logic BaseW '[a] (SimpleRep a) where ... type of ToGenericW ^ arg types^ result type^ The Logic instance does not demand any of these things have any properties at all. It is here, where we actually build the App node, that we demand the properties App terms require. App :: AppRequires s t ds r => t s ds r -> List Term dom -> Term rng

appTerm ∷ ∀ t ds r. AppRequires t ds r ⇒ t ds r → FunTy (MapList Term ds) (Term r)

name ∷ String → Term a → Term a

named ∷ String → Term a → Term a

Give a Term a nameHint, if its a Var, and doesn't already have one, otherwise return the Term unchanged.

data Binder a where

Constructors

(:->) ∷ HasSpec a ⇒ Var a → Pred → Binder a

bind ∷ (HasSpec a, IsPred p) ⇒ (Term a → p) → Binder a

data Weighted f a

Constructors

Weighted

Fields

weight ∷ Maybe Int

thing ∷ f a

mapWeighted ∷ (f a → g b) → Weighted f a → Weighted g b

traverseWeighted ∷ Applicative m ⇒ (f a → m (g a)) → Weighted f a → m (Weighted g a)

data Pred where

Constructors

ElemPred ∷ ∀ a. HasSpec a ⇒ Bool → Term a → NonEmpty a → Pred

Monitor ∷ ((∀ a. Term a → a) → Property → Property) → Pred

And ∷ [Pred] → Pred

Exists

Fields

∷ ((∀ b. Term b → b) → GE a)

Constructive recovery function for checking existential quantification

→ Binder a

→ Pred

Subst ∷ HasSpec a ⇒ Var a → Term a → Pred → Pred

Let ∷ Term a → Binder a → Pred

Assert ∷ Term Bool → Pred

Reifies

Fields

∷ (HasSpec a, HasSpec b)

⇒ Term b

This depends on the a term

→ Term a

→ (a → b)

Recover a useable value from the a term.

→ Pred

DependsOn ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Pred

ForAll ∷ (Forallable t a, HasSpec t, HasSpec a) ⇒ Term t → Binder a → Pred

Case

Fields

∷ HasSpec (SumOver as)

⇒ Term (SumOver as)

→ List (Weighted Binder) as

Each branch of the type is bound with only one variable because as are types. Constructors with multiple arguments are encoded with ProdOver (c.f. Univ).

→ Pred

When ∷ HasSpec Bool ⇒ Term Bool → Pred → Pred

GenHint ∷ HasGenHint a ⇒ Hint a → Term a → Pred

TruePred ∷ Pred

FalsePred ∷ NonEmpty String → Pred

Explain ∷ NonEmpty String → Pred → Pred

class Forallable t e | t → e where

Minimal complete definition

Nothing

Methods

fromForAllSpec ∷ (HasSpec t, HasSpec e) ⇒ Specification e → Specification t

default fromForAllSpec ∷ (HasSpec t, HasSpec e, HasSimpleRep t, TypeSpec t ~ TypeSpec (SimpleRep t), Forallable (SimpleRep t) e, HasSpec (SimpleRep t)) ⇒ Specification e → Specification t

forAllToList ∷ t → [e]

default forAllToList ∷ (HasSimpleRep t, Forallable (SimpleRep t) e) ⇒ t → [e]

class (HasSpec a, Show (Hint a)) ⇒ HasGenHint a where

Hints are things that only affect generation, and not validation. For instance, parameters to control distribution of generated values.

Associated Types

type Hint a

Methods

giveHint ∷ Hint a → Specification a

class Show p ⇒ IsPred p where

Methods

toPred ∷ p → Pred

memberSpecList ∷ [a] → NonEmpty String → Specification a

return a MemberSpec or ans ErrorSpec depending on if xs the null list or not

explainSpec ∷ [String] → Specification a → Specification a

explainSpecOpt ∷ [String] → Specification a → Specification a

equalSpec ∷ a → Specification a

notEqualSpec ∷ ∀ a. HasSpec a ⇒ a → Specification a

notMemberSpec ∷ ∀ a f. (HasSpec a, Foldable f) ⇒ f a → Specification a

constrained ∷ ∀ a p. (IsPred p, HasSpec a) ⇒ (Term a → p) → Specification a

isErrorLike ∷ ∀ a. Specification a → Bool

errorLikeMessage ∷ ∀ a. Specification a → NonEmpty String

fromGESpec ∷ HasCallStack ⇒ GE (Specification a) → Specification a

addToErrorSpec ∷ NonEmpty String → Specification a → Specification a

Add the explanations, if it's an ErrorSpec, else drop them

data WithPrec a

Constructors

WithPrec Int a

parensIf ∷ Bool → Doc ann → Doc ann

prettyPrec ∷ Pretty (WithPrec a) ⇒ Int → a → Doc ann

ppList ∷ ∀ f as ann. All HasSpec as ⇒ (∀ a. HasSpec a ⇒ f a → Doc ann) → List f as → [Doc ann]

ppList_ ∷ ∀ f as ann. (∀ a. f a → Doc ann) → List f as → [Doc ann]

prettyType ∷ ∀ t x. Typeable t ⇒ Doc x

vsep' ∷ [Doc ann] → Doc ann

(/>) ∷ Doc ann → Doc ann → Doc ann infixl 5

short ∷ ∀ a x. (Show a, Typeable a) ⇒ [a] → Doc x

showType ∷ ∀ t. Typeable t ⇒ String

data Fun dom rng where

Constructors

Fun ∷ ∀ t dom rng. AppRequires t dom rng ⇒ t dom rng → Fun dom rng

extractf ∷ ∀ t d r. LogicRequires t ⇒ Fun d r → Maybe (t d r)

appFun ∷ Fun '[x] b → Term x → Term b

sameFun ∷ Fun d1 r1 → Fun d2 r2 → Bool

pattern FromGeneric ∷ ∀ rng. () ⇒ ∀ a. (rng ~ a, GenericRequires a, HasSpec a, AppRequires BaseW '[SimpleRep a] rng) ⇒ Term (SimpleRep a) → Term rng

pattern ToGeneric ∷ ∀ rng. () ⇒ ∀ a. (rng ~ SimpleRep a, GenericRequires a, HasSpec a, AppRequires BaseW '[a] rng) ⇒ Term a → Term rng

Orphan instances

HasSimpleRep ()

Instance details

Associated Types

type SimpleRep ()

type TheSop () ∷ [Type]

Methods

toSimpleRep ∷ () → SimpleRep ()

fromSimpleRep ∷ SimpleRep () → ()

Typeable c ⇒ Show (Evidence c)

Instance details

Methods

showsPrec ∷ Int → Evidence c → ShowS

show ∷ Evidence c → String

showList ∷ [Evidence c] → ShowS