Constrained.GenT

Documentation

data GE a

It distinguishes between two kinds of errors: FatalError and GenError and non-fatal errors.

Constructors

FatalError (NonEmpty (NonEmpty String))

GenError (NonEmpty (NonEmpty String))

Result a

data GenMode

Generation mode - how strict are we about requiring the generator to succeed. This is necessary because sometimes failing to find a value means there is an actual problem (a generator _should_ be satisfiable but for whatever buggy reason it isn't) and sometimes failing to find a value just means there are no values. The latter case is very relevant when you're generating e.g. lists or sets of values that can be empty.

Constructors

Loose

Strict

newtype GenT m a

Constructors

GenT

Fields

runGenT ∷ GenMode → [NonEmpty String] → Gen (m a)

class Monad m ⇒ MonadGenError m where

A class for different types of errors with a stack of explain calls to narrow down problems. The (NonEmpty String) means one cannot cause an Error without at least 1 string to explain it.

Methods

genError ∷ HasCallStack ⇒ NonEmpty String → m a

fatalError ∷ HasCallStack ⇒ NonEmpty String → m a

genErrors ∷ HasCallStack ⇒ NonEmpty (NonEmpty String) → m a

fatalErrors ∷ HasCallStack ⇒ NonEmpty (NonEmpty String) → m a

explain ∷ HasCallStack ⇒ NonEmpty String → m a → m a

genError1 ∷ MonadGenError m ⇒ String → m a

genError with one line of explanation

fatalError1 ∷ MonadGenError m ⇒ String → m a

fatalError with one line of explanation

explain1 ∷ MonadGenError m ⇒ String → m a → m a

explain with one line of explanation

add ∷ NonEmpty a → [NonEmpty a] → NonEmpty (NonEmpty a)

cat ∷ NonEmpty (NonEmpty a) → [NonEmpty a] → NonEmpty (NonEmpty a)

catMessages ∷ NonEmpty (NonEmpty String) → String

Sometimes we have a bunch of genError or fatalError messages we want to combine into one big message. This happens when we want to lift one of these into an input for error

catMessageList ∷ NonEmpty (NonEmpty String) → NonEmpty String

Turn each inner (NonEmpty String) into a String

catGEs ∷ ∀ m a. MonadGenError m ⇒ [GE a] → m [a]

fromGE ∷ HasCallStack ⇒ (NonEmpty (NonEmpty String) → a) → GE a → a

Given a function for handling GenError, and handling FatalError by using error Turn (GE a) into a

errorGE ∷ GE a → a

Turn (GE a) into a both GenError and FatalErrors are handled by using error

isOk ∷ GE a → Bool

runGE ∷ ∀ m r. MonadGenError m ⇒ GE r → m r

fromGEProp ∷ Testable p ⇒ GE p → Property

fromGEDiscard ∷ Testable p ⇒ GE p → Property

headGE ∷ Foldable t ⇒ t a → GE a

listFromGE ∷ GE [a] → [a]

Turn a `GE [a]` to `[a]`, genError goes to `[]` and fatalError to error.

strictGen ∷ GenT m a → Gen (m a)

genFromGenT ∷ GenT GE a → Gen a

resizeT ∷ (Int → Int) → GenT m a → GenT m a

pureGen ∷ Applicative m ⇒ Gen a → GenT m a

listOfT ∷ MonadGenError m ⇒ GenT GE a → GenT m [a]

listOfUntilLenT ∷ MonadGenError m ⇒ GenT GE a → Int → (Int → Bool) → GenT m [a]

Generate a list of elements of length at most goalLen, but accepting failure to get that many elements so long as validLen is true. TODO: possibly one could return "more, fewer, ok" in the validLen instead of Bool

vectorOfT ∷ MonadGenError m ⇒ Int → GenT GE a → GenT m [a]

suchThatT ∷ MonadGenError m ⇒ GenT m a → (a → Bool) → GenT m a infixl 2

suchThatWithTryT ∷ MonadGenError m ⇒ Int → GenT m a → (a → Bool) → GenT m a

scaleT ∷ (Int → Int) → GenT m a → GenT m a

getMode ∷ Applicative m ⇒ GenT m GenMode

getMessages ∷ Applicative m ⇒ GenT m [NonEmpty String]

withMode ∷ GenMode → GenT m a → GenT m a

oneofT ∷ MonadGenError m ⇒ [GenT GE a] → GenT m a

frequencyT ∷ MonadGenError m ⇒ [(Int, GenT GE a)] → GenT m a

chooseT ∷ (Random a, Ord a, Show a, MonadGenError m) ⇒ (a, a) → GenT m a

sizeT ∷ Monad m ⇒ GenT m Int

inspect ∷ ∀ m x. MonadGenError m ⇒ GenT GE x → GenT m (GE x)

Always succeeds, but returns the internal GE structure for analysis

tryGenT ∷ MonadGenError m ⇒ GenT GE a → GenT m (Maybe a)

Ignore all kinds of Errors, by squashing them into Nothing

catchGenT ∷ MonadGenError m ⇒ GenT GE a → GenT m (Either (NonEmpty (NonEmpty String)) a)

catchGen ∷ GenT GE a → Gen (Either (NonEmpty (NonEmpty String)) a)

Pass on the error messages of both kinds of Errors in the Gen (not the GenT) monad

firstGenT ∷ ∀ m a. MonadGenError m ⇒ [GenT GE a] → GenT m (Either [NonEmpty (NonEmpty String)] a)

Return the first successfull result from a list of computations, if they all fail return a list of the error messages from each one.

liftGen ∷ ∀ x. (∀ m. MonadGenError m ⇒ GenT m x) → GenT GE x

dropGen ∷ MonadGenError m ⇒ GenT GE a → GenT m a

frequency2 ∷ ∀ m a. MonadGenError m ⇒ (Int, GenT GE a) → (Int, GenT GE a) → GenT m a

Run one of the actions with frequency proportional to the count. If it fails, run the other.

push ∷ ∀ m a. MonadGenError m ⇒ [String] → m a → m a

like explain for GenT, but uses [String] rather than (NonEmpty String) if the list is null, it becomes the identity

pushGE ∷ ∀ a. [String] → GE a → GE a

like explain for GE, but uses [String] rather than (NonEmpty String) if the list is null, it becomes the identity