Constrained.Generation

Description

All the things that are necessary for generation and shrinking.

Documentation

genFromSpecT ∷ ∀ a m. (HasCallStack, HasSpec a, MonadGenError m) ⇒ Specification a → GenT m a

Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

simplifySpec ∷ HasSpec a ⇒ Specification a → Specification a

optimisePred ∷ Pred → Pred

aggressiveInlining ∷ Pred → Pred

substituteAndSimplifyTerm ∷ Subst → Term a → Term a

Apply a substitution and simplify the resulting term if the substitution changed the term.

simplifyTerm ∷ ∀ a. Term a → Term a

Simplify a Term, if the Term is an App, apply the rewrite rules chosen by the (Logic sym t bs a) instance attached to the function witness f

simplifyPred ∷ Pred → Pred

simplifyPreds ∷ [Pred] → [Pred]

simplifyBinder ∷ Binder a → Binder a

computeSpecSimplified ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Specification a)

Precondition: the Pred defines the `Var a` Runs in GE in order for us to have detailed context on failure.

computeSpec ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Specification a)

Precondition: the `Pred fn` defines the `Var a`. Runs in GE in order for us to have detailed context on failure.

computeSpecBinder ∷ Binder a → GE (Specification a)

computeSpecBinderSimplified ∷ Binder a → GE (Specification a)

caseSpec ∷ ∀ as. HasSpec (SumOver as) ⇒ Maybe String → List (Weighted Specification) as → Specification (SumOver as)

Turn a list of branches into a SumSpec. If all the branches fail return an ErrorSpec. Note the requirement of HasSpec(SumOver).

data SumSpec a b

The Specification for Sums.

Constructors

SumSpecRaw (Maybe String) (Maybe (Int, Int)) (Specification a) (Specification b)

pattern SumSpec ∷ Maybe (Int, Int) → Specification a → Specification b → SumSpec a b

sumType ∷ Maybe String → String

totalWeight ∷ List (Weighted f) as → Maybe Int

genFromSpec ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Specification a → Gen a

A version of genFromSpecT that simply errors if the generator fails

genFromSpecWithSeed ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Int → Int → Specification a → a

A version of genFromSpecT that takes a seed and a size and gives you a result

debugSpec ∷ ∀ a. HasSpec a ⇒ Specification a → IO ()

A version of genFromSpecT that runs in the IO monad. Good for debugging.

shrinkWithSpec ∷ ∀ a. HasSpec a ⇒ Specification a → a → [a]

shrinkFromPreds ∷ HasSpec a ⇒ Pred → Var a → a → [a]

shrinkEnvFromPlan ∷ Env → SolverPlan → [Env]

substStage ∷ Env → SolverStage → SolverStage

normalizeSolverStage ∷ SolverStage → SolverStage

fixupWithSpec ∷ ∀ a. HasSpec a ⇒ Specification a → a → Maybe a

computeHints ∷ [Pred] → Hints

prepareLinearization ∷ Pred → GE SolverPlan

Linearize a predicate, turning it into a list of variables to solve and their defining constraints such that each variable can be solved independently.

flattenPred ∷ Pred → [Pred]

Flatten nested Let, Exists, and And in a `Pred fn`. Let and Exists bound variables become free in the result.

linearize ∷ MonadGenError m ⇒ [Pred] → DependGraph → m [SolverStage]

mergeSolverStage ∷ SolverStage → [SolverStage] → [SolverStage]

Does nothing if the variable is not in the plan already.

prettyPlan ∷ HasSpec a ⇒ Specification a → Doc ann

printPlan ∷ HasSpec a ⇒ Specification a → IO ()

isEmptyPlan ∷ SolverPlan → Bool

stepPlan ∷ MonadGenError m ⇒ Env → SolverPlan → GenT m (Env, SolverPlan)

genFromPreds ∷ ∀ m. MonadGenError m ⇒ Env → Pred → GenT m Env

Generate a satisfying Env for a `p : Pred fn`. The Env contains values for all the free variables in `flattenPred p`.

backPropagation ∷ SolverPlan → SolverPlan

Push as much information we can backwards through the plan.

data EqW ∷ [Type] → Type → Type where

Constructors

EqualW ∷ (Eq a, HasSpec a) ⇒ EqW '[a, a] Bool

(==.) ∷ HasSpec a ⇒ Term a → Term a → Term Bool infix 4

pattern Equal ∷ ∀ b. () ⇒ ∀ a. (b ~ Bool, Eq a, HasSpec a) ⇒ Term a → Term a → Term b

whenTrue ∷ ∀ p. IsPred p ⇒ Term Bool → p → Pred

pinnedBy ∷ ∀ a. HasSpec a ⇒ Var a → Pred → Maybe (Term a)

Is the variable x pinned to some free term in p? (free term meaning that all the variables in the term are free in p).

TODO: complete this with more cases!

saturatePred ∷ Pred → [Pred]

guardSumSpec ∷ ∀ a b. (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ [String] → SumSpec a b → Specification (Sum a b)

combTypeName ∷ Maybe String → Maybe String → Maybe String

type family CountCases a where ...

Equations

CountCases (Sum a b) = 1 + CountCases b

CountCases _ = 1

countCases ∷ ∀ a. KnownNat (CountCases a) ⇒ Int

data SumW dom rng where

Constructors

InjLeftW ∷ SumW '[a] (Sum a b)

InjRightW ∷ SumW '[b] (Sum a b)

injLeft_ ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term a → Term (Sum a b)

injRight_ ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term b → Term (Sum a b)

pattern InjRight ∷ ∀ c. () ⇒ ∀ a b. (c ~ Sum a b, AppRequires SumW '[b] c) ⇒ Term b → Term c

pattern InjLeft ∷ ∀ c. () ⇒ ∀ a b. (c ~ Sum a b, AppRequires SumW '[a] c) ⇒ Term a → Term c

sumWeightL ∷ Maybe (Int, Int) → Doc a

sumWeightR ∷ Maybe (Int, Int) → Doc a

data BoolW (dom ∷ [Type]) (rng ∷ Type) where

Operations on Bool

Constructors

NotW ∷ BoolW '[Bool] Bool

OrW ∷ BoolW '[Bool, Bool] Bool

boolSem ∷ BoolW dom rng → FunTy dom rng

not_ ∷ Term Bool → Term Bool

okOr ∷ Bool → Bool → Specification Bool

We have something like (constant ||. HOLE) must evaluate to need. Return a (Specification Bool) for HOLE, that makes that True.

or_ ∷ Term Bool → Term Bool → Term Bool

data SolverStage where

Constructors

SolverStage

Fields

∷ HasSpec a

⇒ { stageVar ∷ Var a

, stagePreds ∷ [Pred]

, stageSpec ∷ Specification a

} → SolverStage

data SolverPlan

Constructors

SolverPlan

Fields

solverPlan ∷ [SolverStage]

solverDependencies ∷ Graph Name

isTrueSpec ∷ Specification a → Bool

prettyLinear ∷ [SolverStage] → Doc ann

Orphan instances

(HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ HasSpec (Sum a b)
The HasSpec Sum instance

Instance details

Associated Types

type TypeSpec (Sum a b)

type Prerequisites (Sum a b)

Methods

emptySpec ∷ TypeSpec (Sum a b)

combineSpec ∷ TypeSpec (Sum a b) → TypeSpec (Sum a b) → Specification (Sum a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Sum a b) → GenT m (Sum a b)

conformsTo ∷ Sum a b → TypeSpec (Sum a b) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Sum a b) → Sum a b → [Sum a b]

toPreds ∷ Term (Sum a b) → TypeSpec (Sum a b) → Pred

cardinalTypeSpec ∷ TypeSpec (Sum a b) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Sum a b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Sum a b) → BinaryShow

monadConformsTo ∷ Sum a b → TypeSpec (Sum a b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Sum a b) → [Sum a b] → Specification (Sum a b)

guardTypeSpec ∷ [String] → TypeSpec (Sum a b) → Specification (Sum a b)

prerequisites ∷ Evidence (Prerequisites (Sum a b))