Constrained.Generic

Description

How can we automatically inject normal Haskell types into the logic, using GHC.Generics

Documentation

data Prod a b

Constructors

Prod

Fields

prodFst ∷ a

prodSnd ∷ b

type family ProdOver as where ...

Equations

ProdOver '[] = ()

ProdOver '[a] = a

ProdOver (a : as) = Prod a (ProdOver as)

listToProd ∷ (ProdOver as → r) → List Identity as → r

prodToList ∷ ∀ as. TypeList as ⇒ ProdOver as → List Identity as

appendProd ∷ ∀ xs ys. (TypeList xs, TypeList ys) ⇒ ProdOver xs → ProdOver ys → ProdOver (Append xs ys)

splitProd ∷ ∀ xs ys. (TypeList xs, TypeList ys) ⇒ ProdOver (Append xs ys) → Prod (ProdOver xs) (ProdOver ys)

data Sum a b

Generic way to represent Sums

Constructors

SumLeft a

SumRight b

type family SumOver as where ...

Equations

SumOver '[a] = a

SumOver (a : as) = Sum a (SumOver as)

class Typeable (SimpleRep a) ⇒ HasSimpleRep a where

Minimal complete definition

Nothing

Associated Types

type SimpleRep a

type SimpleRep a = SOP (TheSop a)

type TheSop a ∷ [Type]

type TheSop a = SOPOf (Rep a)

Methods

toSimpleRep ∷ a → SimpleRep a

default toSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ a → SimpleRep a

fromSimpleRep ∷ SimpleRep a → a

default fromSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ SimpleRep a → a

type family SimplifyRep f where ...

Equations

SimplifyRep f = SOP (SOPOf f)

class SimpleGeneric rep where

Methods

toSimpleRep' ∷ rep p → SimplifyRep rep

fromSimpleRep' ∷ SimplifyRep rep → rep p

data (c ∷ Symbol) ::: (ts ∷ [Type])

A constructor name with the types of its arguments

type family SOPOf f where ...

Turn a Rep into a list that flattens the sum structre and gives the constructors names: Maybe Int -> '[Nothing ::: '[()], Just ::: '[Int]] Either Int Bool -> '[Left ::: '[Int], Right ::: '[Bool]] data Foo = Foo Int Bool | Bar Double -> '[Foo ::: '[Int, Bool], Bar ::: '[Double]]

Equations

SOPOf (D1 _ f) = SOPOf f

SOPOf (f :+: g) = Append (SOPOf f) (SOPOf g)

SOPOf (C1 ('MetaCons constr _ _) f) = '[constr ::: Constr f]

type family Constr f where ...

Flatten a single constructor

Equations

Constr (S1 _ f) = Constr f

Constr (K1 _ k) = '[k]

Constr U1 = '[()]

Constr (f :*: g) = Append (Constr f) (Constr g)

type family SOP constrs where ...

Turn a list from SOPOf into a Sum over Prod representation.

Equations

SOP '[c ::: prod] = ProdOver prod

SOP ((c ::: prod) : constrs) = Sum (ProdOver prod) (SOP constrs)

type family ConstrOf c sop where ...

Equations

ConstrOf c ((c ::: constr) : sop) = constr

ConstrOf c (_ : sop) = ConstrOf c sop

class Inject c constrs r where

Methods

inject' ∷ (SOP constrs → r) → FunTy (ConstrOf c constrs) r

inject ∷ ∀ c constrs. Inject c constrs (SOP constrs) ⇒ FunTy (ConstrOf c constrs) (SOP constrs)

type family ALG constrs r where ...

An `ALG constrs r` is a function that takes a way to turn every constructor into an r: ``` ALG (SOPOf (Rep (Either Int Bool))) r = (Int -> r) -> (Bool -> r) -> r ```

Equations

ALG '[c ::: prod] r = FunTy prod r → r

ALG ((c ::: prod) : constrs) r = FunTy prod r → ALG constrs r

class SOPLike constrs r where

Methods

algebra ∷ SOP constrs → ALG constrs r

Run a SOP

consts ∷ r → ALG constrs r

Ignore everything in the SOP

class SimpleConstructor rep where

Methods

toSimpleCon' ∷ rep p → ProdOver (Constr rep)

fromSimpleCon' ∷ ProdOver (Constr rep) → rep p

class SopList xs ys where

Construct and deconstruct cases in a SOP

Methods

injectSOPLeft ∷ SOP xs → SOP (Append xs ys)

injectSOPRight ∷ SOP ys → SOP (Append xs ys)

caseSOP ∷ SOP (Append xs ys) → Sum (SOP xs) (SOP ys)