Constrained.List

Documentation

data List (f ∷ k → Type) (as ∷ [k]) where

A heterogeneous list / an inductive tuple. We use this heavily to represent arguments to functions in terms. A term application (c.f. Base) is represented as `App :: fn as b -> List (Term fn) as -> Term fn b` for example.

Constructors

Nil ∷ List f '[]

(:>) ∷ f a → List f as → List f (a ': as) infixr 5

type family Length as where ...

Equations

Length '[] = 0

Length (_ ': as) = 1 + Length as

type family as :! n where ... infixr 5

Equations

'[] :! n = TypeError ('Text "Indexing into empty type-level list")

(a ': as) :! 0 = a

(a ': as) :! n = as :! (n - 1)

class IndexOf n as where

Methods

at ∷ List f as → f (as :! n)

toList ∷ (∀ a. f a → b) → List f as → [b]

mapListC_ ∷ ∀ c f as b. All c as ⇒ (∀ a. c a ⇒ f a → b) → List f as → [b]

mapList ∷ (∀ a. f a → g a) → List f as → List g as

mapListC ∷ ∀ c as f g. All c as ⇒ (∀ a. c a ⇒ f a → g a) → List f as → List g as

mapMList ∷ Applicative m ⇒ (∀ a. f a → m (g a)) → List f as → m (List g as)

mapMListC ∷ ∀ c as f g m. (Applicative m, All c as) ⇒ (∀ a. c a ⇒ f a → m (g a)) → List f as → m (List g as)

foldMapList ∷ Monoid b ⇒ (∀ a. f a → b) → List f as → b

foldMapListC ∷ ∀ c as b f. (All c as, Monoid b) ⇒ (∀ a. c a ⇒ f a → b) → List f as → b

appendList ∷ List f as → List f bs → List f (Append as bs)

lengthList ∷ List f as → Int

type family Append as as' where ...

Append two type-level lists

Equations

Append '[] as' = as'

Append (a ': as) as' = a ': Append as as'

type family MapList (f ∷ k → j) as where ...

Map a type functor over a list

Equations

MapList f '[] = '[]

MapList f (a ': as) = f a ': MapList f as

data ListCtx f (as ∷ [Type]) c where

A List with a hole in it, can be seen as a zipper over type-level lists.

We use this to represent contexts over terms in Base. For example, an application of f to a single variable of type a would be the pair `(fn as b, ListCtx Value as (HOLE a))` where HOLE (c.f. Base) is isomorphic to :~:.

Constructors

(:?) ∷ c a → List f as → ListCtx f (a ': as) c infixr 5

(:!) ∷ f a → ListCtx f as c → ListCtx f (a ': as) c infixr 5

pattern NilCtx ∷ c a → ListCtx f '[a] c

A Convenient pattern for singleton contexts

pattern ListCtx ∷ () ⇒ as'' ~ Append as (a ': as') ⇒ List f as → c a → List f as' → ListCtx f as'' c

A view of a ListCtx where you see the whole context at the same time.

data ListCtxWhole f as c where

Internals for the ListCtx pattern

Constructors

ListCtxWhole ∷ List f as → c a → List f as' → ListCtxWhole f (Append as (a ': as')) c

toWholeCtx ∷ ListCtx f as c → ListCtxWhole f as c

fromWholeCtx ∷ ListCtxWhole f as c → ListCtx f as c

fillListCtx ∷ ListCtx f as c → (∀ a. c a → f a) → List f as

mapListCtx ∷ (∀ a. f a → g a) → ListCtx f as c → ListCtx g as c

mapListCtxC ∷ ∀ c as f g h. All c as ⇒ (∀ a. c a ⇒ f a → g a) → ListCtx f as h → ListCtx g as h

type family FunTy ts res where ...

A function type from ts to res: FunTy '[Int, Bool] Double = Int -> Bool -> Double

Equations

FunTy '[] a = a

FunTy (a ': as) r = a → FunTy as r

listShape ∷ ∀ as. TypeList as ⇒ List (Const ()) as

Materialize the shape of the type list as, this is very useful for avoiding having to write type classes that recurse over as.

class TypeList ts where

Higher-order functions for working on Lists

Methods

uncurryList ∷ FunTy (MapList f ts) r → List f ts → r

uncurryList_ ∷ (∀ a. f a → a) → FunTy ts r → List f ts → r

curryList ∷ (List f ts → r) → FunTy (MapList f ts) r

curryList_ ∷ (∀ a. a → f a) → (List f ts → r) → FunTy ts r

unfoldList ∷ (∀ a as. List f as → f a) → List f ts

type family All (c ∷ k → Constraint) (as ∷ [k]) ∷ Constraint where ...

Equations

All c '[] = ()

All c (a ': as) = (c a, All c as)