Constrained.NumSpec

Documentation

data NumOrdW (dom ∷ [Type]) (rng ∷ Type) where

Constructors

LessOrEqualW ∷ OrdLike a ⇒ NumOrdW '[a, a] Bool

LessW ∷ OrdLike a ⇒ NumOrdW '[a, a] Bool

GreaterOrEqualW ∷ OrdLike a ⇒ NumOrdW '[a, a] Bool

GreaterW ∷ OrdLike a ⇒ NumOrdW '[a, a] Bool

class (Ord a, HasSpec a) ⇒ OrdLike a where

Minimal complete definition

Nothing

Methods

leqSpec ∷ a → Specification a

default leqSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

ltSpec ∷ a → Specification a

default ltSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

geqSpec ∷ a → Specification a

default geqSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

gtSpec ∷ a → Specification a

default gtSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

class MaybeBounded a where

Minimal complete definition

Nothing

Methods

lowerBound ∷ Maybe a

default lowerBound ∷ Bounded a ⇒ Maybe a

upperBound ∷ Maybe a

default upperBound ∷ Bounded a ⇒ Maybe a

newtype Unbounded a

Constructors

Unbounded a

data NumSpec n

Constructors

NumSpecInterval (Maybe n) (Maybe n)

emptyNumSpec ∷ Ord a ⇒ NumSpec a

guardNumSpec ∷ (Ord n, HasSpec n, TypeSpec n ~ NumSpec n) ⇒ [String] → NumSpec n → Specification n

combineNumSpec ∷ (HasSpec n, Ord n, TypeSpec n ~ NumSpec n) ⇒ NumSpec n → NumSpec n → Specification n

genFromNumSpec ∷ (MonadGenError m, Show n, Random n, Ord n, Num n, MaybeBounded n) ⇒ NumSpec n → GenT m n

shrinkWithNumSpec ∷ Arbitrary n ⇒ NumSpec n → n → [n]

constrainInterval ∷ (MonadGenError m, Ord a, Num a, Show a) ⇒ Maybe a → Maybe a → Integer → m (a, a)

conformsToNumSpec ∷ Ord n ⇒ n → NumSpec n → Bool

nubOrd ∷ Ord a ⇒ [a] → [a]

Strip out duplicates (in n-log(n) time, by building an intermediate Set)

nubOrdMemberSpec ∷ Ord a ⇒ String → [a] → Specification a

Builds a MemberSpec, but returns an Error spec if the list is empty

lowBound ∷ Bounded n ⇒ Maybe n → n

highBound ∷ Bounded n ⇒ Maybe n → n

countSpec ∷ ∀ n. (Bounded n, Integral n) ⇒ NumSpec n → Integer

The exact count of the number elements in a Bounded NumSpec

finiteSize ∷ ∀ n. (Integral n, Bounded n) ⇒ Integer

The exact number of elements in a Bounded Integral type.

notInNumSpec ∷ ∀ n. (HasSpec n, TypeSpec n ~ NumSpec n, Bounded n, Integral n) ⇒ NumSpec n → [n] → Specification n

This is an optimizing version of TypeSpec :: TypeSpec n -> [n] -> Specification n for Bounded NumSpecs. notInNumSpec :: Bounded n => TypeSpec n -> [n] -> Specification n We use this function to specialize the (HasSpec t) method typeSpecOpt for Bounded n. So given (TypeSpec interval badlist) we might want to transform it to (MemberSpec goodlist) There are 2 opportunities where this can payoff big time. 1) Suppose the total count of the elements in the interval is < length badlist we can then return (MemberSpec (filter elements (notElem badlist))) this must be smaller than (TypeSpec interval badlist) because the filtered list must be smaller than badlist 2) Suppose the type t is finite with size N. If the length of the badlist > (N/2), then the number of possible good things must be smaller than (length badlist), because (possible good + bad == N), so regardless of the count of the interval (MemberSpec (filter elements (notElem badlist))) is better. Sometimes much better. Example, let n be the finite set {0,1,2,3,4,5,6,7,8,9} and the bad list be [0,1,3,4,5,6,8,9] (TypeSpec [0..9] [0,1,3,4,5,6,8,9]) = filter {0,1,2,3,4,5,6,7,8,9} (notElem [0,1,3,4,5,6,8,9]) = [2,7] So (MemberSpec [2,7]) is better than (TypeSpec [0..9] [0,1,3,4,5,6,8,9]). This works no matter what the count of interval is. We only need the (length badlist > (N/2)).

guardEmpty ∷ (Ord n, Num n) ⇒ Maybe n → Maybe n → NumSpec n → NumSpec n

addNumSpec ∷ (Ord n, Num n) ⇒ NumSpec n → NumSpec n → NumSpec n

subNumSpec ∷ (Ord n, Num n) ⇒ NumSpec n → NumSpec n → NumSpec n

multNumSpec ∷ (Ord n, Num n) ⇒ NumSpec n → NumSpec n → NumSpec n

negNumSpec ∷ Num n ⇒ NumSpec n → NumSpec n

data T x

T is a sort of special version of Maybe, with two Nothings. Given:: NumSpecInterval (Maybe n) (Maybe n) -> Numspec We can't distinguish between the two Nothings in (NumSpecInterval Nothing Nothing) But using (NumSpecInterval NegInf PosInf) we can, In fact we can make a total ordering on T So an ascending Sorted [T x] would all the NegInf on the left and all the PosInf on the right, with the Ok's sorted in between. I.e. [NegInf, NegInf, Ok 3, Ok 6, Ok 12, Pos Inf]

Constructors

NegInf

Ok x

PosInf

unT ∷ T x → Maybe x

neg ∷ Maybe x → T x

Use this on the lower bound. I.e. lo from pair (lo,hi)

pos ∷ Maybe x → T x

Use this on the upper bound. I.e. hi from pair (lo,hi)

multT ∷ Num x ⇒ T x → T x → T x

multiply two (T x), correctly handling the infinities NegInf and PosInf

type Number n = (Num n, Enum n, TypeSpec n ~ NumSpec n, Num (NumSpec n), HasSpec n, Ord n)

What constraints we need to make HasSpec instance for a Haskell numeric type. By abstracting over this, we can avoid making actual HasSpec instances until all the requirements (HasSpec Bool, HasSpec(Sum a b)) have been met in Constrained.TheKnot.

addSpecInt ∷ Number n ⇒ Specification n → Specification n → Specification n

subSpecInt ∷ Number n ⇒ Specification n → Specification n → Specification n

multSpecInt ∷ Number n ⇒ Specification n → Specification n → Specification n

operateSpec ∷ Number n ⇒ String → (n → n → n) → (TypeSpec n → TypeSpec n → TypeSpec n) → Specification n → Specification n → Specification n

let n be some numeric type, and f and ft be operations on n and (TypeSpec n) Then lift these operations from (TypeSpec n) to (Specification n) Normally f will be a (Num n) instance method (+,-,*) on n, and ft will be a a (Num (TypeSpec n)) instance method (+,-,*) on (TypeSpec n) But this will work for any operations f and ft with the right types

cardinality ∷ ∀ a. (Number Integer, HasSpec a) ⇒ Specification a → Specification Integer

Put some (admittedly loose bounds) on the number of solutions that genFromTypeSpec might return. For lots of types, there is no way to be very accurate. Here we lift the HasSpec methods cardinalTrueSpec and cardinalTypeSpec from (TypeSpec Integer) to (Specification Integer)

cardinalNumSpec ∷ ∀ n. (Integral n, MaybeBounded n) ⇒ NumSpec n → Specification Integer

A generic function to use as an instance for the HasSpec method cardinalTypeSpec :: HasSpec a => TypeSpec a -> Specification Integer for types n such that (TypeSpec n ~ NumSpec n)

class (Num a, HasSpec a) ⇒ NumLike a where

Minimal complete definition

Nothing

Methods

subtractSpec ∷ a → TypeSpec a → Specification a

default subtractSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, NumLike (SimpleRep a)) ⇒ a → TypeSpec a → Specification a

negateSpec ∷ TypeSpec a → Specification a

default negateSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, NumLike (SimpleRep a)) ⇒ TypeSpec a → Specification a

safeSubtract ∷ a → a → Maybe a

default safeSubtract ∷ (HasSimpleRep a, NumLike (SimpleRep a)) ⇒ a → a → Maybe a

data IntW (as ∷ [Type]) b where

Constructors

AddW ∷ NumLike a ⇒ IntW '[a, a] a

NegateW ∷ NumLike a ⇒ IntW '[a] a

type Numeric a = (HasSpec a, Ord a, Num a, TypeSpec a ~ NumSpec a, MaybeBounded a)

addFn ∷ ∀ a. NumLike a ⇒ Term a → Term a → Term a

negateFn ∷ ∀ a. NumLike a ⇒ Term a → Term a

Orphan instances

Arbitrary Natural

Instance details

Methods

arbitrary ∷ Gen Natural

shrink ∷ Natural → [Natural]

Random Natural

Instance details

Methods

randomR ∷ RandomGen g ⇒ (Natural, Natural) → g → (Natural, g)

random ∷ RandomGen g ⇒ g → (Natural, g)

randomRs ∷ RandomGen g ⇒ (Natural, Natural) → g → [Natural]

randoms ∷ RandomGen g ⇒ g → [Natural]

Uniform Natural

Instance details

Methods

uniformM ∷ StatefulGen g m ⇒ g → m Natural

Number Integer ⇒ Num (Specification Integer)
This is very liberal, since in lots of cases it returns TrueSpec. for example all operations on SuspendedSpec, and certain operations between TypeSpec and MemberSpec. Perhaps we should remove it. Only the addSpec (+) and multSpec (*) methods are used. But, it is kind of cool ... In Fact we can use this to make Num(Specification n) instance for any n. But, only Integer is safe, because in all other types (+) and especially (-) can lead to overflow or underflow failures.

Instance details

Methods

(+) ∷ Specification Integer → Specification Integer → Specification Integer

(-) ∷ Specification Integer → Specification Integer → Specification Integer

(*) ∷ Specification Integer → Specification Integer → Specification Integer

negate ∷ Specification Integer → Specification Integer

abs ∷ Specification Integer → Specification Integer

signum ∷ Specification Integer → Specification Integer

fromInteger ∷ Integer → Specification Integer

NumLike a ⇒ Num (Term a)

Instance details

Methods

(+) ∷ Term a → Term a → Term a

(-) ∷ Term a → Term a → Term a

(*) ∷ Term a → Term a → Term a

negate ∷ Term a → Term a

abs ∷ Term a → Term a

signum ∷ Term a → Term a

fromInteger ∷ Integer → Term a

Random (Ratio Integer)

Instance details

Methods

randomR ∷ RandomGen g ⇒ (Ratio Integer, Ratio Integer) → g → (Ratio Integer, g)

random ∷ RandomGen g ⇒ g → (Ratio Integer, g)

randomRs ∷ RandomGen g ⇒ (Ratio Integer, Ratio Integer) → g → [Ratio Integer]

randoms ∷ RandomGen g ⇒ g → [Ratio Integer]