Constrained.Properties

Description

Useful properties for debugging custom HasSpec instances.

Documentation

conformsToSpecProp ∷ ∀ a. HasSpec a ⇒ a → Specification a → Property

forAllSpecShow ∷ (HasSpec a, Testable p) ⇒ Specification a → (a → String) → (a → p) → Property

forAllSpec ∷ (HasSpec a, Testable p) ⇒ Specification a → (a → p) → Property

forAllSpecDiscard ∷ (HasSpec a, Testable p) ⇒ Specification a → (a → p) → Property

prop_sound ∷ HasSpec a ⇒ Specification a → Property

prop_constrained_satisfies_sound ∷ HasSpec a ⇒ Specification a → Property

prop_constrained_explained ∷ HasSpec a ⇒ Specification a → Property

prop_complete ∷ HasSpec a ⇒ Specification a → Property

`prop_complete ps` assumes that ps is satisfiable

prop_constrained_satisfies_complete ∷ HasSpec a ⇒ Specification a → Property

prop_shrink_sound ∷ HasSpec a ⇒ Specification a → Property

prop_conformEmpty ∷ ∀ a. HasSpec a ⇒ a → Property

prop_univSound ∷ TestableFn → Property

main ∷ IO ()

prop_gen_sound ∷ ∀ a. HasSpec a ⇒ Specification a → Property

specType ∷ Specification a → String

showCtxWith ∷ ∀ s t as b. (AppRequires s t as b, Fill as) ⇒ t s as b → TestableCtx as → String

Builds a Term from A FunSym and an appropriate (List Term) of args computed from the TestableCtx, using fillCList, which fills the HOLE with variable v0 (uncurryList $ appTerm ProdW) (Lit True :> Lit False :> Nil) ===> prod_ True False

runOnCtx ∷ ∀ s t as b x. (AppRequires s t as b, Fill as) ⇒ CList 'Pre as x x → t s as b → x → b

Takes a CList, replaces the hole with x, and turns it into a (List Value as), then uses the semantics of the function symbol fn applying it to each item in the List in turn, to get a resulting answer of type b

class Fill as where

Fill in the HOLE in a CList, and then turn it into a regular List

Methods

fillCList ∷ ∀ f hole m. (Ctx hole hole → f hole) → (∀ a. Literal a ⇒ a → f a) → CList m as hole hole → List f as

data TestableFn where

Constructors

TestableFn ∷ (Fill as, Arbitrary (Specification b), Typeable (FunTy as b), AppRequires s t as b) ⇒ t s as b → TestableFn

genCtxForFn ∷ ∀ s t as b. AppRequires s t as b ⇒ t s as b → Gen (TestableCtx as)

ex2 ∷ Prod (Value Bool) (Value Bool)

data TestableCtx as where

Constructors

TestableCtx ∷ (HasSpec a, All HasSpec as) ⇒ CList 'Pre as a a → TestableCtx as

genTestableCtx ∷ (All HasSpec as, TypeList as) ⇒ Gen (TestableCtx as)

genPre ∷ ∀ f as. All HasSpec as ⇒ Int → List f as → Gen (TestableCtx as)

genPost ∷ ∀ f as. All HasSpec as ⇒ List f as → Gen (PolyCList as)

shrinkCList ∷ All HasSpec as ⇒ CList m as x y → [CList m as x y]

genTestableFn ∷ Gen TestableFn

prop_mapSpec ∷ (HasSpec a, AppRequires s t '[a] b) ⇒ t s '[a] b → Specification a → Property

prop_propagateSpecSound ∷ (HasSpec a, AppRequires s t '[a] b) ⇒ t s '[a] b → b → Property