Constrained.Spec.Generics

Documentation

data GenericsFn fn args res

type IsNormalType a = (Cases a ~ '[a], Args a ~ '[a], IsProd a, CountCases a ~ 1)

fst_ ∷ ∀ fn a b. (HasSpec fn a, HasSpec fn b) ⇒ Term fn (a, b) → Term fn a

snd_ ∷ ∀ fn a b. (HasSpec fn a, HasSpec fn b) ⇒ Term fn (a, b) → Term fn b

pair_ ∷ ∀ fn a b. (HasSpec fn a, HasSpec fn b) ⇒ Term fn a → Term fn b → Term fn (a, b)

left_ ∷ (HasSpec fn a, HasSpec fn b, IsNormalType a, IsNormalType b) ⇒ Term fn a → Term fn (Either a b)

right_ ∷ (HasSpec fn a, HasSpec fn b, IsNormalType a, IsNormalType b) ⇒ Term fn b → Term fn (Either a b)

cJust_ ∷ (HasSpec fn a, IsNormalType a) ⇒ Term fn a → Term fn (Maybe a)

cNothing_ ∷ (HasSpec fn a, IsNormalType a) ⇒ Term fn (Maybe a)

caseOn ∷ ∀ fn a. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a), SimpleRep a ~ SumOver (Cases (SimpleRep a)), TypeList (Cases (SimpleRep a))) ⇒ Term fn a → FunTy (MapList (Weighted (Binder fn)) (Cases (SimpleRep a))) (Pred fn)

branch ∷ ∀ fn p a. (HasSpec fn a, All (HasSpec fn) (Args a), IsPred p fn, IsProd a) ⇒ FunTy (MapList (Term fn) (Args a)) p → Weighted (Binder fn) a

branchW ∷ ∀ fn p a. (HasSpec fn a, All (HasSpec fn) (Args a), IsPred p fn, IsProd a) ⇒ Int → FunTy (MapList (Term fn) (Args a)) p → Weighted (Binder fn) a

forAll' ∷ ∀ fn t a p. (Forallable t a, Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSpec fn t, HasSpec fn (SimpleRep a), HasSimpleRep a, All (HasSpec fn) (Args (SimpleRep a)), IsPred p fn, IsProd (SimpleRep a), HasSpec fn a) ⇒ Term fn t → FunTy (MapList (Term fn) (Args (SimpleRep a))) p → Pred fn

Like forAll but pattern matches on the `Term fn a`

constrained' ∷ ∀ a fn p. (Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSpec fn (SimpleRep a), HasSimpleRep a, All (HasSpec fn) (Args (SimpleRep a)), IsProd (SimpleRep a), HasSpec fn a, IsPred p fn) ⇒ FunTy (MapList (Term fn) (Args (SimpleRep a))) p → Specification fn a

Like constrained but pattern matches on the bound `Term fn a`

reify' ∷ ∀ fn a b p. (Cases (SimpleRep b) ~ '[SimpleRep b], TypeSpec fn b ~ TypeSpec fn (SimpleRep b), HasSpec fn (SimpleRep b), HasSimpleRep b, All (HasSpec fn) (Args (SimpleRep b)), IsProd (SimpleRep b), HasSpec fn a, HasSpec fn b, IsPred p fn) ⇒ Term fn a → (a → b) → FunTy (MapList (Term fn) (Args (SimpleRep b))) p → Pred fn

Like reify but pattern matches on the bound `Term fn b`

con ∷ ∀ c a r fn. (SimpleRep a ~ SOP (TheSop a), TypeSpec fn a ~ TypeSpec fn (SOP (TheSop a)), TypeList (ConstrOf c (TheSop a)), HasSpec fn a, HasSimpleRep a, r ~ FunTy (MapList (Term fn) (ConstrOf c (TheSop a))) (Term fn a), ResultType r ~ Term fn a, SOPTerm c fn (TheSop a), ConstrTerm fn (ConstrOf c (TheSop a))) ⇒ r

onCon ∷ ∀ c a fn p. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, HasSpec fn a, HasSpec fn (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All (HasSpec fn) (Cases (SOP (TheSop a))), HasSpec fn (ProdOver (ConstrOf c (TheSop a))), IsPred p fn, Args (ProdOver (ConstrOf c (TheSop a))) ~ ConstrOf c (TheSop a), All (HasSpec fn) (ConstrOf c (TheSop a)), IsProd (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term fn a → FunTy (MapList (Term fn) (ConstrOf c (TheSop a))) p → Pred fn

isCon ∷ ∀ c a fn. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, HasSpec fn a, HasSpec fn (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All (HasSpec fn) (Cases (SOP (TheSop a))), HasSpec fn (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term fn a → Pred fn

sel ∷ ∀ n fn a c as. (SimpleRep a ~ ProdOver as, TheSop a ~ '[c ::: as], TypeSpec fn a ~ TypeSpec fn (ProdOver as), Select fn n as, HasSpec fn a, HasSpec fn (ProdOver as), HasSimpleRep a) ⇒ Term fn a → Term fn (At n as)

match ∷ ∀ fn p a. (HasSpec fn a, IsProductType fn a, IsPred p fn) ⇒ Term fn a → FunTy (MapList (Term fn) (ProductAsList a)) p → Pred fn

onJust ∷ ∀ fn a p. (BaseUniverse fn, HasSpec fn a, IsNormalType a, IsPred p fn) ⇒ Term fn (Maybe a) → (Term fn a → p) → Pred fn

isJust ∷ ∀ fn a. (BaseUniverse fn, HasSpec fn a, IsNormalType a) ⇒ Term fn (Maybe a) → Pred fn

ifElse ∷ (BaseUniverse fn, IsPred p fn, IsPred q fn) ⇒ Term fn Bool → p → q → Pred fn

chooseSpec ∷ HasSpec fn a ⇒ (Int, Specification fn a) → (Int, Specification fn a) → Specification fn a

Orphan instances

(IsNormalType a, HasSpec fn a) ⇒ HasSpec fn (Maybe a)

Instance details

Associated Types

type TypeSpec fn (Maybe a)

type Prerequisites fn (Maybe a)

Methods

emptySpec ∷ TypeSpec fn (Maybe a)

combineSpec ∷ TypeSpec fn (Maybe a) → TypeSpec fn (Maybe a) → Specification fn (Maybe a)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (Maybe a) → GenT m (Maybe a)

conformsTo ∷ Maybe a → TypeSpec fn (Maybe a) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (Maybe a) → Maybe a → [Maybe a]

toPreds ∷ Term fn (Maybe a) → TypeSpec fn (Maybe a) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (Maybe a) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (Maybe a) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (Maybe a) → BinaryShow

monadConformsTo ∷ Maybe a → TypeSpec fn (Maybe a) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (Maybe a) → [Maybe a] → Specification fn (Maybe a)

prerequisites ∷ Evidence (Prerequisites fn (Maybe a))

(HasSpec fn a, IsNormalType a, HasSpec fn b, IsNormalType b) ⇒ HasSpec fn (Either a b)

Instance details

Associated Types

type TypeSpec fn (Either a b)

type Prerequisites fn (Either a b)

Methods

emptySpec ∷ TypeSpec fn (Either a b)

combineSpec ∷ TypeSpec fn (Either a b) → TypeSpec fn (Either a b) → Specification fn (Either a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (Either a b) → GenT m (Either a b)

conformsTo ∷ Either a b → TypeSpec fn (Either a b) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (Either a b) → Either a b → [Either a b]

toPreds ∷ Term fn (Either a b) → TypeSpec fn (Either a b) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (Either a b) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (Either a b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (Either a b) → BinaryShow

monadConformsTo ∷ Either a b → TypeSpec fn (Either a b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (Either a b) → [Either a b] → Specification fn (Either a b)

prerequisites ∷ Evidence (Prerequisites fn (Either a b))

(HasSpec fn a, HasSpec fn b) ⇒ HasSpec fn (a, b)

Instance details

Associated Types

type TypeSpec fn (a, b)

type Prerequisites fn (a, b)

Methods

emptySpec ∷ TypeSpec fn (a, b)

combineSpec ∷ TypeSpec fn (a, b) → TypeSpec fn (a, b) → Specification fn (a, b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (a, b) → GenT m (a, b)

conformsTo ∷ (a, b) → TypeSpec fn (a, b) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (a, b) → (a, b) → [(a, b)]

toPreds ∷ Term fn (a, b) → TypeSpec fn (a, b) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (a, b) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (a, b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (a, b) → BinaryShow

monadConformsTo ∷ (a, b) → TypeSpec fn (a, b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (a, b) → [(a, b)] → Specification fn (a, b)

prerequisites ∷ Evidence (Prerequisites fn (a, b))

(HasSpec fn a, HasSpec fn b, HasSpec fn c) ⇒ HasSpec fn (a, b, c)

Instance details

Associated Types

type TypeSpec fn (a, b, c)

type Prerequisites fn (a, b, c)

Methods

emptySpec ∷ TypeSpec fn (a, b, c)

combineSpec ∷ TypeSpec fn (a, b, c) → TypeSpec fn (a, b, c) → Specification fn (a, b, c)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (a, b, c) → GenT m (a, b, c)

conformsTo ∷ (a, b, c) → TypeSpec fn (a, b, c) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (a, b, c) → (a, b, c) → [(a, b, c)]

toPreds ∷ Term fn (a, b, c) → TypeSpec fn (a, b, c) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (a, b, c) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (a, b, c) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (a, b, c) → BinaryShow

monadConformsTo ∷ (a, b, c) → TypeSpec fn (a, b, c) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (a, b, c) → [(a, b, c)] → Specification fn (a, b, c)

prerequisites ∷ Evidence (Prerequisites fn (a, b, c))

(HasSpec fn a, HasSpec fn b, HasSpec fn c, HasSpec fn d) ⇒ HasSpec fn (a, b, c, d)

Instance details

Associated Types

type TypeSpec fn (a, b, c, d)

type Prerequisites fn (a, b, c, d)

Methods

emptySpec ∷ TypeSpec fn (a, b, c, d)

combineSpec ∷ TypeSpec fn (a, b, c, d) → TypeSpec fn (a, b, c, d) → Specification fn (a, b, c, d)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (a, b, c, d) → GenT m (a, b, c, d)

conformsTo ∷ (a, b, c, d) → TypeSpec fn (a, b, c, d) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (a, b, c, d) → (a, b, c, d) → [(a, b, c, d)]

toPreds ∷ Term fn (a, b, c, d) → TypeSpec fn (a, b, c, d) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (a, b, c, d) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (a, b, c, d) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (a, b, c, d) → BinaryShow

monadConformsTo ∷ (a, b, c, d) → TypeSpec fn (a, b, c, d) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (a, b, c, d) → [(a, b, c, d)] → Specification fn (a, b, c, d)

prerequisites ∷ Evidence (Prerequisites fn (a, b, c, d))

(HasSpec fn a, HasSpec fn b, HasSpec fn c, HasSpec fn d, HasSpec fn e) ⇒ HasSpec fn (a, b, c, d, e)

Instance details

Associated Types

type TypeSpec fn (a, b, c, d, e)

type Prerequisites fn (a, b, c, d, e)

Methods

emptySpec ∷ TypeSpec fn (a, b, c, d, e)

combineSpec ∷ TypeSpec fn (a, b, c, d, e) → TypeSpec fn (a, b, c, d, e) → Specification fn (a, b, c, d, e)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (a, b, c, d, e) → GenT m (a, b, c, d, e)

conformsTo ∷ (a, b, c, d, e) → TypeSpec fn (a, b, c, d, e) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (a, b, c, d, e) → (a, b, c, d, e) → [(a, b, c, d, e)]

toPreds ∷ Term fn (a, b, c, d, e) → TypeSpec fn (a, b, c, d, e) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (a, b, c, d, e) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (a, b, c, d, e) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (a, b, c, d, e) → BinaryShow

monadConformsTo ∷ (a, b, c, d, e) → TypeSpec fn (a, b, c, d, e) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (a, b, c, d, e) → [(a, b, c, d, e)] → Specification fn (a, b, c, d, e)

prerequisites ∷ Evidence (Prerequisites fn (a, b, c, d, e))

(HasSpec fn a, HasSpec fn b, HasSpec fn c, HasSpec fn d, HasSpec fn e, HasSpec fn g) ⇒ HasSpec fn (a, b, c, d, e, g)

Instance details

Associated Types

type TypeSpec fn (a, b, c, d, e, g)

type Prerequisites fn (a, b, c, d, e, g)

Methods

emptySpec ∷ TypeSpec fn (a, b, c, d, e, g)

combineSpec ∷ TypeSpec fn (a, b, c, d, e, g) → TypeSpec fn (a, b, c, d, e, g) → Specification fn (a, b, c, d, e, g)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (a, b, c, d, e, g) → GenT m (a, b, c, d, e, g)

conformsTo ∷ (a, b, c, d, e, g) → TypeSpec fn (a, b, c, d, e, g) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (a, b, c, d, e, g) → (a, b, c, d, e, g) → [(a, b, c, d, e, g)]

toPreds ∷ Term fn (a, b, c, d, e, g) → TypeSpec fn (a, b, c, d, e, g) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (a, b, c, d, e, g) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (a, b, c, d, e, g) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (a, b, c, d, e, g) → BinaryShow

monadConformsTo ∷ (a, b, c, d, e, g) → TypeSpec fn (a, b, c, d, e, g) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (a, b, c, d, e, g) → [(a, b, c, d, e, g)] → Specification fn (a, b, c, d, e, g)

prerequisites ∷ Evidence (Prerequisites fn (a, b, c, d, e, g))

(HasSpec fn a, HasSpec fn b, HasSpec fn c, HasSpec fn d, HasSpec fn e, HasSpec fn g, HasSpec fn h) ⇒ HasSpec fn (a, b, c, d, e, g, h)

Instance details

Associated Types

type TypeSpec fn (a, b, c, d, e, g, h)

type Prerequisites fn (a, b, c, d, e, g, h)

Methods

emptySpec ∷ TypeSpec fn (a, b, c, d, e, g, h)

combineSpec ∷ TypeSpec fn (a, b, c, d, e, g, h) → TypeSpec fn (a, b, c, d, e, g, h) → Specification fn (a, b, c, d, e, g, h)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec fn (a, b, c, d, e, g, h) → GenT m (a, b, c, d, e, g, h)

conformsTo ∷ (a, b, c, d, e, g, h) → TypeSpec fn (a, b, c, d, e, g, h) → Bool

shrinkWithTypeSpec ∷ TypeSpec fn (a, b, c, d, e, g, h) → (a, b, c, d, e, g, h) → [(a, b, c, d, e, g, h)]

toPreds ∷ Term fn (a, b, c, d, e, g, h) → TypeSpec fn (a, b, c, d, e, g, h) → Pred fn

cardinalTypeSpec ∷ TypeSpec fn (a, b, c, d, e, g, h) → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

typeSpecHasError ∷ TypeSpec fn (a, b, c, d, e, g, h) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec fn (a, b, c, d, e, g, h) → BinaryShow

monadConformsTo ∷ (a, b, c, d, e, g, h) → TypeSpec fn (a, b, c, d, e, g, h) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec fn (a, b, c, d, e, g, h) → [(a, b, c, d, e, g, h)] → Specification fn (a, b, c, d, e, g, h)

prerequisites ∷ Evidence (Prerequisites fn (a, b, c, d, e, g, h))

HasSimpleRep (Maybe a)

Instance details

Associated Types

type SimpleRep (Maybe a)

type TheSop (Maybe a) ∷ [Type]

Methods

toSimpleRep ∷ Maybe a → SimpleRep (Maybe a)

fromSimpleRep ∷ SimpleRep (Maybe a) → Maybe a

FunctionLike (GenericsFn fn)

Instance details

Methods

sem ∷ ∀ (as ∷ [Type]) b. GenericsFn fn as b → FunTy as b

BaseUniverse fn ⇒ Functions (GenericsFn fn) fn

Instance details

Methods

propagateSpecFun ∷ ∀ (as ∷ [Type]) a b. (TypeList as, Typeable as, HasSpec fn a, HasSpec fn b, All (HasSpec fn) as) ⇒ GenericsFn fn as b → ListCtx Value as (HOLE a) → Specification fn b → Specification fn a

rewriteRules ∷ ∀ (as ∷ [Type]) b. (TypeList as, Typeable as, HasSpec fn b, All (HasSpec fn) as) ⇒ GenericsFn fn as b → List (Term fn) as → Maybe (Term fn b)

mapTypeSpec ∷ (HasSpec fn a, HasSpec fn b) ⇒ GenericsFn fn '[a] b → TypeSpec fn a → Specification fn b

BaseUniverse fn ⇒ Functions (SumFn fn) fn

Instance details

Methods

propagateSpecFun ∷ ∀ (as ∷ [Type]) a b. (TypeList as, Typeable as, HasSpec fn a, HasSpec fn b, All (HasSpec fn) as) ⇒ SumFn fn as b → ListCtx Value as (HOLE a) → Specification fn b → Specification fn a

rewriteRules ∷ ∀ (as ∷ [Type]) b. (TypeList as, Typeable as, HasSpec fn b, All (HasSpec fn) as) ⇒ SumFn fn as b → List (Term fn) as → Maybe (Term fn b)

mapTypeSpec ∷ (HasSpec fn a, HasSpec fn b) ⇒ SumFn fn '[a] b → TypeSpec fn a → Specification fn b

(HasSpec fn a, HasSpec fn b, Arbitrary (FoldSpec fn a), Arbitrary (FoldSpec fn b)) ⇒ Arbitrary (FoldSpec fn (a, b))

Instance details

Methods

arbitrary ∷ Gen (FoldSpec fn (a, b))

shrink ∷ FoldSpec fn (a, b) → [FoldSpec fn (a, b)]

HasSimpleRep (Either a b)

Instance details

Associated Types

type SimpleRep (Either a b)

type TheSop (Either a b) ∷ [Type]

Methods

toSimpleRep ∷ Either a b → SimpleRep (Either a b)

fromSimpleRep ∷ SimpleRep (Either a b) → Either a b

HasSimpleRep (a, b)

Instance details

Associated Types

type SimpleRep (a, b)

type TheSop (a, b) ∷ [Type]

Methods

toSimpleRep ∷ (a, b) → SimpleRep (a, b)

fromSimpleRep ∷ SimpleRep (a, b) → (a, b)

HasSimpleRep (a, b, c)

Instance details

Associated Types

type SimpleRep (a, b, c)

type TheSop (a, b, c) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c) → SimpleRep (a, b, c)

fromSimpleRep ∷ SimpleRep (a, b, c) → (a, b, c)

HasSimpleRep (a, b, c, d)

Instance details

Associated Types

type SimpleRep (a, b, c, d)

type TheSop (a, b, c, d) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d) → SimpleRep (a, b, c, d)

fromSimpleRep ∷ SimpleRep (a, b, c, d) → (a, b, c, d)

HasSimpleRep (a, b, c, d, e)

Instance details

Associated Types

type SimpleRep (a, b, c, d, e)

type TheSop (a, b, c, d, e) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d, e) → SimpleRep (a, b, c, d, e)

fromSimpleRep ∷ SimpleRep (a, b, c, d, e) → (a, b, c, d, e)

HasSimpleRep (a, b, c, d, e, g)

Instance details

Associated Types

type SimpleRep (a, b, c, d, e, g)

type TheSop (a, b, c, d, e, g) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d, e, g) → SimpleRep (a, b, c, d, e, g)

fromSimpleRep ∷ SimpleRep (a, b, c, d, e, g) → (a, b, c, d, e, g)

HasSimpleRep (a, b, c, d, e, g, h)

Instance details

Associated Types

type SimpleRep (a, b, c, d, e, g, h)

type TheSop (a, b, c, d, e, g, h) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d, e, g, h) → SimpleRep (a, b, c, d, e, g, h)

fromSimpleRep ∷ SimpleRep (a, b, c, d, e, g, h) → (a, b, c, d, e, g, h)