Constrained.Spec.Size

Documentation

genFromSizeSpec ∷ MonadGenError m ⇒ Specification Integer → GenT m Integer

Because Sizes should always be >= 0, We provide this alternate generator that can be used to replace (genFromSpecT @Integer), to ensure this important property

data SizeW (s ∷ Symbol) (dom ∷ [Type]) rng ∷ Type where

Constructors

SizeOfW ∷ ∀ n. Sized n ⇒ SizeW "sizeOf_" '[n] Integer

sizeOfFn ∷ ∀ a. (HasSpec a, Sized a) ⇒ Fun '[a] Integer

sizeOf_ ∷ (HasSpec t, Sized t) ⇒ Term t → Term Integer

rangeSize ∷ Integer → Integer → SizeSpec

between ∷ (HasSpec a, TypeSpec a ~ NumSpec a) ⇒ a → a → Specification a

maxSpec ∷ Specification Integer → Specification Integer

The widest interval whose largest element is admitted by the original spec

type SizeSpec = NumSpec Integer

class Sized t where

Minimal complete definition

Nothing

Methods

sizeOf ∷ t → Integer

default sizeOf ∷ (HasSimpleRep t, Sized (SimpleRep t)) ⇒ t → Integer

liftSizeSpec ∷ HasSpec t ⇒ SizeSpec → [Integer] → Specification t

default liftSizeSpec ∷ (HasSpec t, HasSimpleRep t, Sized (SimpleRep t), HasSpec (SimpleRep t), TypeSpec t ~ TypeSpec (SimpleRep t)) ⇒ SizeSpec → [Integer] → Specification t

liftMemberSpec ∷ HasSpec t ⇒ [Integer] → Specification t

default liftMemberSpec ∷ (HasSpec t, HasSpec (SimpleRep t), HasSimpleRep t, Sized (SimpleRep t), TypeSpec t ~ TypeSpec (SimpleRep t)) ⇒ [Integer] → Specification t

sizeOfTypeSpec ∷ HasSpec t ⇒ TypeSpec t → Specification Integer

default sizeOfTypeSpec ∷ (HasSpec (SimpleRep t), Sized (SimpleRep t), TypeSpec t ~ TypeSpec (SimpleRep t)) ⇒ TypeSpec t → Specification Integer

hasSize ∷ (HasSpec t, Sized t) ⇒ SizeSpec → Specification t

Orphan instances

(Sized t, HasSpec t) ⇒ Logic "sizeOf_" SizeW '[t] Integer

Instance details

Methods

info ∷ SizeW "sizeOf_" '[t] Integer → String

propagate ∷ Context "sizeOf_" SizeW '[t] Integer hole → Specification Integer → Specification hole

rewriteRules ∷ SizeW "sizeOf_" '[t] Integer → List Term '[t] → Evidence (AppRequires "sizeOf_" SizeW '[t] Integer) → Maybe (Term Integer)

mapTypeSpec ∷ ('[t] ~ '[a], Integer ~ b, HasSpec a, HasSpec b) ⇒ SizeW "sizeOf_" '[a] b → TypeSpec a → Specification b

saturate ∷ SizeW "sizeOf_" '[t] Bool → List Term '[t] → [Pred]