Constrained.Spec.SumProd

Documentation

type IsNormalType a = (Cases a ~ '[a], Args a ~ '[a], IsProd a, CountCases a ~ 1)

type IsProductType a = (HasSimpleRep a, Cases (SimpleRep a) ~ '[SimpleRep a], SimpleRep a ~ SumOver (Cases (SimpleRep a)), IsProd (SimpleRep a), HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a), All HasSpec (ProductAsList a))

type ProductAsList a = Args (SimpleRep a)

caseOn ∷ ∀ a. (HasSpec a, HasSpec (SimpleRep a), HasSimpleRep a, TypeSpec a ~ TypeSpec (SimpleRep a), SimpleRep a ~ SumOver (Cases (SimpleRep a)), TypeList (Cases (SimpleRep a))) ⇒ Term a → FunTy (MapList (Weighted Binder) (Cases (SimpleRep a))) Pred

branch ∷ ∀ p a. (HasSpec a, All HasSpec (Args a), IsPred p, IsProd a) ⇒ FunTy (MapList Term (Args a)) p → Weighted Binder a

branchW ∷ ∀ p a. (HasSpec a, All HasSpec (Args a), IsPred p, IsProd a) ⇒ Int → FunTy (MapList Term (Args a)) p → Weighted Binder a

forAll' ∷ ∀ t a p. (Forallable t a, Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec a ~ TypeSpec (SimpleRep a), HasSpec t, HasSpec (SimpleRep a), HasSimpleRep a, All HasSpec (Args (SimpleRep a)), IsPred p, IsProd (SimpleRep a), HasSpec a) ⇒ Term t → FunTy (MapList Term (Args (SimpleRep a))) p → Pred

Like forAll but pattern matches on the `Term a`

constrained' ∷ ∀ a p. (Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec a ~ TypeSpec (SimpleRep a), HasSpec (SimpleRep a), HasSimpleRep a, All HasSpec (Args (SimpleRep a)), IsProd (SimpleRep a), HasSpec a, IsPred p) ⇒ FunTy (MapList Term (Args (SimpleRep a))) p → Specification a

Like constrained but pattern matches on the bound `Term a`

reify' ∷ ∀ a b p. (Cases (SimpleRep b) ~ '[SimpleRep b], TypeSpec b ~ TypeSpec (SimpleRep b), HasSpec (SimpleRep b), HasSimpleRep b, All HasSpec (Args (SimpleRep b)), IsProd (SimpleRep b), HasSpec a, HasSpec b, IsPred p) ⇒ Term a → (a → b) → FunTy (MapList Term (Args (SimpleRep b))) p → Pred

Like reify but pattern matches on the bound `Term b`

con ∷ ∀ c a r. (SimpleRep a ~ SOP (TheSop a), TypeSpec a ~ TypeSpec (SOP (TheSop a)), TypeList (ConstrOf c (TheSop a)), HasSpec a, HasSimpleRep a, r ~ FunTy (MapList Term (ConstrOf c (TheSop a))) (Term a), ResultType r ~ Term a, SOPTerm c (TheSop a), ConstrTerm (ConstrOf c (TheSop a))) ⇒ r

onCon ∷ ∀ c a p. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, HasSpec a, HasSpec (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All HasSpec (Cases (SOP (TheSop a))), HasSpec (ProdOver (ConstrOf c (TheSop a))), IsPred p, Args (ProdOver (ConstrOf c (TheSop a))) ~ ConstrOf c (TheSop a), All HasSpec (ConstrOf c (TheSop a)), IsProd (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term a → FunTy (MapList Term (ConstrOf c (TheSop a))) p → Pred

isCon ∷ ∀ c a. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, HasSpec a, HasSpec (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All HasSpec (Cases (SOP (TheSop a))), HasSpec (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term a → Pred

sel ∷ ∀ n a c as. (SimpleRep a ~ ProdOver as, TheSop a ~ '[c ::: as], TypeSpec a ~ TypeSpec (ProdOver as), Select n as, HasSpec a, HasSpec (ProdOver as), HasSimpleRep a) ⇒ Term a → Term (At n as)

match ∷ ∀ p a. (HasSpec a, IsProductType a, IsPred p) ⇒ Term a → FunTy (MapList Term (ProductAsList a)) p → Pred

onJust ∷ ∀ a p. (HasSpec a, IsNormalType a, IsPred p) ⇒ Term (Maybe a) → (Term a → p) → Pred

isJust ∷ ∀ a. (HasSpec a, IsNormalType a) ⇒ Term (Maybe a) → Pred

chooseSpec ∷ HasSpec a ⇒ (Int, Specification a) → (Int, Specification a) → Specification a

ChooseSpec is one of the ways we can Or two Specs together This works for any kind of type that has a HasSpec instance. If your type is a Sum type. One can use CaseOn which is much easier.

left_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term a → Term (Either a b)

right_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term b → Term (Either a b)

cJust_ ∷ (HasSpec a, IsNormalType a) ⇒ Term a → Term (Maybe a)

cNothing_ ∷ (HasSpec a, IsNormalType a) ⇒ Term (Maybe a)

fst_ ∷ (HasSpec x, HasSpec y) ⇒ Term (x, y) → Term x

snd_ ∷ (HasSpec x, HasSpec y) ⇒ Term (x, y) → Term y

fstW ∷ (HasSpec a, HasSpec b) ⇒ FunW "composeFn" '[(a, b)] a

sndW ∷ (HasSpec a, HasSpec b) ⇒ FunW "composeFn" '[(a, b)] b

pair_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term a → Term b → Term (a, b)

leftFn ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term a → Term (Sum a b)

rightFn ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term b → Term (Sum a b)

prodFst_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term a

prodSnd_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term b

prod_ ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Term (Prod a b)

data PairSpec a b

Constructors

Cartesian (Specification a) (Specification b)

Orphan instances

(HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Logic "leftFn" BaseW '[a] (Sum a b)

Instance details

Methods

info ∷ BaseW "leftFn" '[a] (Sum a b) → String

propagate ∷ Context "leftFn" BaseW '[a] (Sum a b) hole → Specification (Sum a b) → Specification hole

rewriteRules ∷ BaseW "leftFn" '[a] (Sum a b) → List Term '[a] → Evidence (AppRequires "leftFn" BaseW '[a] (Sum a b)) → Maybe (Term (Sum a b))

mapTypeSpec ∷ ('[a] ~ '[a0], Sum a b ~ b0, HasSpec a0, HasSpec b0) ⇒ BaseW "leftFn" '[a0] b0 → TypeSpec a0 → Specification b0

saturate ∷ BaseW "leftFn" '[a] Bool → List Term '[a] → [Pred]

(HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Logic "rightFn" BaseW '[b] (Sum a b)

Instance details

Methods

info ∷ BaseW "rightFn" '[b] (Sum a b) → String

propagate ∷ Context "rightFn" BaseW '[b] (Sum a b) hole → Specification (Sum a b) → Specification hole

rewriteRules ∷ BaseW "rightFn" '[b] (Sum a b) → List Term '[b] → Evidence (AppRequires "rightFn" BaseW '[b] (Sum a b)) → Maybe (Term (Sum a b))

mapTypeSpec ∷ ('[b] ~ '[a0], Sum a b ~ b0, HasSpec a0, HasSpec b0) ⇒ BaseW "rightFn" '[a0] b0 → TypeSpec a0 → Specification b0

saturate ∷ BaseW "rightFn" '[b] Bool → List Term '[b] → [Pred]

(HasSpec a, HasSpec b, Arbitrary (FoldSpec a), Arbitrary (FoldSpec b)) ⇒ Arbitrary (FoldSpec (a, b))

Instance details

Methods

arbitrary ∷ Gen (FoldSpec (a, b))

shrink ∷ FoldSpec (a, b) → [FoldSpec (a, b)]

(IsNormalType a, HasSpec a) ⇒ HasSpec (Maybe a)

Instance details

Associated Types

type TypeSpec (Maybe a)

type Prerequisites (Maybe a)

Methods

emptySpec ∷ TypeSpec (Maybe a)

combineSpec ∷ TypeSpec (Maybe a) → TypeSpec (Maybe a) → Specification (Maybe a)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Maybe a) → GenT m (Maybe a)

conformsTo ∷ Maybe a → TypeSpec (Maybe a) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Maybe a) → Maybe a → [Maybe a]

toPreds ∷ Term (Maybe a) → TypeSpec (Maybe a) → Pred

cardinalTypeSpec ∷ TypeSpec (Maybe a) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Maybe a) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Maybe a) → BinaryShow

monadConformsTo ∷ Maybe a → TypeSpec (Maybe a) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Maybe a) → [Maybe a] → Specification (Maybe a)

guardTypeSpec ∷ [String] → TypeSpec (Maybe a) → Specification (Maybe a)

prerequisites ∷ Evidence (Prerequisites (Maybe a))

Typeable a ⇒ HasSimpleRep (Maybe a)

Instance details

Associated Types

type SimpleRep (Maybe a)

type TheSop (Maybe a) ∷ [Type]

Methods

toSimpleRep ∷ Maybe a → SimpleRep (Maybe a)

fromSimpleRep ∷ SimpleRep (Maybe a) → Maybe a

(HasSpec a, IsNormalType a, HasSpec b, IsNormalType b) ⇒ HasSpec (Either a b)

Instance details

Associated Types

type TypeSpec (Either a b)

type Prerequisites (Either a b)

Methods

emptySpec ∷ TypeSpec (Either a b)

combineSpec ∷ TypeSpec (Either a b) → TypeSpec (Either a b) → Specification (Either a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Either a b) → GenT m (Either a b)

conformsTo ∷ Either a b → TypeSpec (Either a b) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Either a b) → Either a b → [Either a b]

toPreds ∷ Term (Either a b) → TypeSpec (Either a b) → Pred

cardinalTypeSpec ∷ TypeSpec (Either a b) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Either a b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Either a b) → BinaryShow

monadConformsTo ∷ Either a b → TypeSpec (Either a b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Either a b) → [Either a b] → Specification (Either a b)

guardTypeSpec ∷ [String] → TypeSpec (Either a b) → Specification (Either a b)

prerequisites ∷ Evidence (Prerequisites (Either a b))

(HasSpec a, HasSpec b) ⇒ HasSpec (a, b)

Instance details

Associated Types

type TypeSpec (a, b)

type Prerequisites (a, b)

Methods

emptySpec ∷ TypeSpec (a, b)

combineSpec ∷ TypeSpec (a, b) → TypeSpec (a, b) → Specification (a, b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (a, b) → GenT m (a, b)

conformsTo ∷ (a, b) → TypeSpec (a, b) → Bool

shrinkWithTypeSpec ∷ TypeSpec (a, b) → (a, b) → [(a, b)]

toPreds ∷ Term (a, b) → TypeSpec (a, b) → Pred

cardinalTypeSpec ∷ TypeSpec (a, b) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (a, b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (a, b) → BinaryShow

monadConformsTo ∷ (a, b) → TypeSpec (a, b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (a, b) → [(a, b)] → Specification (a, b)

guardTypeSpec ∷ [String] → TypeSpec (a, b) → Specification (a, b)

prerequisites ∷ Evidence (Prerequisites (a, b))

(Typeable a, Typeable b) ⇒ HasSimpleRep (Either a b)

Instance details

Associated Types

type SimpleRep (Either a b)

type TheSop (Either a b) ∷ [Type]

Methods

toSimpleRep ∷ Either a b → SimpleRep (Either a b)

fromSimpleRep ∷ SimpleRep (Either a b) → Either a b

(Typeable a, Typeable b) ⇒ HasSimpleRep (a, b)

Instance details

Associated Types

type SimpleRep (a, b)

type TheSop (a, b) ∷ [Type]

Methods

toSimpleRep ∷ (a, b) → SimpleRep (a, b)

fromSimpleRep ∷ SimpleRep (a, b) → (a, b)

(HasSpec a, HasSpec b, HasSpec c) ⇒ HasSpec (a, b, c)

Instance details

Associated Types

type TypeSpec (a, b, c)

type Prerequisites (a, b, c)

Methods

emptySpec ∷ TypeSpec (a, b, c)

combineSpec ∷ TypeSpec (a, b, c) → TypeSpec (a, b, c) → Specification (a, b, c)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (a, b, c) → GenT m (a, b, c)

conformsTo ∷ (a, b, c) → TypeSpec (a, b, c) → Bool

shrinkWithTypeSpec ∷ TypeSpec (a, b, c) → (a, b, c) → [(a, b, c)]

toPreds ∷ Term (a, b, c) → TypeSpec (a, b, c) → Pred

cardinalTypeSpec ∷ TypeSpec (a, b, c) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (a, b, c) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (a, b, c) → BinaryShow

monadConformsTo ∷ (a, b, c) → TypeSpec (a, b, c) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (a, b, c) → [(a, b, c)] → Specification (a, b, c)

guardTypeSpec ∷ [String] → TypeSpec (a, b, c) → Specification (a, b, c)

prerequisites ∷ Evidence (Prerequisites (a, b, c))

(Typeable a, Typeable b, Typeable c) ⇒ HasSimpleRep (a, b, c)

Instance details

Associated Types

type SimpleRep (a, b, c)

type TheSop (a, b, c) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c) → SimpleRep (a, b, c)

fromSimpleRep ∷ SimpleRep (a, b, c) → (a, b, c)

(HasSpec a, HasSpec b, HasSpec c, HasSpec d) ⇒ HasSpec (a, b, c, d)

Instance details

Associated Types

type TypeSpec (a, b, c, d)

type Prerequisites (a, b, c, d)

Methods

emptySpec ∷ TypeSpec (a, b, c, d)

combineSpec ∷ TypeSpec (a, b, c, d) → TypeSpec (a, b, c, d) → Specification (a, b, c, d)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (a, b, c, d) → GenT m (a, b, c, d)

conformsTo ∷ (a, b, c, d) → TypeSpec (a, b, c, d) → Bool

shrinkWithTypeSpec ∷ TypeSpec (a, b, c, d) → (a, b, c, d) → [(a, b, c, d)]

toPreds ∷ Term (a, b, c, d) → TypeSpec (a, b, c, d) → Pred

cardinalTypeSpec ∷ TypeSpec (a, b, c, d) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (a, b, c, d) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (a, b, c, d) → BinaryShow

monadConformsTo ∷ (a, b, c, d) → TypeSpec (a, b, c, d) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (a, b, c, d) → [(a, b, c, d)] → Specification (a, b, c, d)

guardTypeSpec ∷ [String] → TypeSpec (a, b, c, d) → Specification (a, b, c, d)

prerequisites ∷ Evidence (Prerequisites (a, b, c, d))

(Typeable a, Typeable b, Typeable c, Typeable d) ⇒ HasSimpleRep (a, b, c, d)

Instance details

Associated Types

type SimpleRep (a, b, c, d)

type TheSop (a, b, c, d) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d) → SimpleRep (a, b, c, d)

fromSimpleRep ∷ SimpleRep (a, b, c, d) → (a, b, c, d)

(HasSpec a, HasSpec b, HasSpec c, HasSpec d, HasSpec e) ⇒ HasSpec (a, b, c, d, e)

Instance details

Associated Types

type TypeSpec (a, b, c, d, e)

type Prerequisites (a, b, c, d, e)

Methods

emptySpec ∷ TypeSpec (a, b, c, d, e)

combineSpec ∷ TypeSpec (a, b, c, d, e) → TypeSpec (a, b, c, d, e) → Specification (a, b, c, d, e)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (a, b, c, d, e) → GenT m (a, b, c, d, e)

conformsTo ∷ (a, b, c, d, e) → TypeSpec (a, b, c, d, e) → Bool

shrinkWithTypeSpec ∷ TypeSpec (a, b, c, d, e) → (a, b, c, d, e) → [(a, b, c, d, e)]

toPreds ∷ Term (a, b, c, d, e) → TypeSpec (a, b, c, d, e) → Pred

cardinalTypeSpec ∷ TypeSpec (a, b, c, d, e) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (a, b, c, d, e) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (a, b, c, d, e) → BinaryShow

monadConformsTo ∷ (a, b, c, d, e) → TypeSpec (a, b, c, d, e) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (a, b, c, d, e) → [(a, b, c, d, e)] → Specification (a, b, c, d, e)

guardTypeSpec ∷ [String] → TypeSpec (a, b, c, d, e) → Specification (a, b, c, d, e)

prerequisites ∷ Evidence (Prerequisites (a, b, c, d, e))

(Typeable a, Typeable b, Typeable c, Typeable d, Typeable e) ⇒ HasSimpleRep (a, b, c, d, e)

Instance details

Associated Types

type SimpleRep (a, b, c, d, e)

type TheSop (a, b, c, d, e) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d, e) → SimpleRep (a, b, c, d, e)

fromSimpleRep ∷ SimpleRep (a, b, c, d, e) → (a, b, c, d, e)

(HasSpec a, HasSpec b, HasSpec c, HasSpec d, HasSpec e, HasSpec g) ⇒ HasSpec (a, b, c, d, e, g)

Instance details

Associated Types

type TypeSpec (a, b, c, d, e, g)

type Prerequisites (a, b, c, d, e, g)

Methods

emptySpec ∷ TypeSpec (a, b, c, d, e, g)

combineSpec ∷ TypeSpec (a, b, c, d, e, g) → TypeSpec (a, b, c, d, e, g) → Specification (a, b, c, d, e, g)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (a, b, c, d, e, g) → GenT m (a, b, c, d, e, g)

conformsTo ∷ (a, b, c, d, e, g) → TypeSpec (a, b, c, d, e, g) → Bool

shrinkWithTypeSpec ∷ TypeSpec (a, b, c, d, e, g) → (a, b, c, d, e, g) → [(a, b, c, d, e, g)]

toPreds ∷ Term (a, b, c, d, e, g) → TypeSpec (a, b, c, d, e, g) → Pred

cardinalTypeSpec ∷ TypeSpec (a, b, c, d, e, g) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (a, b, c, d, e, g) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (a, b, c, d, e, g) → BinaryShow

monadConformsTo ∷ (a, b, c, d, e, g) → TypeSpec (a, b, c, d, e, g) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (a, b, c, d, e, g) → [(a, b, c, d, e, g)] → Specification (a, b, c, d, e, g)

guardTypeSpec ∷ [String] → TypeSpec (a, b, c, d, e, g) → Specification (a, b, c, d, e, g)

prerequisites ∷ Evidence (Prerequisites (a, b, c, d, e, g))

(Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Typeable g) ⇒ HasSimpleRep (a, b, c, d, e, g)

Instance details

Associated Types

type SimpleRep (a, b, c, d, e, g)

type TheSop (a, b, c, d, e, g) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d, e, g) → SimpleRep (a, b, c, d, e, g)

fromSimpleRep ∷ SimpleRep (a, b, c, d, e, g) → (a, b, c, d, e, g)

(HasSpec a, HasSpec b, HasSpec c, HasSpec d, HasSpec e, HasSpec g, HasSpec h) ⇒ HasSpec (a, b, c, d, e, g, h)

Instance details

Associated Types

type TypeSpec (a, b, c, d, e, g, h)

type Prerequisites (a, b, c, d, e, g, h)

Methods

emptySpec ∷ TypeSpec (a, b, c, d, e, g, h)

combineSpec ∷ TypeSpec (a, b, c, d, e, g, h) → TypeSpec (a, b, c, d, e, g, h) → Specification (a, b, c, d, e, g, h)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (a, b, c, d, e, g, h) → GenT m (a, b, c, d, e, g, h)

conformsTo ∷ (a, b, c, d, e, g, h) → TypeSpec (a, b, c, d, e, g, h) → Bool

shrinkWithTypeSpec ∷ TypeSpec (a, b, c, d, e, g, h) → (a, b, c, d, e, g, h) → [(a, b, c, d, e, g, h)]

toPreds ∷ Term (a, b, c, d, e, g, h) → TypeSpec (a, b, c, d, e, g, h) → Pred

cardinalTypeSpec ∷ TypeSpec (a, b, c, d, e, g, h) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (a, b, c, d, e, g, h) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (a, b, c, d, e, g, h) → BinaryShow

monadConformsTo ∷ (a, b, c, d, e, g, h) → TypeSpec (a, b, c, d, e, g, h) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (a, b, c, d, e, g, h) → [(a, b, c, d, e, g, h)] → Specification (a, b, c, d, e, g, h)

guardTypeSpec ∷ [String] → TypeSpec (a, b, c, d, e, g, h) → Specification (a, b, c, d, e, g, h)

prerequisites ∷ Evidence (Prerequisites (a, b, c, d, e, g, h))

(Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Typeable g, Typeable h) ⇒ HasSimpleRep (a, b, c, d, e, g, h)

Instance details

Associated Types

type SimpleRep (a, b, c, d, e, g, h)

type TheSop (a, b, c, d, e, g, h) ∷ [Type]

Methods

toSimpleRep ∷ (a, b, c, d, e, g, h) → SimpleRep (a, b, c, d, e, g, h)

fromSimpleRep ∷ SimpleRep (a, b, c, d, e, g, h) → (a, b, c, d, e, g, h)