Constrained.Syntax

Description

This module contains operations and tranformations on Syntax, Term, Pred, etc. 1) Computing Free Variables 2) Substitution 3) Renaming 4) internal helper functions 5) Syntacic only transformations

Documentation

mkNamed ∷ String → Q Pat

mkNamedExpr ∷ String → Q Exp

var ∷ QuasiQuoter

freeVarNames ∷ ∀ t. HasVariables t ⇒ t → Set Int

newtype FreeVars

Constructors

FreeVars

Fields

unFreeVars ∷ Map Name Int

restrictedTo ∷ FreeVars → Set Name → FreeVars

memberOf ∷ Name → FreeVars → Bool

count ∷ Name → FreeVars → Int

freeVar ∷ Name → FreeVars

singleton ∷ Name → Int → FreeVars

without ∷ Foldable t ⇒ FreeVars → t Name → FreeVars

class HasVariables a where

Minimal complete definition

freeVars

Methods

freeVars ∷ a → FreeVars

freeVarSet ∷ a → Set Name

countOf ∷ Name → a → Int

appearsIn ∷ Name → a → Bool

type Subst = [SubstEntry]

data SubstEntry where

Constructors

(:=) ∷ HasSpec a ⇒ Var a → Term a → SubstEntry

backwardsSubstitution ∷ ∀ a. HasSpec a ⇒ Subst → Term a → Term a

fastInequality ∷ Term a → Term b → Bool

Sound but not complete inequality on terms

substituteTerm ∷ ∀ a. Subst → Term a → Term a

substituteTerm' ∷ ∀ a. Subst → Term a → Writer Any (Term a)

substituteBinder ∷ HasSpec a ⇒ Var a → Term a → Binder b → Binder b

substitutePred ∷ HasSpec a ⇒ Var a → Term a → Pred → Pred

substTerm ∷ Env → Term a → Term a

substBinder ∷ Env → Binder a → Binder a

substPred ∷ Env → Pred → Pred

unBind ∷ a → Binder a → Pred

data Name where

Constructors

Name ∷ HasSpec a ⇒ Var a → Name

fromLits ∷ List Term as → Maybe (List Value as)

fromLit ∷ Term a → Maybe (Value a)

isLit ∷ Term a → Bool

name ∷ String → Term a → Term a

named ∷ String → Term a → Term a

Give a Term a nameHint, if its a Var, and doesn't already have one, otherwise return the Term unchanged.

mkCase ∷ HasSpec (SumOver as) ⇒ Term (SumOver as) → List (Weighted Binder) as → Pred

runCaseOn ∷ SumOver as → List Binder as → (∀ a. HasSpec a ⇒ Var a → a → Pred → r) → r

letSubexpressionElimination ∷ HasSpec Bool ⇒ Pred → Pred

letFloating ∷ Pred → Pred

assertExplain ∷ IsPred p ⇒ NonEmpty String → p → Pred

assert ∷ IsPred p ⇒ p → Pred

forAll ∷ (Forallable t a, HasSpec t, HasSpec a, IsPred p) ⇒ Term t → (Term a → p) → Pred

mkForAll ∷ (Forallable t a, HasSpec t, HasSpec a) ⇒ Term t → Binder a → Pred

exists ∷ ∀ a p. (HasSpec a, IsPred p) ⇒ ((∀ b. Term b → b) → GE a) → (Term a → p) → Pred

unsafeExists ∷ ∀ a p. (HasSpec a, IsPred p) ⇒ (Term a → p) → Pred

letBind ∷ (HasSpec a, IsPred p) ⇒ Term a → (Term a → p) → Pred

reify ∷ (HasSpec a, HasSpec b, IsPred p) ⇒ Term a → (a → b) → (Term b → p) → Pred

assertReified ∷ (HasSpec Bool, HasSpec a) ⇒ Term a → (a → Bool) → Pred

requires (HasSpec Bool)

explanation ∷ NonEmpty String → Pred → Pred

Wrap an Explain around a Pred, unless there is a simpler form.

monitor ∷ ((∀ a. Term a → a) → Property → Property) → Pred

Add QuickCheck monitoring (e.g. collect or counterexample) to a predicate. To use the monitoring in a property call monitorSpec on the Specification containing the monitoring and a value generated from the specification.

reifies ∷ (HasSpec a, HasSpec b) ⇒ Term b → Term a → (a → b) → Pred

dependsOn ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Pred

lit ∷ HasSpec a ⇒ a → Term a

genHint ∷ ∀ t. HasGenHint t ⇒ Hint t → Term t → Pred

envFromPred ∷ Env → Pred → GE Env

runTermE ∷ ∀ a. Env → Term a → Either (NonEmpty String) a

runTerm ∷ MonadGenError m ⇒ Env → Term a → m a

data SolverStage where

Constructors

SolverStage

Fields

∷ HasSpec a

⇒ { stageVar ∷ Var a

, stagePreds ∷ [Pred]

, stageSpec ∷ Specification a

} → SolverStage

data SolverPlan

Constructors

SolverPlan

Fields

solverPlan ∷ [SolverStage]

solverDependencies ∷ Graph Name

isTrueSpec ∷ Specification a → Bool

prettyLinear ∷ [SolverStage] → Doc ann

regularize ∷ HasVariables t ⇒ Var a → t → Var a

regularizeBinder ∷ Binder a → Binder a

regularizeNamesPred ∷ Pred → Pred

regularizeNames ∷ Specification a → Specification a

toCtx ∷ ∀ m v a. (MonadGenError m, HasCallStack, HasSpec a, HasSpec v) ⇒ Var v → Term a → m (Ctx v a)

Construct a Context from a Var and a Term (which we hope has exactly one occurrence of that Var). Runs monadically, and fails with a error message if that property does not hold.

data HiddenClist mode ds v where

Constructors

Hide ∷ CList mode ds v y → HiddenClist mode ds v

toCtxList ∷ ∀ m v as. (All HasSpec as, HasSpec v, MonadGenError m, HasCallStack) ⇒ Var v → List Term as → m (HiddenClist 'Pre as v)

data PolyCList as where

Constructors

Poly ∷ (∀ i j. CList 'Post as i j) → PolyCList as

Orphan instances

Rename Pred

Instance details

Methods

rename ∷ Typeable x ⇒ Var x → Var x → Pred → Pred

Rename (Binder a)

Instance details

Methods

rename ∷ Typeable x ⇒ Var x → Var x → Binder a → Binder a

Rename (Term a)

Instance details

Methods

rename ∷ Typeable x ⇒ Var x → Var x → Term a → Term a

Pretty (Var a)

Instance details

Methods

pretty ∷ Var a → Doc ann

prettyList ∷ [Var a] → Doc ann

Rename (f a) ⇒ Rename (Weighted f a)

Instance details

Methods

rename ∷ Typeable x ⇒ Var x → Var x → Weighted f a → Weighted f a