Constrained.Test

Documentation

prop_conformEmpty ∷ ∀ fn a. HasSpec fn a ⇒ a → Property

prop_propagateSpec ∷ (HasSpec fn a, HasSpec fn b) ⇒ Var b → Term fn a → Spec fn a → Property

prop_mapSpec ∷ (HasSpec fn a, HasSpec fn b) ⇒ fn '[a] b → Spec fn a → Property

prop_genInverse ∷ (HasSpec fn a, HasSpec fn b) ⇒ fn '[a] b → Spec fn a → b → Property

prop_sound ∷ HasSpec fn a ⇒ Spec fn a → Property

prop_complete ∷ HasSpec fn a ⇒ Spec fn a → Property

`prop_complete ps` assumes that ps is satisfiable

genTest ∷ HasSpec BaseFn a ⇒ Spec BaseFn a → IO a

testAll ∷ IO ()

tests ∷ TestTree

numberyTests ∷ TestTree

sizeTests ∷ TestTree

type Numbery a = (Foldy BaseFn a, OrdLike BaseFn a, NumLike BaseFn a, Ord a, Enum a)

data NumberyType where

Constructors

N ∷ (Typeable a, Numbery a) ⇒ Proxy a → NumberyType

testNumberyListSpec ∷ String → (∀ a. Numbery a ⇒ Spec BaseFn [a]) → TestTree

testSpec ∷ HasSpec fn a ⇒ String → Spec fn a → TestTree

leqPair ∷ Spec BaseFn (Int, Int)

setPair ∷ Spec BaseFn (Set (Int, Int))

setSpec ∷ Spec BaseFn (Set Int)

compositionalSpec ∷ Spec BaseFn (Set Int)

simplePairSpec ∷ Spec BaseFn (Int, Int)

trickyCompositional ∷ Spec BaseFn (Int, Int)

newtype NotASet a

Constructors

NotASet (Set a)

emptyListSpec ∷ Spec BaseFn ([Int], NotASet (Either Int Int, Int))

eitherSpec ∷ Spec BaseFn (Either Int Int)

maybeSpec ∷ Spec BaseFn (Set (Maybe Int))

eitherSetSpec ∷ Spec BaseFn (Set (Either Int Int), Set (Either Int Int), Set (Either Int Int))

data Foo

Constructors

Foo Int

Bar Int Int

fooSpec ∷ Spec BaseFn Foo

intSpec ∷ Spec BaseFn (Int, Int)

mapElemSpec ∷ Spec BaseFn (Map Int (Bool, Int))

mapElemKeySpec ∷ Spec BaseFn Int

intRangeSpec ∷ Int → Spec BaseFn Int

testRewriteSpec ∷ Spec BaseFn ((Int, Int), (Int, Int))

mapPairSpec ∷ Spec BaseFn (Map Int Int, Set Int)

mapEmptyDomainSpec ∷ Spec BaseFn (Map Int Int)

setPairSpec ∷ Spec BaseFn (Set Int, Set Int)

fixedSetSpec ∷ Spec BaseFn (Set Int)

setOfPairLetSpec ∷ Spec BaseFn (Set (Int, Int))

setSingletonSpec ∷ Spec BaseFn (Set (Int, Int))

pairSingletonSpec ∷ Spec BaseFn (Int, Int)

eitherSimpleSetSpec ∷ Spec BaseFn (Set (Either Int Int))

forAllAnySpec ∷ Spec BaseFn (Set Int)

weirdSetPairSpec ∷ Spec BaseFn ([Int], Set (Either Int Int))

maybeJustSetSpec ∷ Spec BaseFn (Set (Maybe Int))

notSubsetSpec ∷ Spec BaseFn (Set Int, Set Int)

emptySetSpec ∷ Spec BaseFn (Set Int)

emptyEitherMemberSpec ∷ Spec BaseFn (Set (Either Int Int))

emptyEitherSpec ∷ Spec BaseFn (Set (Either Int Int))

knownDomainMap ∷ Spec BaseFn (Map Int Int)

parallelLet ∷ Spec BaseFn (Int, Int)

letExists ∷ Spec BaseFn (Int, Int)

letExistsLet ∷ Spec BaseFn (Int, Int)

notSubset ∷ Spec BaseFn (Set Int)

unionSized ∷ Spec BaseFn (Set Int)

listSum ∷ Numbery a ⇒ Spec BaseFn [a]

listSumForall ∷ Numbery a ⇒ Spec BaseFn [a]

listSumRange ∷ Numbery a ⇒ Spec BaseFn [a]

listSumRangeUpper ∷ Numbery a ⇒ Spec BaseFn [a]

listSumRangeRange ∷ Numbery a ⇒ Spec BaseFn [a]

listSumElemRange ∷ Numbery a ⇒ Spec BaseFn [a]

listSumPair ∷ Numbery a ⇒ Spec BaseFn [(a, Int)]

listEmpty ∷ Spec BaseFn [Int]

dependencyWeirdness ∷ Spec BaseFn (Int, Int, Int)

foldTrueCases ∷ Spec BaseFn (Either Int Int)

foldSingleCase ∷ Spec BaseFn Int

parallelLetPair ∷ Spec BaseFn (Int, Int)

data Three

Constructors

One

Two

Three

mapSizeConstrained ∷ Spec BaseFn (Map Three Int)

orPair ∷ Spec BaseFn [(Int, Int)]

allZeroTree ∷ Spec BaseFn (BinTree Int)

isBST ∷ Spec BaseFn (BinTree Int)

noChildrenSameTree ∷ Spec BaseFn (BinTree Int)

type RoseFn = Fix (OneofL (RoseTreeFn ': BaseFns))

allZeroRoseTree ∷ Spec RoseFn (RoseTree Int)

noSameChildrenRoseTree ∷ Spec RoseFn (RoseTree Int)

successiveChildren ∷ Spec RoseFn (RoseTree Int)

successiveChildren8 ∷ Spec RoseFn (RoseTree Int)

roseTreeList ∷ Spec RoseFn [RoseTree Int]

pairListError ∷ Spec BaseFn [(Int, Int)]

listMustSizeIssue ∷ Spec BaseFn [Int]

roseTreePairs ∷ Spec RoseFn (RoseTree ([Int], [Int]))

roseTreeMaybe ∷ Spec RoseFn (RoseTree (Maybe (Int, Int)))

badTreeInteraction ∷ Spec RoseFn (RoseTree (Either Int Int))

sizeAddOrSub1 ∷ Spec BaseFn Integer

sizeAddOrSub2 ∷ Spec BaseFn Integer

sizeAddOrSub3 ∷ Spec BaseFn Integer

sizeAddOrSub4 ∷ Spec BaseFn Integer

We expect a negative Integer, so ltSpec tests for that.

sizeAddOrSub5 ∷ Spec BaseFn Integer

listSubSize ∷ Spec BaseFn [Int]

setSubSize ∷ Spec BaseFn (Set Int)

mapSubSize ∷ Spec BaseFn (Map Int Int)

hasSizeList ∷ Spec BaseFn [Int]

hasSizeSet ∷ Spec BaseFn (Set Int)

hasSizeMap ∷ Spec BaseFn (Map Int Int)

noInfinity ∷ Gen (NumSpec Integer)

When we multiply intervals, we get a bounding box, around the possible values. When the intervals have infinities, the bounding box can be very loose. In fact the order in which we multiply intervals with infinities can affect how loose the bounding box is. So ((NegInf, n) * (a, b)) * (c,d) AND (NegInf, n) * ((a, b) * (c,d)) may have different bounding boxes To test the associative laws we must have no infinities, and then the associative law will hold.

plusNegate ∷ NumSpec Integer → NumSpec Integer → Property

commutesNumSpec ∷ NumSpec Integer → NumSpec Integer → Property

assocNumSpec ∷ NumSpec Integer → NumSpec Integer → NumSpec Integer → Property

commuteTimes ∷ NumSpec Integer → NumSpec Integer → Property

assocNumSpecTimes ∷ Gen Property

negNegate ∷ NumSpec Integer → Property

scaleNumSpec ∷ NumSpec Integer → Property

scaleOne ∷ NumSpec Integer → Property

numNumSpecTree ∷ TestTree

runTestSpec ∷ HasSpec BaseFn t ⇒ Spec BaseFn t → IO ()

generateSpec ∷ ∀ fn a. HasSpec fn a ⇒ Spec fn a → IO a

Orphan instances

Arbitrary (NumSpec Integer)

Instance details

Methods

arbitrary ∷ Gen (NumSpec Integer)

shrink ∷ NumSpec Integer → [NumSpec Integer]