Constrained.TheKnot

Description

All the things that are mutually recursive.

Documentation

data SumSpec a b

The Specification for Sums.

Constructors

SumSpecRaw (Maybe String) (Maybe (Int, Int)) (Specification a) (Specification b)

pattern SumSpec ∷ Maybe (Int, Int) → Specification a → Specification b → SumSpec a b

guardSumSpec ∷ ∀ a b. (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ [String] → SumSpec a b → Specification (Sum a b)

combTypeName ∷ Maybe String → Maybe String → Maybe String

type family CountCases a where ...

Equations

CountCases (Sum a b) = 1 + CountCases b

CountCases _ = 1

countCases ∷ ∀ a. KnownNat (CountCases a) ⇒ Int

totalWeight ∷ List (Weighted f) as → Maybe Int

sumType ∷ Maybe String → String

caseBoolSpec ∷ HasSpec a ⇒ Specification Bool → (Bool → Specification a) → Specification a

data BoolW (sym ∷ Symbol) (dom ∷ [Type]) (rng ∷ Type) where

Operations on Bool

Constructors

NotW ∷ BoolW "not_" '[Bool] Bool

OrW ∷ BoolW "or_" '[Bool, Bool] Bool

boolSem ∷ BoolW sym dom rng → FunTy dom rng

not_ ∷ Term Bool → Term Bool

okOr ∷ Bool → Bool → Specification Bool

We have something like (constant ||. HOLE) must evaluate to need. Return a (Specification Bool) for HOLE, that makes that True.

or_ ∷ Term Bool → Term Bool → Term Bool

(==.) ∷ HasSpec a ⇒ Term a → Term a → Term Bool infix 4

toPredsNumSpec ∷ OrdLike n ⇒ Term n → NumSpec n → Pred

(<=.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(<.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(>=.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(>.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

simplifySpec ∷ HasSpec a ⇒ Specification a → Specification a

ifElse ∷ (IsPred p, IsPred q) ⇒ Term Bool → p → q → Pred

If the `Specification Bool` doesn't constrain the boolean you will get a TrueSpec out.

whenTrue ∷ ∀ p. IsPred p ⇒ Term Bool → p → Pred

pinnedBy ∷ ∀ a. HasSpec a ⇒ Var a → Pred → Maybe (Term a)

Is the variable x pinned to some free term in p? (free term meaning that all the variables in the term are free in p).

TODO: complete this with more cases!

optimisePred ∷ Pred → Pred

aggressiveInlining ∷ Pred → Pred

substituteAndSimplifyTerm ∷ Subst → Term a → Term a

Apply a substitution and simplify the resulting term if the substitution changed the term.

simplifyTerm ∷ ∀ a. Term a → Term a

Simplify a Term, if the Term is an App, apply the rewrite rules chosen by the (Logic sym t bs a) instance attached to the function witness f

simplifyPred ∷ Pred → Pred

simplifyPreds ∷ [Pred] → [Pred]

simplifyBinder ∷ Binder a → Binder a

computeSpecSimplified ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Specification a)

Precondition: the Pred defines the `Var a` Runs in GE in order for us to have detailed context on failure.

computeSpec ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Specification a)

Precondition: the `Pred fn` defines the `Var a`. Runs in GE in order for us to have detailed context on failure.

computeSpecBinder ∷ Binder a → GE (Specification a)

computeSpecBinderSimplified ∷ Binder a → GE (Specification a)

sumWeightL ∷ Maybe (Int, Int) → Doc a

sumWeightR ∷ Maybe (Int, Int) → Doc a

caseSpec ∷ ∀ as. HasSpec (SumOver as) ⇒ Maybe String → List (Weighted Specification) as → Specification (SumOver as)

Turn a list of branches into a SumSpec. If all the branches fail return an ErrorSpec. Note the requirement of HasSpec(SumOver).

genFromSpecT ∷ ∀ a m. (HasCallStack, HasSpec a, MonadGenError m) ⇒ Specification a → GenT m a

Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

genFromSpec ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Specification a → Gen a

A version of genFromSpecT that simply errors if the generator fails

genFromSpecWithSeed ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Int → Int → Specification a → a

A version of genFromSpecT that takes a seed and a size and gives you a result

debugSpec ∷ ∀ a. HasSpec a ⇒ Specification a → IO ()

A version of genFromSpecT that runs in the IO monad. Good for debugging.

type DependGraph = Graph Name

dependency ∷ HasVariables t ⇒ Name → t → DependGraph

irreflexiveDependencyOn ∷ ∀ t t'. (HasVariables t, HasVariables t') ⇒ t → t' → DependGraph

noDependencies ∷ HasVariables t ⇒ t → DependGraph

type Hints = DependGraph

respecting ∷ Hints → DependGraph → DependGraph

solvableFrom ∷ Name → Set Name → DependGraph → Bool

computeDependencies ∷ Pred → DependGraph

computeBinderDependencies ∷ Binder a → DependGraph

computeTermDependencies ∷ Term a → DependGraph

computeTermDependencies' ∷ Term a → (DependGraph, Set Name)

shrinkWithSpec ∷ ∀ a. HasSpec a ⇒ Specification a → a → [a]

shrinkFromPreds ∷ HasSpec a ⇒ Pred → Var a → a → [a]

shrinkEnvFromPlan ∷ Env → SolverPlan → [Env]

substStage ∷ Env → SolverStage → SolverStage

normalizeSolverStage ∷ SolverStage → SolverStage

fixupWithSpec ∷ ∀ a. HasSpec a ⇒ Specification a → a → Maybe a

computeHints ∷ [Pred] → Hints

prepareLinearization ∷ Pred → GE SolverPlan

Linearize a predicate, turning it into a list of variables to solve and their defining constraints such that each variable can be solved independently.

flattenPred ∷ Pred → [Pred]

Flatten nested Let, Exists, and And in a `Pred fn`. Let and Exists bound variables become free in the result.

linearize ∷ MonadGenError m ⇒ [Pred] → DependGraph → m [SolverStage]

mergeSolverStage ∷ SolverStage → [SolverStage] → [SolverStage]

Does nothing if the variable is not in the plan already.

prettyPlan ∷ HasSpec a ⇒ Specification a → Doc ann

printPlan ∷ HasSpec a ⇒ Specification a → IO ()

isEmptyPlan ∷ SolverPlan → Bool

stepPlan ∷ MonadGenError m ⇒ Env → SolverPlan → GenT m (Env, SolverPlan)

genFromPreds ∷ ∀ m. MonadGenError m ⇒ Env → Pred → GenT m Env

Generate a satisfying Env for a `p : Pred fn`. The Env contains values for all the free variables in `flattenPred p`.

backPropagation ∷ SolverPlan → SolverPlan

Push as much information we can backwards through the plan.

mapSpec ∷ ∀ s t a b. (Logic s t '[a] b, HasSpec a, HasSpec b) ⇒ t s '[a] b → Specification a → Specification b

Functor like property for Specification, but instead of a Haskell function (a -> b), it takes a function symbol (t c s '[a] b) from a to b. Note, in this context, a function symbol is some constructor of a witnesstype. Eg. ProdFstW, InjRightW, SingletonW, etc. NOT the lifted versions like fst_ singleton_, which construct Terms. We had to wait until here to define this because it depends on Semigroup property of Specification, and Asserting equality

saturatePred ∷ Pred → [Pred]

pairView ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Maybe (Term a, Term b)

cartesian ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ Specification a → Specification b → Specification (Prod a b)

data PairSpec a b

Constructors

Cartesian (Specification a) (Specification b)

prodFst_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term a

prodSnd_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term b

sameFst ∷ Eq a1 ⇒ a1 → [Prod a1 a2] → [a2]

sameSnd ∷ Eq a1 ⇒ a1 → [Prod a2 a1] → [a2]

prod_ ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Term (Prod a b)

Orphan instances

HasSpec Integer

Instance details

Associated Types

type TypeSpec Integer

type Prerequisites Integer

Methods

emptySpec ∷ TypeSpec Integer

combineSpec ∷ TypeSpec Integer → TypeSpec Integer → Specification Integer

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec Integer → GenT m Integer

conformsTo ∷ Integer → TypeSpec Integer → Bool

shrinkWithTypeSpec ∷ TypeSpec Integer → Integer → [Integer]

toPreds ∷ Term Integer → TypeSpec Integer → Pred

cardinalTypeSpec ∷ TypeSpec Integer → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec Integer → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec Integer → BinaryShow

monadConformsTo ∷ Integer → TypeSpec Integer → Writer [String] Bool

typeSpecOpt ∷ TypeSpec Integer → [Integer] → Specification Integer

guardTypeSpec ∷ [String] → TypeSpec Integer → Specification Integer

prerequisites ∷ Evidence (Prerequisites Integer)

HasSpec Bool

Instance details

Associated Types

type TypeSpec Bool

type Prerequisites Bool

Methods

emptySpec ∷ TypeSpec Bool

combineSpec ∷ TypeSpec Bool → TypeSpec Bool → Specification Bool

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec Bool → GenT m Bool

conformsTo ∷ Bool → TypeSpec Bool → Bool

shrinkWithTypeSpec ∷ TypeSpec Bool → Bool → [Bool]

toPreds ∷ Term Bool → TypeSpec Bool → Pred

cardinalTypeSpec ∷ TypeSpec Bool → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec Bool → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec Bool → BinaryShow

monadConformsTo ∷ Bool → TypeSpec Bool → Writer [String] Bool

typeSpecOpt ∷ TypeSpec Bool → [Bool] → Specification Bool

guardTypeSpec ∷ [String] → TypeSpec Bool → Specification Bool

prerequisites ∷ Evidence (Prerequisites Bool)

HasSimpleRep Bool

Instance details

Associated Types

type SimpleRep Bool

type TheSop Bool ∷ [Type]

Methods

toSimpleRep ∷ Bool → SimpleRep Bool

fromSimpleRep ∷ SimpleRep Bool → Bool

OrdLike a ⇒ Logic "<." NumOrdW '[a, a] Bool

Instance details

Methods

info ∷ NumOrdW "<." '[a, a] Bool → String

propagate ∷ Context "<." NumOrdW '[a, a] Bool hole → Specification Bool → Specification hole

rewriteRules ∷ NumOrdW "<." '[a, a] Bool → List Term '[a, a] → Evidence (AppRequires "<." NumOrdW '[a, a] Bool) → Maybe (Term Bool)

mapTypeSpec ∷ ('[a, a] ~ '[a0], Bool ~ b, HasSpec a0, HasSpec b) ⇒ NumOrdW "<." '[a0] b → TypeSpec a0 → Specification b

saturate ∷ NumOrdW "<." '[a, a] Bool → List Term '[a, a] → [Pred]

OrdLike a ⇒ Logic "<=." NumOrdW '[a, a] Bool

Instance details

Methods

info ∷ NumOrdW "<=." '[a, a] Bool → String

propagate ∷ Context "<=." NumOrdW '[a, a] Bool hole → Specification Bool → Specification hole

rewriteRules ∷ NumOrdW "<=." '[a, a] Bool → List Term '[a, a] → Evidence (AppRequires "<=." NumOrdW '[a, a] Bool) → Maybe (Term Bool)

mapTypeSpec ∷ ('[a, a] ~ '[a0], Bool ~ b, HasSpec a0, HasSpec b) ⇒ NumOrdW "<=." '[a0] b → TypeSpec a0 → Specification b

saturate ∷ NumOrdW "<=." '[a, a] Bool → List Term '[a, a] → [Pred]

(HasSpec Bool, Eq a, Typeable a) ⇒ Logic "==." BaseW '[a, a] Bool

Instance details

Methods

info ∷ BaseW "==." '[a, a] Bool → String

propagate ∷ Context "==." BaseW '[a, a] Bool hole → Specification Bool → Specification hole

rewriteRules ∷ BaseW "==." '[a, a] Bool → List Term '[a, a] → Evidence (AppRequires "==." BaseW '[a, a] Bool) → Maybe (Term Bool)

mapTypeSpec ∷ ('[a, a] ~ '[a0], Bool ~ b, HasSpec a0, HasSpec b) ⇒ BaseW "==." '[a0] b → TypeSpec a0 → Specification b

saturate ∷ BaseW "==." '[a, a] Bool → List Term '[a, a] → [Pred]

OrdLike a ⇒ Logic ">." NumOrdW '[a, a] Bool

Instance details

Methods

info ∷ NumOrdW ">." '[a, a] Bool → String

propagate ∷ Context ">." NumOrdW '[a, a] Bool hole → Specification Bool → Specification hole

rewriteRules ∷ NumOrdW ">." '[a, a] Bool → List Term '[a, a] → Evidence (AppRequires ">." NumOrdW '[a, a] Bool) → Maybe (Term Bool)

mapTypeSpec ∷ ('[a, a] ~ '[a0], Bool ~ b, HasSpec a0, HasSpec b) ⇒ NumOrdW ">." '[a0] b → TypeSpec a0 → Specification b

saturate ∷ NumOrdW ">." '[a, a] Bool → List Term '[a, a] → [Pred]

OrdLike a ⇒ Logic ">=." NumOrdW '[a, a] Bool

Instance details

Methods

info ∷ NumOrdW ">=." '[a, a] Bool → String

propagate ∷ Context ">=." NumOrdW '[a, a] Bool hole → Specification Bool → Specification hole

rewriteRules ∷ NumOrdW ">=." '[a, a] Bool → List Term '[a, a] → Evidence (AppRequires ">=." NumOrdW '[a, a] Bool) → Maybe (Term Bool)

mapTypeSpec ∷ ('[a, a] ~ '[a0], Bool ~ b, HasSpec a0, HasSpec b) ⇒ NumOrdW ">=." '[a0] b → TypeSpec a0 → Specification b

saturate ∷ NumOrdW ">=." '[a, a] Bool → List Term '[a, a] → [Pred]

(HasSpec Bool, TypeSpec Bool ~ SumSpec () ()) ⇒ Logic "not_" BoolW '[Bool] Bool

Instance details

Methods

info ∷ BoolW "not_" '[Bool] Bool → String

propagate ∷ Context "not_" BoolW '[Bool] Bool hole → Specification Bool → Specification hole

rewriteRules ∷ BoolW "not_" '[Bool] Bool → List Term '[Bool] → Evidence (AppRequires "not_" BoolW '[Bool] Bool) → Maybe (Term Bool)

mapTypeSpec ∷ ('[Bool] ~ '[a], Bool ~ b, HasSpec a, HasSpec b) ⇒ BoolW "not_" '[a] b → TypeSpec a → Specification b

saturate ∷ BoolW "not_" '[Bool] Bool → List Term '[Bool] → [Pred]

HasSpec Bool ⇒ Logic "or_" BoolW '[Bool, Bool] Bool

Instance details

Methods

info ∷ BoolW "or_" '[Bool, Bool] Bool → String

propagate ∷ Context "or_" BoolW '[Bool, Bool] Bool hole → Specification Bool → Specification hole

rewriteRules ∷ BoolW "or_" '[Bool, Bool] Bool → List Term '[Bool, Bool] → Evidence (AppRequires "or_" BoolW '[Bool, Bool] Bool) → Maybe (Term Bool)

mapTypeSpec ∷ ('[Bool, Bool] ~ '[a], Bool ~ b, HasSpec a, HasSpec b) ⇒ BoolW "or_" '[a] b → TypeSpec a → Specification b

saturate ∷ BoolW "or_" '[Bool, Bool] Bool → List Term '[Bool, Bool] → [Pred]

(HasSpec a, HasSpec b) ⇒ Logic "prodFst_" BaseW '[Prod a b] a

Instance details

Methods

info ∷ BaseW "prodFst_" '[Prod a b] a → String

propagate ∷ Context "prodFst_" BaseW '[Prod a b] a hole → Specification a → Specification hole

rewriteRules ∷ BaseW "prodFst_" '[Prod a b] a → List Term '[Prod a b] → Evidence (AppRequires "prodFst_" BaseW '[Prod a b] a) → Maybe (Term a)

mapTypeSpec ∷ ('[Prod a b] ~ '[a0], a ~ b0, HasSpec a0, HasSpec b0) ⇒ BaseW "prodFst_" '[a0] b0 → TypeSpec a0 → Specification b0

saturate ∷ BaseW "prodFst_" '[Prod a b] Bool → List Term '[Prod a b] → [Pred]

(HasSpec a, HasSpec b) ⇒ Logic "prodSnd_" BaseW '[Prod a b] b

Instance details

Methods

info ∷ BaseW "prodSnd_" '[Prod a b] b → String

propagate ∷ Context "prodSnd_" BaseW '[Prod a b] b hole → Specification b → Specification hole

rewriteRules ∷ BaseW "prodSnd_" '[Prod a b] b → List Term '[Prod a b] → Evidence (AppRequires "prodSnd_" BaseW '[Prod a b] b) → Maybe (Term b)

mapTypeSpec ∷ ('[Prod a b] ~ '[a0], b ~ b0, HasSpec a0, HasSpec b0) ⇒ BaseW "prodSnd_" '[a0] b0 → TypeSpec a0 → Specification b0

saturate ∷ BaseW "prodSnd_" '[Prod a b] Bool → List Term '[Prod a b] → [Pred]

(HasSpec a, HasSpec b) ⇒ Logic "prod_" BaseW '[a, b] (Prod a b)

Instance details

Methods

info ∷ BaseW "prod_" '[a, b] (Prod a b) → String

propagate ∷ Context "prod_" BaseW '[a, b] (Prod a b) hole → Specification (Prod a b) → Specification hole

rewriteRules ∷ BaseW "prod_" '[a, b] (Prod a b) → List Term '[a, b] → Evidence (AppRequires "prod_" BaseW '[a, b] (Prod a b)) → Maybe (Term (Prod a b))

mapTypeSpec ∷ ('[a, b] ~ '[a0], Prod a b ~ b0, HasSpec a0, HasSpec b0) ⇒ BaseW "prod_" '[a0] b0 → TypeSpec a0 → Specification b0

saturate ∷ BaseW "prod_" '[a, b] Bool → List Term '[a, b] → [Pred]

(HasSpec a, Arbitrary (TypeSpec a)) ⇒ Arbitrary (Specification a)

Instance details

Methods

arbitrary ∷ Gen (Specification a)

shrink ∷ Specification a → [Specification a]

(HasSpec a, HasSpec b) ⇒ HasSpec (Prod a b)

Instance details

Associated Types

type TypeSpec (Prod a b)

type Prerequisites (Prod a b)

Methods

emptySpec ∷ TypeSpec (Prod a b)

combineSpec ∷ TypeSpec (Prod a b) → TypeSpec (Prod a b) → Specification (Prod a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Prod a b) → GenT m (Prod a b)

conformsTo ∷ Prod a b → TypeSpec (Prod a b) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Prod a b) → Prod a b → [Prod a b]

toPreds ∷ Term (Prod a b) → TypeSpec (Prod a b) → Pred

cardinalTypeSpec ∷ TypeSpec (Prod a b) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Prod a b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Prod a b) → BinaryShow

monadConformsTo ∷ Prod a b → TypeSpec (Prod a b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Prod a b) → [Prod a b] → Specification (Prod a b)

guardTypeSpec ∷ [String] → TypeSpec (Prod a b) → Specification (Prod a b)

prerequisites ∷ Evidence (Prerequisites (Prod a b))

(HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ HasSpec (Sum a b)
The HasSpec Sum instance

Instance details

Associated Types

type TypeSpec (Sum a b)

type Prerequisites (Sum a b)

Methods

emptySpec ∷ TypeSpec (Sum a b)

combineSpec ∷ TypeSpec (Sum a b) → TypeSpec (Sum a b) → Specification (Sum a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Sum a b) → GenT m (Sum a b)

conformsTo ∷ Sum a b → TypeSpec (Sum a b) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Sum a b) → Sum a b → [Sum a b]

toPreds ∷ Term (Sum a b) → TypeSpec (Sum a b) → Pred

cardinalTypeSpec ∷ TypeSpec (Sum a b) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Sum a b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Sum a b) → BinaryShow

monadConformsTo ∷ Sum a b → TypeSpec (Sum a b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Sum a b) → [Sum a b] → Specification (Sum a b)

guardTypeSpec ∷ [String] → TypeSpec (Sum a b) → Specification (Sum a b)

prerequisites ∷ Evidence (Prerequisites (Sum a b))