Constrained.Univ

Documentation

class FunctionLike fn where

Methods

sem ∷ fn as b → FunTy as b

The semantics of a function is given by sem

class IsMember (fn ∷ [Type] → Type → Type) fnU (path ∷ [PathElem]) where

Evidence that fn exists somewhere in fnU. To make it possible for fnU to consist of a balanced tree of functions we need the path to help GHC disambiguate instances of this class during solving.

Methods

injectFn0 ∷ fn as b → fnU as b

Inject a specific function symbol into a function universe.

extractFn0 ∷ fnU as b → Maybe (fn as b)

type Member fn univ = IsMember fn univ (Path fn univ)

injectFn ∷ ∀ fn fnU as b. Member fn fnU ⇒ fn as b → fnU as b

extractFn ∷ ∀ fn fnU as b. Member fn fnU ⇒ fnU as b → Maybe (fn as b)

data PathElem

Constructors

PFix

PLeft

PRight

newtype Fix (fn ∷ Univ → Univ) (as ∷ [Type]) b

Constructors

Fix (fn (Fix fn) as b)

type Univ = [Type] → Type → Type

data Oneof fn fn' (fnU ∷ [Type] → Type → Type) (as ∷ [Type]) b

Constructors

OneofLeft !(fn fnU as b)

OneofRight !(fn' fnU as b)

type family OneofL as where ...

Build a balanced tree of Oneof from a list of function universes. NOTE: it is important that this is a balanced tree as that reduces the amount of overhead in injectFn and extractFn from `O(n)` to `O(log(n))`. Surprisingly, we've observed this making more than 10% difference in runtime on some generation tasks even though the list as is typically small (< 20 elements).

Equations

OneofL '[t] = t

OneofL (t ': ts) = Insert t (OneofL ts)

type family Insert a t where ...

Equations

Insert a (Oneof fn fn') = Oneof fn' (Insert a fn)

Insert a t = Oneof a t

type family MPath fn fnU ∷ Maybe [PathElem] where ...

Compute the path to look up fn in fnU

Equations

MPath fn fn = 'Just '[]

MPath fn (Oneof fn' fn'' fnU) = Resolve (MPath fn (fn' fnU)) (MPath fn (fn'' fnU))

MPath fn (Fix fn') = AddFix (MPath fn (fn' (Fix fn')))

MPath _ _ = 'Nothing

type family AddFix mpath where ...

Equations

AddFix ('Just path) = 'Just ('PFix ': path)

AddFix 'Nothing = 'Nothing

type family Resolve a b where ...

Equations

Resolve ('Just as) _ = 'Just ('PLeft ': as)

Resolve _ ('Just bs) = 'Just ('PRight ': bs)

Resolve _ _ = 'Nothing

type family FromJust m where ...

Equations

FromJust ('Just a) = a

type family Path fn fn' where ...

Equations

Path fn fn' = FromJust (MPath fn fn')

equalFn ∷ ∀ fn a. (Eq a, Member (EqFn fn) fn) ⇒ fn '[a, a] Bool

data EqFn (fn ∷ [Type] → Type → Type) as b where

Constructors

Equal ∷ Eq a ⇒ EqFn fn '[a, a] Bool

notFn ∷ ∀ fn. Member (BoolFn fn) fn ⇒ fn '[Bool] Bool

orFn ∷ ∀ fn. Member (BoolFn fn) fn ⇒ fn '[Bool, Bool] Bool

data BoolFn (fn ∷ [Type] → Type → Type) as b where

Operations on Bool. One might expect And :: BoolFn fn '[Bool, Bool] Bool But this is problematic because a Term like ( p x &&. q y ) has to solve for x or y first, choose x, so once x is fixed, it might be impossible to solve for y Luckily we can get conjuction by using Preds, in fact using Preds, the x and y can even be the same variable. Or can be problematic too (by using a form of DeMorgan's laws: x && y = not(not x || not y)) but while that is possible, there are many scenarios which can only be specified using Or. If one inadvertantly uses a form of DeMorgan's laws, the worst that can happen is an ErrorSpec is returned, and the user can reformulate using Preds. So Or is incomplete, but sound, when it works.

Constructors

Not ∷ BoolFn fn '[Bool] Bool

Or ∷ BoolFn fn '[Bool, Bool] Bool

data Prod a b

Constructors

Prod

Fields

prodFst ∷ a

prodSnd ∷ b

fstFn ∷ ∀ fn a b. Member (PairFn fn) fn ⇒ fn '[Prod a b] a

sndFn ∷ ∀ fn a b. Member (PairFn fn) fn ⇒ fn '[Prod a b] b

pairFn ∷ ∀ fn a b. Member (PairFn fn) fn ⇒ fn '[a, b] (Prod a b)

data PairFn (fn ∷ [Type] → Type → Type) args res where

Constructors

Pair ∷ PairFn fn '[a, b] (Prod a b)

Fst ∷ PairFn fn '[Prod a b] a

Snd ∷ PairFn fn '[Prod a b] b

type family ProdOver as where ...

Equations

ProdOver '[] = ()

ProdOver '[a] = a

ProdOver (a ': as) = Prod a (ProdOver as)

listToProd ∷ (ProdOver as → r) → List Identity as → r

prodToList ∷ ∀ as. TypeList as ⇒ ProdOver as → List Identity as

appendProd ∷ ∀ xs ys. (TypeList xs, TypeList ys) ⇒ ProdOver xs → ProdOver ys → ProdOver (Append xs ys)

splitProd ∷ ∀ xs ys. (TypeList xs, TypeList ys) ⇒ ProdOver (Append xs ys) → Prod (ProdOver xs) (ProdOver ys)

data Sum a b

Constructors

SumLeft a

SumRight b

injLeftFn ∷ ∀ fn a b. Member (SumFn fn) fn ⇒ fn '[a] (Sum a b)

injRightFn ∷ ∀ fn a b. Member (SumFn fn) fn ⇒ fn '[b] (Sum a b)

data SumFn (fn ∷ [Type] → Type → Type) args res where

Constructors

InjLeft ∷ SumFn fn '[a] (Sum a b)

InjRight ∷ SumFn fn '[b] (Sum a b)

type family SumOver as where ...

Equations

SumOver '[a] = a

SumOver (a ': as) = Sum a (SumOver as)

singletonFn ∷ ∀ fn a. (Member (SetFn fn) fn, Ord a) ⇒ fn '[a] (Set a)

unionFn ∷ ∀ fn a. (Member (SetFn fn) fn, Ord a, Show a, Typeable a) ⇒ fn '[Set a, Set a] (Set a)

subsetFn ∷ ∀ fn a. (Member (SetFn fn) fn, Ord a, Show a, Typeable a) ⇒ fn '[Set a, Set a] Bool

memberFn ∷ ∀ fn a. (Member (SetFn fn) fn, Ord a, Show a, Typeable a) ⇒ fn '[a, Set a] Bool

elemFn ∷ ∀ fn a. (Member (SetFn fn) fn, Eq a, Show a, Typeable a) ⇒ fn '[a, [a]] Bool

disjointFn ∷ ∀ fn a. (Member (SetFn fn) fn, Ord a, Show a, Typeable a) ⇒ fn '[Set a, Set a] Bool

fromListFn ∷ ∀ fn a. (Member (SetFn fn) fn, Ord a) ⇒ fn '[[a]] (Set a)

data SetFn (fn ∷ [Type] → Type → Type) args res where

Constructors

Subset ∷ (Ord a, Show a, Typeable a) ⇒ SetFn fn '[Set a, Set a] Bool

Disjoint ∷ (Ord a, Show a, Typeable a) ⇒ SetFn fn '[Set a, Set a] Bool

Member ∷ (Ord a, Show a, Typeable a) ⇒ SetFn fn '[a, Set a] Bool

Singleton ∷ Ord a ⇒ SetFn fn '[a] (Set a)

Union ∷ (Ord a, Show a, Typeable a) ⇒ SetFn fn '[Set a, Set a] (Set a)

Elem ∷ (Eq a, Show a, Typeable a) ⇒ SetFn fn '[a, [a]] Bool

FromList ∷ Ord a ⇒ SetFn fn '[[a]] (Set a)

data MapFn (fn ∷ [Type] → Type → Type) args res where

Constructors

Dom ∷ Ord k ⇒ MapFn fn '[Map k v] (Set k)

Rng ∷ MapFn fn '[Map k v] [v]

Lookup ∷ Ord k ⇒ MapFn fn '[k, Map k v] (Maybe v)

domFn ∷ ∀ fn k v. (Member (MapFn fn) fn, Ord k) ⇒ fn '[Map k v] (Set k)

rngFn ∷ ∀ fn k v. Member (MapFn fn) fn ⇒ fn '[Map k v] [v]

lookupFn ∷ ∀ fn k v. (Member (MapFn fn) fn, Ord k) ⇒ fn '[k, Map k v] (Maybe v)