Test.Minimal.Model

Documentation

data IntegerSym (dom ∷ [Type]) rng where

Constructors

PlusW ∷ IntegerSym '[Integer, Integer] Integer

MinusW ∷ IntegerSym '[Integer, Integer] Integer

NegateW ∷ IntegerSym '[Integer] Integer

LessOrEqW ∷ IntegerSym '[Integer, Integer] Bool

GreaterOrEqW ∷ IntegerSym '[Integer, Integer] Bool

negateRange ∷ Range → Range

minus ∷ Integer → Integer → Integer

geqSpec ∷ Integer → Spec Integer

leqSpec ∷ Integer → Spec Integer

gtSpec ∷ Integer → Spec Integer

ltSpec ∷ Integer → Spec Integer

(<=.) ∷ Term Integer → Term Integer → Term Bool

(>=.) ∷ Term Integer → Term Integer → Term Bool

(+.) ∷ Term Integer → Term Integer → Term Integer

(-.) ∷ Term Integer → Term Integer → Term Integer

negate_ ∷ Term Integer → Term Integer

data Range

Constructors

Interval (Maybe Integer) (Maybe Integer)

constrainInterval ∷ MonadGenError m ⇒ Maybe Integer → Maybe Integer → Integer → m (Integer, Integer)

data BoolSym (dom ∷ [Type]) rng where

Constructors

NotW ∷ BoolSym '[Bool] Bool

not_ ∷ Term Bool → Term Bool

data SetSym (dom ∷ [Type]) rng where

Constructors

MemberW ∷ (HasSpec a, Ord a) ⇒ SetSym [a, Set a] Bool

SizeW ∷ (HasSpec a, Ord a) ⇒ SetSym '[Set a] Integer

SubsetW ∷ (HasSpec a, Ord a) ⇒ SetSym [Set a, Set a] Bool

setSize ∷ Set a → Integer

size_ ∷ (HasSpec s, Ord s) ⇒ Term (Set s) → Term Integer

subset_ ∷ (HasSpec s, Ord s) ⇒ Term (Set s) → Term (Set s) → Term Bool

member_ ∷ (Ord a, HasSpec a) ⇒ Term a → Term (Set a) → Term Bool

data SetSpec a

Constructors

SetSpec

Fields

setMust ∷ Set a

setAll ∷ Spec a

setCount ∷ Spec Integer

guardSetSpec ∷ (HasSpec a, Ord a) ⇒ SetSpec a → Spec (Set a)

knownUpperBound ∷ Spec Integer → Maybe Integer

pattern Pair ∷ ∀ c. () ⇒ ∀ a b. (c ~ (a, b), HasSpec a, HasSpec b) ⇒ Term a → Term b → Term c

data PairSym (dom ∷ [Type]) rng where

Constructors

FstW ∷ PairSym '[(a, b)] a

SndW ∷ PairSym '[(a, b)] b

PairW ∷ PairSym '[a, b] (a, b)

sameFst ∷ Eq a1 ⇒ a1 → [(a1, a2)] → [a2]

sameSnd ∷ Eq a1 ⇒ a1 → [(a2, a1)] → [a2]

fst_ ∷ (HasSpec a, HasSpec b) ⇒ Term (a, b) → Term a

snd_ ∷ (HasSpec a, HasSpec b) ⇒ Term (a, b) → Term b

pair_ ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Term (a, b)

data PairSpec a b

Constructors

Cartesian (Spec a) (Spec b)

guardPair ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ Spec a → Spec b → Spec (a, b)

data EitherSym (dom ∷ [Type]) rng where

Constructors

LeftW ∷ EitherSym '[a] (Either a b)

RightW ∷ EitherSym '[b] (Either a b)

left_ ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term (Either a b)

right_ ∷ (HasSpec a, HasSpec b) ⇒ Term b → Term (Either a b)

data SumSpec a b

Constructors

SumSpec a b

guardSum ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ Spec a → Spec b → Spec (Either a b)

data ListSym (dom ∷ [Type]) rng where

Constructors

ElemW ∷ Eq a ⇒ ListSym [a, [a]] Bool

LengthW ∷ ListSym '[[a]] Integer

genFromSpecT ∷ ∀ a m. (HasCallStack, HasSpec a, MonadGenError m) ⇒ Spec a → GenT m a

Generalize genFromTypeSpec from `TypeSpec t` to `Spec t` Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

genFromSpec ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Spec a → Gen a

A version of genFromSpecT that simply errors if the generator fails

debugSpec ∷ ∀ a. HasSpec a ⇒ Spec a → IO ()

A version of genFromSpecT that runs in the IO monad. Good for debugging.

genFromPreds ∷ ∀ m. MonadGenError m ⇒ Env → Pred → GenT m Env

Generate a satisfying Env for a `p : Pred fn`. The Env contains values for all the free variables in `flattenPred p`.

simplifySpec ∷ HasSpec a ⇒ Spec a → Spec a

fromGESpec ∷ HasCallStack ⇒ GE (Spec a) → Spec a

Turn GenError into ErrorSpec, and FatalError into error

optimisePred ∷ Pred → Pred

aggressiveInlining ∷ Pred → Pred

propagateSpecM ∷ ∀ v a m. (Monad m, HasSpec v) ⇒ Spec a → m (Ctx v a) → m (Spec v)

Lifts propagateSpec to take a Monadic Ctx

computeSpecSimplified ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Spec a)

Precondition: the Pred defines the `Var a` Runs in GE in order for us to have detailed context on failure.

computeSpec ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Spec a)

Precondition: the `Pred fn` defines the `Var a`. Runs in GE in order for us to have detailed context on failure.

computeSpecBinder ∷ Binder a → GE (Spec a)

computeSpecBinderSimplified ∷ Binder a → GE (Spec a)

substStage ∷ Env → SolverStage → SolverStage

normalizeSolverStage ∷ SolverStage → SolverStage

type Hints = DependGraph

type DependGraph = Graph Name

dependency ∷ HasVariables t ⇒ Name → t → DependGraph

irreflexiveDependencyOn ∷ ∀ t t'. (HasVariables t, HasVariables t') ⇒ t → t' → DependGraph

noDependencies ∷ HasVariables t ⇒ t → DependGraph

respecting ∷ Hints → DependGraph → DependGraph

solvableFrom ∷ Name → Set Name → DependGraph → Bool

computeHints ∷ [Pred] → Hints

saturatePred ∷ Pred → [Pred]

prepareLinearization ∷ Pred → GE SolverPlan

Linearize a predicate, turning it into a list of variables to solve and their defining constraints such that each variable can be solved independently.

flattenPred ∷ Pred → [Pred]

Flatten nested Let, Exists, and And in a `Pred fn`. Let and Exists bound variables become free in the result.

linearize ∷ MonadGenError m ⇒ [Pred] → DependGraph → m [SolverStage]

mergeSolverStage ∷ SolverStage → [SolverStage] → [SolverStage]

Does nothing if the variable is not in the plan already.

prettyPlan ∷ HasSpec a ⇒ Spec a → Doc ann

printPlan ∷ HasSpec a ⇒ Spec a → IO ()

isEmptyPlan ∷ SolverPlan → Bool

stepPlan ∷ MonadGenError m ⇒ Env → SolverPlan → GenT m (Env, SolverPlan)

computeDependencies ∷ Pred → DependGraph

computeBinderDependencies ∷ Binder a → DependGraph

computeTermDependencies ∷ Term a → DependGraph

computeTermDependencies' ∷ Term a → (DependGraph, Set Name)

backPropagation ∷ SolverPlan → SolverPlan

Push as much information we can backwards through the plan.

spec9 ∷ Spec (Set Integer)

Orphan instances

HasSpec Integer

Instance details

Associated Types

type TypeSpec Integer

Methods

anySpec ∷ TypeSpec Integer

combineSpec ∷ TypeSpec Integer → TypeSpec Integer → Spec Integer

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec Integer → GenT m Integer

conformsTo ∷ Integer → TypeSpec Integer → Bool

toPreds ∷ Term Integer → TypeSpec Integer → Pred

guardTypeSpec ∷ TypeSpec Integer → Spec Integer

HasSpec Bool

Instance details

Associated Types

type TypeSpec Bool

Methods

anySpec ∷ TypeSpec Bool

combineSpec ∷ TypeSpec Bool → TypeSpec Bool → Spec Bool

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec Bool → GenT m Bool

conformsTo ∷ Bool → TypeSpec Bool → Bool

toPreds ∷ Term Bool → TypeSpec Bool → Pred

guardTypeSpec ∷ TypeSpec Bool → Spec Bool

(Container (Set a) a, Ord a, HasSpec a) ⇒ HasSpec (Set a)

Instance details

Associated Types

type TypeSpec (Set a)

Methods

anySpec ∷ TypeSpec (Set a)

combineSpec ∷ TypeSpec (Set a) → TypeSpec (Set a) → Spec (Set a)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Set a) → GenT m (Set a)

conformsTo ∷ Set a → TypeSpec (Set a) → Bool

toPreds ∷ Term (Set a) → TypeSpec (Set a) → Pred

guardTypeSpec ∷ TypeSpec (Set a) → Spec (Set a)

Ord s ⇒ Container (Set s) s

Instance details

Methods

fromForAllSpec ∷ Spec s → Spec (Set s)

forAllToList ∷ Set s → [s]

(HasSpec a, HasSpec b) ⇒ HasSpec (Either a b)

Instance details

Associated Types

type TypeSpec (Either a b)

Methods

anySpec ∷ TypeSpec (Either a b)

combineSpec ∷ TypeSpec (Either a b) → TypeSpec (Either a b) → Spec (Either a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Either a b) → GenT m (Either a b)

conformsTo ∷ Either a b → TypeSpec (Either a b) → Bool

toPreds ∷ Term (Either a b) → TypeSpec (Either a b) → Pred

guardTypeSpec ∷ TypeSpec (Either a b) → Spec (Either a b)

(HasSpec a, HasSpec b) ⇒ HasSpec (a, b)

Instance details

Associated Types

type TypeSpec (a, b)

Methods

anySpec ∷ TypeSpec (a, b)

combineSpec ∷ TypeSpec (a, b) → TypeSpec (a, b) → Spec (a, b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (a, b) → GenT m (a, b)

conformsTo ∷ (a, b) → TypeSpec (a, b) → Bool

toPreds ∷ Term (a, b) → TypeSpec (a, b) → Pred

guardTypeSpec ∷ TypeSpec (a, b) → Spec (a, b)