Test.Minimal.Syntax

Documentation

forAll ∷ (Container t a, HasSpec t, HasSpec a) ⇒ Term t → (Term a → Pred) → Pred

mkForAll ∷ (Container t a, HasSpec t, HasSpec a) ⇒ Term t → Binder a → Pred

exists ∷ ∀ a. HasSpec a ⇒ ((∀ b. Term b → b) → GE a) → (Term a → Pred) → Pred

unsafeExists ∷ ∀ a. HasSpec a ⇒ (Term a → Pred) → Pred

notMemberSpec ∷ ∀ a f. (HasSpec a, Foldable f) ⇒ f a → Spec a

isErrorLike ∷ ∀ a. Spec a → Bool

hasError ∷ ∀ a. Spec a → Maybe (NonEmpty String)

handleErrors ∷ Spec a → Spec b → (Spec a → Spec b → Spec c) → Spec c

Given two Spec, return an ErrorSpec if one or more is an ErrorSpec If neither is an ErrorSpec apply the continuation f

addToErrorSpec ∷ NonEmpty String → Spec a → Spec a

Add the explanations, if it's an ErrorSpec, else drop them

memberSpecList ∷ [a] → NonEmpty String → Spec a

return a MemberSpec or ans ErrorSpec depending on if xs the null list or not

satisfies ∷ ∀ a. HasSpec a ⇒ Term a → Spec a → Pred

runTermE ∷ ∀ a. Env → Term a → Either (NonEmpty String) a

runTerm ∷ MonadGenError m ⇒ Env → Term a → m a

conformsToSpec ∷ ∀ a. HasSpec a ⇒ a → Spec a → Bool

checkPredE ∷ Env → NonEmpty String → Pred → Maybe (NonEmpty String)

runCaseOn ∷ Either a b → Binder a → Binder b → (∀ x. HasSpec x ⇒ Var x → x → Pred → r) → r

checkPredsE ∷ NonEmpty String → Env → [Pred] → Maybe (NonEmpty String)

Like checkPred, But it takes [Pred] rather than a single Pred, and it builds a much more involved explanation if it fails. Does the Pred evaluate to True under the given Env? If it doesn't, an involved explanation appears in the (Just message) If it does, then it returns Nothing

data Name where

Constructors

Name ∷ HasSpec a ⇒ Var a → Name

freeVarNames ∷ ∀ t. HasVariables t ⇒ t → Set Int

newtype FreeVars

Constructors

FreeVars

Fields

unFreeVars ∷ Map Name Int

restrictedTo ∷ FreeVars → Set Name → FreeVars

memberOf ∷ Name → FreeVars → Bool

count ∷ Name → FreeVars → Int

freeVar ∷ Name → FreeVars

singleton ∷ Name → Int → FreeVars

without ∷ Foldable t ⇒ FreeVars → t Name → FreeVars

class HasVariables a where

Minimal complete definition

freeVars

Methods

freeVars ∷ a → FreeVars

freeVarSet ∷ a → Set Name

countOf ∷ Name → a → Int

appearsIn ∷ Name → a → Bool

fromLits ∷ List Term as → Maybe (List Value as)

fromLit ∷ Term a → Maybe (Value a)

isLit ∷ Term a → Bool

type Subst = [SubstEntry]

data SubstEntry where

Constructors

(:=) ∷ HasSpec a ⇒ Var a → Term a → SubstEntry

backwardsSubstitution ∷ ∀ a. HasSpec a ⇒ Subst → Term a → Term a

fastInequality ∷ Term a → Term b → Bool

Sound but not complete inequality on terms

substituteTerm ∷ ∀ a. Subst → Term a → Term a

substituteTerm' ∷ ∀ a. Subst → Term a → Writer Any (Term a)

substituteBinder ∷ HasSpec a ⇒ Var a → Term a → Binder b → Binder b

substitutePred ∷ HasSpec a ⇒ Var a → Term a → Pred → Pred

substTerm ∷ Env → Term a → Term a

substBinder ∷ Env → Binder a → Binder a

substPred ∷ Env → Pred → Pred

substSpec ∷ Env → Spec a → Spec a

substSolverStage ∷ Env → SolverStage → SolverStage

substPlan ∷ Env → SolverPlan → SolverPlan

unBind ∷ a → Binder a → Pred

regularize ∷ HasVariables t ⇒ Var a → t → Var a

regularizeBinder ∷ Binder a → Binder a

regularizeNamesPred ∷ Pred → Pred

regularizeNames ∷ Spec a → Spec a

substituteAndSimplifyTerm ∷ Subst → Term a → Term a

Apply a substitution and simplify the resulting term if the substitution changed the term.

simplifyTerm ∷ ∀ a. Term a → Term a

Simplify a Term, if the Term is an App, apply the rewrite rules chosen by the (Semantics t) instance attached to the App to the function witness f

simplifyPred ∷ Pred → Pred

mkCase ∷ HasSpec (Either a b) ⇒ Term (Either a b) → Binder a → Binder b → Pred

simplifyPreds ∷ [Pred] → [Pred]

simplifyBinder ∷ Binder a → Binder a

toPred ∷ Bool → Pred

letFloating ∷ Pred → Pred

letSubexpressionElimination ∷ HasSpec Bool ⇒ Pred → Pred

data SolverStage where

Constructors

SolverStage

Fields

∷ HasSpec a

⇒ { stageVar ∷ Var a

, stagePreds ∷ [Pred]

, stageSpec ∷ Spec a

} → SolverStage

data SolverPlan

Constructors

SolverPlan

Fields

solverPlan ∷ [SolverStage]

solverDependencies ∷ Graph Name

isTrueSpec ∷ Spec a → Bool

prettyLinear ∷ [SolverStage] → Doc ann

data EqSym ∷ [Type] → Type → Type where

Constructors

EqualW ∷ (Eq a, HasSpec a) ⇒ EqSym '[a, a] Bool

equalSpec ∷ a → Spec a

notEqualSpec ∷ ∀ a. HasSpec a ⇒ a → Spec a

caseBoolSpec ∷ (HasSpec Bool, HasSpec a) ⇒ Spec Bool → (Bool → Spec a) → Spec a

(==.) ∷ (HasSpec Bool, HasSpec a) ⇒ Term a → Term a → Term Bool infix 4

getWitness ∷ ∀ t t' d r. (AppRequires t d r, Typeable t') ⇒ t d r → Maybe (t' d r)

pattern Equal ∷ ∀ b. () ⇒ ∀ a. (b ~ Bool, Eq a, HasSpec a) ⇒ Term a → Term a → Term b

pinnedBy ∷ ∀ a. HasSpec a ⇒ Var a → Pred → Maybe (Term a)

Is the variable x pinned to some free term in p? (free term meaning that all the variables in the term are free in p).

TODO: complete this with more cases!

Orphan instances

Monoid Pred

Instance details

Methods

mempty ∷ Pred

mappend ∷ Pred → Pred → Pred

mconcat ∷ [Pred] → Pred

Semigroup Pred

Instance details

Methods

(<>) ∷ Pred → Pred → Pred

sconcat ∷ NonEmpty Pred → Pred

stimes ∷ Integral b ⇒ b → Pred → Pred

Rename Pred

Instance details

Methods

rename ∷ Typeable x ⇒ Var x → Var x → Pred → Pred

HasSpec a ⇒ Monoid (Spec a)

Instance details

Methods

mempty ∷ Spec a

mappend ∷ Spec a → Spec a → Spec a

mconcat ∷ [Spec a] → Spec a

HasSpec a ⇒ Semigroup (Spec a)

Instance details

Methods

(<>) ∷ Spec a → Spec a → Spec a

sconcat ∷ NonEmpty (Spec a) → Spec a

stimes ∷ Integral b ⇒ b → Spec a → Spec a

Rename (Binder a)

Instance details

Methods

rename ∷ Typeable x ⇒ Var x → Var x → Binder a → Binder a

Rename (Term a)

Instance details

Methods

rename ∷ Typeable x ⇒ Var x → Var x → Term a → Term a

Pretty (Var a)

Instance details

Methods

pretty ∷ Var a → Doc ann

prettyList ∷ [Var a] → Doc ann