Constrained

Description

This module exports the user-facing interface for the library. It is supposed to contain everything you need to write constraints and implement HasSpec and HasSimpleRep instances.

Documentation

data Spec fn a where

A `Spec fn a` denotes a set of as

Constructors

MemberSpec
Elements of a known set

Fields

∷ OrdSet a

It must be an element of this set

→ Spec fn a

ErrorSpec ∷ [String] → Spec fn a
The empty set

SuspendedSpec
The set described by some predicates over the bound variable.

TODO: possibly we want to use a Binder here

Fields

∷ HasSpec fn a

⇒ Var a

This variable ranges over values denoted by the spec

→ Pred fn

And the variable is subject to these constraints

→ Spec fn a

TypeSpec
A type-specific spec

Fields

∷ HasSpec fn a

⇒ TypeSpec fn a

→ OrdSet a

It can't be any of the elements of this set

→ Spec fn a

TrueSpec ∷ Spec fn a
Anything

data Pred (fn ∷ [Type] → Type → Type) where

Constructors

Block ∷ [Pred fn] → Pred fn

Exists

Fields

∷ ((∀ b. Term fn b → b) → GE a)

Constructive recovery function for checking existential quantification

→ Binder fn a

→ Pred fn

Subst ∷ HasSpec fn a ⇒ Var a → Term fn a → Pred fn → Pred fn

Let ∷ Term fn a → Binder fn a → Pred fn

Assert ∷ BaseUniverse fn ⇒ [String] → Term fn Bool → Pred fn

Reify

Fields

∷ (HasSpec fn a, HasSpec fn b)

⇒ Term fn a

→ (a → b)

Recover a useable value from the a term.

→ Binder fn b

Everything in here depends on the a term.

→ Pred fn

DependsOn ∷ (HasSpec fn a, HasSpec fn b) ⇒ Term fn a → Term fn b → Pred fn

ForAll ∷ (Forallable t a, HasSpec fn t, HasSpec fn a) ⇒ Term fn t → Binder fn a → Pred fn

Case

Fields

∷ HasSpec fn (SumOver as)

⇒ Term fn (SumOver as)

→ List (Binder fn) as

Each branch of the type is bound with only one variable because as are types. Constructors with multiple arguments are encoded with ProdOver (c.f. Univ).

→ Pred fn

GenHint ∷ HasGenHint fn a ⇒ Hint a → Term fn a → Pred fn

TruePred ∷ Pred fn

FalsePred ∷ [String] → Pred fn

data Term (fn ∷ [Type] → Type → Type) a

Typed first order terms with function symbols from fn.

data Binder fn a where

Constructors

(:->) ∷ HasSpec fn a ⇒ Var a → Pred fn → Binder fn a

class (Typeable a, Eq a, Show a, Show (TypeSpec fn a), BaseUniverse fn) ⇒ HasSpec fn a where

This class provides the interface that allows you to extend the language to handle a new type. In the case of types that have a generic representation (via HasSimpleRep) that already has an instance of HasSpec it is sufficient to provide an empty instance. However, for types that are used together with specific functions in the function universe fn it may be necessary to provide a specific implementation of HasSpec. This is typically necessary when the Spec for the generic representation does not permit an implementation of propagateSpecFun for some function.

The basic types provided in the language, Set, `[]`, Map, Int, Word64, (,), Either, etc. have instances of this class (technically (,) and Either have instances derived from the instances for their generic Prod and Sum implementations).

Minimal complete definition

Nothing

Associated Types

type TypeSpec (fn ∷ [Type] → Type → Type) a

The `TypeSpec fn a` is the type-specific `Spec fn a`.

type TypeSpec fn a = TypeSpec fn (SimpleRep a)

type Prerequisites fn a ∷ Constraint

Prerequisites for the instance that are sometimes necessary when working with e.g. Specs or functions in the universe.

type Prerequisites fn a = ()

Methods

emptySpec ∷ TypeSpec fn a

default emptySpec ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ TypeSpec fn a

combineSpec ∷ TypeSpec fn a → TypeSpec fn a → Spec fn a

default combineSpec ∷ (HasSimpleRep a, HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ TypeSpec fn a → TypeSpec fn a → Spec fn a

genFromTypeSpec ∷ (HasCallStack, MonadGenError m) ⇒ TypeSpec fn a → GenT m a

Generate a value that satisfies the Spec. The key property for this generator is soundness: ∀ a ∈ genFromTypeSpec spec. a conformsTo spec

default genFromTypeSpec ∷ (HasSimpleRep a, HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ (HasCallStack, MonadGenError m) ⇒ TypeSpec fn a → GenT m a

conformsTo ∷ HasCallStack ⇒ a → TypeSpec fn a → Bool

Check conformance to the spec.

default conformsTo ∷ (HasSimpleRep a, HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ HasCallStack ⇒ a → TypeSpec fn a → Bool

toPreds ∷ Term fn a → TypeSpec fn a → Pred fn

Convert a spec to predicates: The key property here is: ∀ a. a conformsTo spec == a conformsTo constrained (t -> toPreds t spec)

default toPreds ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ Term fn a → TypeSpec fn a → Pred fn

prerequisites ∷ Evidence (Prerequisites fn a)

Materialize the Prerequisites dictionary. It should not be necessary to implement this function manually.

default prerequisites ∷ Prerequisites fn a ⇒ Evidence (Prerequisites fn a)

class HasSimpleRep a where

Minimal complete definition

Nothing

Associated Types

type SimpleRep a

type SimpleRep a = SOP (TheSop a)

type TheSop a ∷ [Type]

type TheSop a = SOPOf (Rep a)

Methods

toSimpleRep ∷ a → SimpleRep a

default toSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ a → SimpleRep a

fromSimpleRep ∷ SimpleRep a → a

default fromSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ SimpleRep a → a

class (Num a, HasSpec fn a) ⇒ NumLike fn a where

Minimal complete definition

Nothing

Methods

subtractSpec ∷ a → TypeSpec fn a → Spec fn a

default subtractSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, NumLike fn (SimpleRep a)) ⇒ a → TypeSpec fn a → Spec fn a

negateSpec ∷ TypeSpec fn a → Spec fn a

default negateSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, NumLike fn (SimpleRep a)) ⇒ TypeSpec fn a → Spec fn a

safeSubtract ∷ a → a → Maybe a

default safeSubtract ∷ (HasSimpleRep a, NumLike fn (SimpleRep a)) ⇒ a → a → Maybe a

class HasSpec fn a ⇒ OrdLike fn a where

Minimal complete definition

Nothing

Methods

leqSpec ∷ a → Spec fn a

default leqSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Spec fn a

ltSpec ∷ a → Spec fn a

default ltSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Spec fn a

geqSpec ∷ a → Spec fn a

default geqSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Spec fn a

gtSpec ∷ a → Spec fn a

default gtSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Spec fn a

class Forallable t e | t → e where

Minimal complete definition

Nothing

Methods

forAllSpec ∷ (HasSpec fn t, HasSpec fn e, BaseUniverse fn) ⇒ Spec fn e → Spec fn t

default forAllSpec ∷ (HasSpec fn t, HasSpec fn e, HasSimpleRep t, TypeSpec fn t ~ TypeSpec fn (SimpleRep t), Forallable (SimpleRep t) e, HasSpec fn (SimpleRep t)) ⇒ Spec fn e → Spec fn t

forAllToList ∷ t → [e]

default forAllToList ∷ (HasSimpleRep t, Forallable (SimpleRep t) e) ⇒ t → [e]

class HasSpec fn a ⇒ Foldy fn a where

Methods

genList ∷ (BaseUniverse fn, MonadGenError m) ⇒ Spec fn a → Spec fn a → GenT m [a]

theAddFn ∷ fn '[a, a] a

theZero ∷ a

class FunctionLike fn where

Methods

sem ∷ fn as b → FunTy as b

The semantics of a function is given by sem

class (∀ as b. Show (f as b), ∀ as b. Eq (f as b), Typeable f, FunctionLike f) ⇒ Functions f fn where

Minimal complete definition

propagateSpecFun, mapTypeSpec

Methods

propagateSpecFun ∷ (TypeList as, Typeable as, HasSpec fn a, All (HasSpec fn) as) ⇒ f as b → ListCtx Value as (HOLE a) → Spec fn b → Spec fn a

rewriteRules ∷ (TypeList as, Typeable as, HasSpec fn b, All (HasSpec fn) as) ⇒ f as b → [RewriteRule fn b]

mapTypeSpec ∷ (HasSpec fn a, HasSpec fn b) ⇒ f '[a] b → TypeSpec fn a → Spec fn b

data HOLE a b where

This is used together with ListCtx to form just the arguments to `f vs Ctx vs'` - replacing Ctx with HOLE - to provide to propagateSpecFun.

Constructors

HOLE ∷ HOLE a a

type BaseUniverse fn = (Functions fn fn, Member (EqFn fn) fn, Member (SetFn fn) fn, Member (BoolFn fn) fn, Member (PairFn fn) fn, Member (IntFn fn) fn, Member (OrdFn fn) fn, Member (GenericsFn fn) fn, Member (ListFn fn) fn, Member (SumFn fn) fn, Member (MapFn fn) fn, Member (FunFn fn) fn)

A minimal Universe contains functions for a bunch of different things.

class Member (fn ∷ [Type] → Type → Type) fnU

Minimal complete definition

injectFn

type BaseFn = Fix (OneofL BaseFns)

type BaseFns = '[ListFn, SumFn, MapFn, SetFn, FunFn, GenericsFn, PairFn, IntFn, BoolFn, EqFn, OrdFn]

data Fix fn (as ∷ [Type]) b

type family OneofL as where ...

Equations

OneofL '[t] = t

OneofL (t ': ts) = Oneof t (OneofL ts)

type IsPred p fn = (PredLike p, UnivConstr p fn)

type IsNormalType a = (Cases a ~ '[a], Args a ~ '[a], IsProd a)

data Value a where

Constructors

Value ∷ Show a ⇒ !a → Value a

type family SOP constrs where ...

Turn a list from SOPOf into a Sum over Prod representation.

Equations

SOP '[c ::: prod] = ProdOver prod

SOP ((c ::: prod) ': constrs) = Sum (ProdOver prod) (SOP constrs)

data (c ∷ Symbol) ::: (ts ∷ [Type])

A constructor name with the types of its arguments

data MapFn (fn ∷ [Type] → Type → Type) args res

data FunFn fn args res

data PairSpec fn a b

Constructors

Cartesian (Spec fn a) (Spec fn b)

data NumSpec n

Constructors

NumSpecInterval (Maybe n) (Maybe n)

data MapSpec fn k v

Constructors

MapSpec

Fields

mapSpecMustKeys ∷ Set k

mapSpecMustValues ∷ [v]

mapSpecSize ∷ Spec fn Int

mapSpecElem ∷ Spec fn (k, v)

mapSpecFold ∷ FoldSpec fn v

data FoldSpec (fn ∷ [Type] → Type → Type) a where

Constructors

NoFold ∷ FoldSpec fn a

FoldSpec ∷ ∀ b fn a. (HasSpec fn a, HasSpec fn b, Foldy fn b, Member (ListFn fn) fn, BaseUniverse fn) ⇒ fn '[a] b → Spec fn b → FoldSpec fn a

genFromSpec ∷ ∀ fn a m. (HasCallStack, HasSpec fn a, MonadGenError m) ⇒ Spec fn a → GenT m a

Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

genFromSpec_ ∷ ∀ fn a. (HasCallStack, HasSpec fn a) ⇒ Spec fn a → Gen a

A version of genFromSpec that simply errors if the generator fails

conformsToSpec ∷ ∀ fn a. HasSpec fn a ⇒ a → Spec fn a → Bool

conformsToSpecProp ∷ ∀ fn a. HasSpec fn a ⇒ a → Spec fn a → Property

typeSpec ∷ HasSpec fn a ⇒ TypeSpec fn a → Spec fn a

con ∷ ∀ c a r fn. (SimpleRep a ~ SOP (TheSop a), TypeSpec fn a ~ TypeSpec fn (SOP (TheSop a)), TypeList (ConstrOf c (TheSop a)), HasSpec fn a, HasSimpleRep a, r ~ FunTy (MapList (Term fn) (ConstrOf c (TheSop a))) (Term fn a), ResultType r ~ Term fn a, SOPTerm c fn (TheSop a), ConstrTerm fn (ConstrOf c (TheSop a))) ⇒ r

caseBoolSpec ∷ HasSpec fn a ⇒ Spec fn Bool → (Bool → Spec fn a) → Spec fn a

If the `Spec fn Bool` doesn't constrain the boolean you will get a TrueSpec out.

constrained ∷ ∀ a fn p. (IsPred p fn, HasSpec fn a) ⇒ (Term fn a → p) → Spec fn a

constrained' ∷ ∀ fn a p. (Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSpec fn (SimpleRep a), HasSimpleRep a, All (HasSpec fn) (Args (SimpleRep a)), IsProd (SimpleRep a), HasSpec fn a, IsPred p fn) ⇒ FunTy (MapList (Term fn) (Args (SimpleRep a))) p → Spec fn a

satisfies ∷ ∀ fn a. HasSpec fn a ⇒ Term fn a → Spec fn a → Pred fn

letBind ∷ (HasSpec fn a, IsPred p fn) ⇒ Term fn a → (Term fn a → p) → Pred fn

match ∷ ∀ fn p a. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, Cases (SimpleRep a) ~ '[SimpleRep a], SimpleRep a ~ SumOver (Cases (SimpleRep a)), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), IsProd (SimpleRep a), All (HasSpec fn) (Args (SimpleRep a)), IsPred p fn) ⇒ Term fn a → FunTy (MapList (Term fn) (Args (SimpleRep a))) p → Pred fn

assert ∷ (BaseUniverse fn, IsPred p fn) ⇒ p → Pred fn

assertExplain ∷ (BaseUniverse fn, IsPred p fn) ⇒ [String] → p → Pred fn

caseOn ∷ ∀ fn a. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a), SimpleRep a ~ SumOver (Cases (SimpleRep a)), TypeList (Cases (SimpleRep a))) ⇒ Term fn a → FunTy (MapList (Binder fn) (Cases (SimpleRep a))) (Pred fn)

branch ∷ ∀ fn p a. (HasSpec fn a, All (HasSpec fn) (Args a), IsPred p fn, IsProd a) ⇒ FunTy (MapList (Term fn) (Args a)) p → Binder fn a

reify ∷ (HasSpec fn a, HasSpec fn b, IsPred p fn) ⇒ Term fn a → (a → b) → (Term fn b → p) → Pred fn

dependsOn ∷ (HasSpec fn a, HasSpec fn b) ⇒ Term fn a → Term fn b → Pred fn

forAll ∷ (Forallable t a, HasSpec fn t, HasSpec fn a, IsPred p fn) ⇒ Term fn t → (Term fn a → p) → Pred fn

forAll' ∷ ∀ fn t a p. (Forallable t a, Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSpec fn t, HasSpec fn (SimpleRep a), HasSimpleRep a, All (HasSpec fn) (Args (SimpleRep a)), IsPred p fn, IsProd (SimpleRep a), HasSpec fn a) ⇒ Term fn t → FunTy (MapList (Term fn) (Args (SimpleRep a))) p → Pred fn

exists ∷ ∀ a p fn. (HasSpec fn a, IsPred p fn) ⇒ ((∀ b. Term fn b → b) → GE a) → (Term fn a → p) → Pred fn

unsafeExists ∷ ∀ a p fn. (HasSpec fn a, IsPred p fn) ⇒ (Term fn a → p) → Pred fn

fst_ ∷ ∀ fn a b. (HasSpec fn a, HasSpec fn b) ⇒ Term fn (a, b) → Term fn a

snd_ ∷ ∀ fn a b. (HasSpec fn a, HasSpec fn b) ⇒ Term fn (a, b) → Term fn b

pair_ ∷ ∀ fn a b. (HasSpec fn a, HasSpec fn b) ⇒ Term fn a → Term fn b → Term fn (a, b)

(<=.) ∷ (Ord a, OrdLike fn a) ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(<.) ∷ (Ord a, OrdLike fn a) ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(==.) ∷ HasSpec fn a ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(/=.) ∷ HasSpec fn a ⇒ Term fn a → Term fn a → Term fn Bool infix 4

member_ ∷ ∀ a fn. (HasSpec fn a, Ord a) ⇒ Term fn a → Term fn (Set a) → Term fn Bool

subset_ ∷ (HasSpec fn a, Ord a) ⇒ Term fn (Set a) → Term fn (Set a) → Term fn Bool

disjoint_ ∷ (HasSpec fn a, Ord a) ⇒ Term fn (Set a) → Term fn (Set a) → Term fn Bool

singleton_ ∷ (HasSpec fn a, Ord a) ⇒ Term fn a → Term fn (Set a)

elem_ ∷ ∀ a fn. HasSpec fn a ⇒ Term fn a → Term fn [a] → Term fn Bool

not_ ∷ BaseUniverse fn ⇒ Term fn Bool → Term fn Bool

foldMap_ ∷ (BaseUniverse fn, Foldy fn b, HasSpec fn a) ⇒ fn '[a] b → Term fn [a] → Term fn b

sum_ ∷ (BaseUniverse fn, Member (FunFn fn) fn, Foldy fn a) ⇒ Term fn [a] → Term fn a

size_ ∷ ∀ a fn. (HasSpec fn (Set a), Ord a) ⇒ Term fn (Set a) → Term fn Int

rng_ ∷ (HasSpec fn k, HasSpec fn v, Ord k) ⇒ Term fn (Map k v) → Term fn [v]

dom_ ∷ (HasSpec fn (Map k v), HasSpec fn k, Ord k) ⇒ Term fn (Map k v) → Term fn (Set k)

left_ ∷ (HasSpec fn a, HasSpec fn b) ⇒ Term fn a → Term fn (Either a b)

right_ ∷ (HasSpec fn a, HasSpec fn b) ⇒ Term fn b → Term fn (Either a b)

toGeneric_ ∷ ∀ a fn. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Term fn a → Term fn (SimpleRep a)

fromGeneric_ ∷ ∀ a fn. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Term fn (SimpleRep a) → Term fn a

injectFn ∷ Member fn fnU ⇒ fn as b → fnU as b

Inject a specific function symbol into a function universe.

app ∷ (HasSpec fn b, Typeable as, TypeList as, All (HasSpec fn) as) ⇒ fn as b → FunTy (MapList (Term fn) as) (Term fn b)

lit ∷ Show a ⇒ a → Term fn a

composeFn ∷ (Member (FunFn fn) fn, HasSpec fn b, Show (fn '[a] b), Show (fn '[b] c), Eq (fn '[a] b), Eq (fn '[b] c)) ⇒ fn '[b] c → fn '[a] b → fn '[a] c

fstFn ∷ ∀ fn a b. Member (PairFn fn) fn ⇒ fn '[Prod a b] a

sndFn ∷ ∀ fn a b. Member (PairFn fn) fn ⇒ fn '[Prod a b] b

addFn ∷ ∀ fn a. NumLike fn a ⇒ fn '[a, a] a

toGenericFn ∷ ∀ fn a. (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ fn '[a] (SimpleRep a)

toSimpleRepSpec ∷ ∀ a fn. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Spec fn a → Spec fn (SimpleRep a)

fromSimpleRepSpec ∷ ∀ a fn. (HasSpec fn a, HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Spec fn (SimpleRep a) → Spec fn a

algebra ∷ SOPLike constrs r ⇒ SOP constrs → ALG constrs r

Run a SOP

inject ∷ ∀ c constrs. Inject c constrs (SOP constrs) ⇒ FunTy (ConstrOf c constrs) (SOP constrs)

toPred ∷ (BaseUniverse fn, IsPred p fn) ⇒ p → Pred fn

module Constrained.List

module Constrained.GenT