Constrained.API

Documentation

type GenericallyInstantiated a = (AssertComputes (SimpleRep a) (((Text "Trying to use a generic instantiation of " :<>: ShowType a) :<>: Text ", likely in a HasSpec instance.") :$$: Text "However, the type has no definition of SimpleRep, likely because of a missing instance of HasSimpleRep."), HasSimpleRep a, HasSpec (SimpleRep a), TypeSpec a ~ TypeSpec (SimpleRep a))

class LogicRequires t ⇒ Logic t where

Logical operations are one that support reasoning about how a function symbol relates to logical properties, that we call Specification's

Minimal complete definition

propagate | propagateTypeSpec, propagateMemberSpec

Methods

propagateTypeSpec ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → TypeSpec b → [b] → Specification a

propagateMemberSpec ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → NonEmpty b → Specification a

propagate ∷ (AppRequires t as b, HasSpec a) ⇒ t as b → ListCtx Value as (HOLE a) → Specification b → Specification a

rewriteRules ∷ (TypeList dom, Typeable dom, HasSpec rng, All HasSpec dom) ⇒ t dom rng → List Term dom → Evidence (AppRequires t dom rng) → Maybe (Term rng)

mapTypeSpec ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ t '[a] b → TypeSpec a → Specification b

saturate ∷ t dom Bool → List Term dom → [Pred]

class Syntax t ⇒ Semantics (t ∷ [Type] → Type → Type) where

Semantic operations are ones that give the function symbol, meaning as a function. I.e. how to apply the function to a list of arguments and return a value.

Methods

semantics ∷ ∀ d r. t d r → FunTy d r

class Syntax (t ∷ [Type] → Type → Type) where

A First-order typed logic has 4 components 1. Terms (Variables (x), Constants (5), and Applications (F x 5) Applications, apply a function symbol to a list of arguments: (FunctionSymbol term1 .. termN) 2. Predicates (Ordered, Odd, ...) 3. Connectives (And, Or, Not, =>, ...) 4. Quantifiers (Forall, Exists)

The Syntax, Semantics, and Logic classes implement new function symbols in the first order logic. Note that a function symbol is first order data, that uniquely identifies a higher order function. The three classes supply varying levels of functionality, relating to the Syntax, Semantics, and Logical operations of the function symbol.

Syntactic operations are ones that have to do with the structure and appearence of the type.

Minimal complete definition

Nothing

Methods

inFix ∷ ∀ dom rng. t dom rng → Bool

prettyWit ∷ ∀ dom rng ann. (All HasSpec dom, HasSpec rng) ⇒ t dom rng → List Term dom → Int → Maybe (Doc ann)

class (HasSpec a, NumLike a, Logic IntW) ⇒ Foldy a where

Minimal complete definition

genList, genSizedList, noNegativeValues

Methods

genList ∷ MonadGenError m ⇒ Specification a → Specification a → GenT m [a]

theAddFn ∷ IntW '[a, a] a

theZero ∷ a

genSizedList ∷ MonadGenError m ⇒ Specification Integer → Specification a → Specification a → GenT m [a]

noNegativeValues ∷ Bool

data NumSpec n

Constructors

NumSpecInterval (Maybe n) (Maybe n)

class MaybeBounded a where

Minimal complete definition

Nothing

Methods

lowerBound ∷ Maybe a

default lowerBound ∷ Bounded a ⇒ Maybe a

upperBound ∷ Maybe a

default upperBound ∷ Bounded a ⇒ Maybe a

data NonEmpty a

Constructors

a :| [a]

data Specification a where

A `Specification a` denotes a set of as

Constructors

ExplainSpec ∷ [String] → Specification a → Specification a
Explain a Specification

MemberSpec
Elements of a known set

Fields

∷ NonEmpty a

It must be an element of this OrdSet (List). Try hard not to put duplicates in the List.

→ Specification a

ErrorSpec ∷ NonEmpty String → Specification a
The empty set

SuspendedSpec
The set described by some predicates over the bound variable.

Fields

∷ HasSpec a

⇒ Var a

This variable ranges over values denoted by the spec

→ Pred

And the variable is subject to these constraints

→ Specification a

TypeSpec
A type-specific spec

Fields

∷ HasSpec a

⇒ TypeSpec a

→ [a]

It can't be any of the elements of this set

→ Specification a

TrueSpec ∷ Specification a
Anything

data Term a where

Constructors

App ∷ ∀ t dom rng. AppRequires t dom rng ⇒ t dom rng → List Term dom → Term rng

Lit ∷ (Typeable a, Eq a, Show a) ⇒ a → Term a

V ∷ HasSpec a ⇒ Var a → Term a

data Fun dom rng where

Constructors

Fun ∷ ∀ t dom rng. AppRequires t dom rng ⇒ t dom rng → Fun dom rng

name ∷ String → Term a → Term a

named ∷ String → Term a → Term a

Give a Term a nameHint, if its a Var, and doesn't already have one, otherwise return the Term unchanged.

data Pred where

Constructors

ElemPred ∷ ∀ a. HasSpec a ⇒ Bool → Term a → NonEmpty a → Pred

Monitor ∷ ((∀ a. Term a → a) → Property → Property) → Pred

And ∷ [Pred] → Pred

Exists

Fields

∷ ((∀ b. Term b → b) → GE a)

Constructive recovery function for checking existential quantification

→ Binder a

→ Pred

Subst ∷ HasSpec a ⇒ Var a → Term a → Pred → Pred

Let ∷ Term a → Binder a → Pred

Assert ∷ Term Bool → Pred

Reifies

Fields

∷ (HasSpec a, HasSpec b)

⇒ Term b

This depends on the a term

→ Term a

→ (a → b)

Recover a useable value from the a term.

→ Pred

DependsOn ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Pred

ForAll ∷ (Forallable t a, HasSpec t, HasSpec a) ⇒ Term t → Binder a → Pred

Case

Fields

∷ HasSpec (SumOver as)

⇒ Term (SumOver as)

→ List (Weighted Binder) as

Each branch of the type is bound with only one variable because as are types. Constructors with multiple arguments are encoded with ProdOver (c.f. Univ).

→ Pred

When ∷ HasSpec Bool ⇒ Term Bool → Pred → Pred

GenHint ∷ HasGenHint a ⇒ Hint a → Term a → Pred

TruePred ∷ Pred

FalsePred ∷ NonEmpty String → Pred

Explain ∷ NonEmpty String → Pred → Pred

class (Typeable a, Eq a, Show a, TypeSpecEqShow a) ⇒ HasSpec a where

Minimal complete definition

Nothing

Associated Types

type TypeSpec a

The `TypeSpec a` is the type-specific `Specification a`.

type TypeSpec a = TypeSpec (SimpleRep a)

type Prerequisites a ∷ Constraint

Prerequisites for the instance that are sometimes necessary when working with e.g. Specifications or functions in the universe.

type Prerequisites a = ()

Methods

emptySpec ∷ TypeSpec a

default emptySpec ∷ GenericallyInstantiated a ⇒ TypeSpec a

combineSpec ∷ TypeSpec a → TypeSpec a → Specification a

default combineSpec ∷ GenericallyInstantiated a ⇒ TypeSpec a → TypeSpec a → Specification a

genFromTypeSpec ∷ (HasCallStack, MonadGenError m) ⇒ TypeSpec a → GenT m a

Generate a value that satisfies the Specification. The key property for this generator is soundness: ∀ a ∈ genFromTypeSpec spec. a conformsTo spec

default genFromTypeSpec ∷ (GenericallyInstantiated a, HasCallStack, MonadGenError m) ⇒ TypeSpec a → GenT m a

conformsTo ∷ HasCallStack ⇒ a → TypeSpec a → Bool

Check conformance to the spec.

default conformsTo ∷ (GenericallyInstantiated a, HasCallStack) ⇒ a → TypeSpec a → Bool

shrinkWithTypeSpec ∷ TypeSpec a → a → [a]

Shrink an a with the aide of a Specification

default shrinkWithTypeSpec ∷ GenericallyInstantiated a ⇒ TypeSpec a → a → [a]

toPreds ∷ Term a → TypeSpec a → Pred

Convert a spec to predicates: The key property here is: ∀ a. a conformsTo spec == a conformsTo constrained (t -> toPreds t spec)

default toPreds ∷ GenericallyInstantiated a ⇒ Term a → TypeSpec a → Pred

cardinalTypeSpec ∷ TypeSpec a → Specification Integer

Compute an upper and lower bound on the number of solutions genFromTypeSpec might return

default cardinalTypeSpec ∷ GenericallyInstantiated a ⇒ TypeSpec a → Specification Integer

cardinalTrueSpec ∷ Specification Integer

A bound on the number of solutions `genFromTypeSpec TrueSpec` can produce. For a type with finite elements, we can get a much more accurate answer than TrueSpec

typeSpecHasError ∷ TypeSpec a → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec a → BinaryShow

monadConformsTo ∷ a → TypeSpec a → Writer [String] Bool

typeSpecOpt ∷ TypeSpec a → [a] → Specification a

For some types (especially finite ones) there may be much better ways to construct a Specification than the default method of just adding a large bad list to TypSpec. This function allows each HasSpec instance to decide.

guardTypeSpec ∷ [String] → TypeSpec a → Specification a

This can be used to detect self inconsistencies in a (TypeSpec t) Note this is similar to typeSpecHasError, and the default value for typeSpecHasError is written in terms of guardTypeSpec Both typeSpecHasError and guardTypeSpec can be set individually.

prerequisites ∷ Evidence (Prerequisites a)

Materialize the Prerequisites dictionary. It should not be necessary to implement this function manually.

default prerequisites ∷ Prerequisites a ⇒ Evidence (Prerequisites a)

class Typeable (SimpleRep a) ⇒ HasSimpleRep a where

Minimal complete definition

Nothing

Associated Types

type SimpleRep a

type SimpleRep a = SOP (TheSop a)

type TheSop a ∷ [Type]

type TheSop a = SOPOf (Rep a)

Methods

toSimpleRep ∷ a → SimpleRep a

default toSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ a → SimpleRep a

fromSimpleRep ∷ SimpleRep a → a

default fromSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ SimpleRep a → a

class (Ord a, HasSpec a) ⇒ OrdLike a where

Minimal complete definition

Nothing

Methods

leqSpec ∷ a → Specification a

default leqSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

ltSpec ∷ a → Specification a

default ltSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

geqSpec ∷ a → Specification a

default geqSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

gtSpec ∷ a → Specification a

default gtSpec ∷ (TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, OrdLike (SimpleRep a)) ⇒ a → Specification a

conformsToSpecE ∷ ∀ a. HasSpec a ⇒ a → Specification a → NonEmpty String → Maybe (NonEmpty String)

conformsToSpec with explanation. Nothing if (conformsToSpec a spec), but (Just explanations) if not(conformsToSpec a spec).

conformsToSpec ∷ HasSpec a ⇒ a → Specification a → Bool

satisfies ∷ ∀ a. HasSpec a ⇒ Term a → Specification a → Pred

genFromSpecT ∷ ∀ a m. (HasCallStack, HasSpec a, MonadGenError m) ⇒ Specification a → GenT m a

Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

genFromSpec ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Specification a → Gen a

A version of genFromSpecT that simply errors if the generator fails

debugSpec ∷ ∀ a. HasSpec a ⇒ Specification a → IO ()

A version of genFromSpecT that runs in the IO monad. Good for debugging.

genFromSpecWithSeed ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Int → Int → Specification a → a

A version of genFromSpecT that takes a seed and a size and gives you a result

simplifySpec ∷ HasSpec a ⇒ Specification a → Specification a

cardinality ∷ ∀ a. (Number Integer, HasSpec a) ⇒ Specification a → Specification Integer

Put some (admittedly loose bounds) on the number of solutions that genFromTypeSpec might return. For lots of types, there is no way to be very accurate. Here we lift the HasSpec methods cardinalTrueSpec and cardinalTypeSpec from (TypeSpec Integer) to (Specification Integer)

ifElse ∷ (IsPred p, IsPred q) ⇒ Term Bool → p → q → Pred

If the `Specification Bool` doesn't constrain the boolean you will get a TrueSpec out.

whenTrue ∷ ∀ p. IsPred p ⇒ Term Bool → p → Pred

simplifyTerm ∷ ∀ a. Term a → Term a

Simplify a Term, if the Term is an App, apply the rewrite rules chosen by the (Logic sym t bs a) instance attached to the function witness f

constrained ∷ ∀ a p. (IsPred p, HasSpec a) ⇒ (Term a → p) → Specification a

assertExplain ∷ IsPred p ⇒ NonEmpty String → p → Pred

assert ∷ IsPred p ⇒ p → Pred

forAll ∷ (Forallable t a, HasSpec t, HasSpec a, IsPred p) ⇒ Term t → (Term a → p) → Pred

exists ∷ ∀ a p. (HasSpec a, IsPred p) ⇒ ((∀ b. Term b → b) → GE a) → (Term a → p) → Pred

unsafeExists ∷ ∀ a p. (HasSpec a, IsPred p) ⇒ (Term a → p) → Pred

letBind ∷ (HasSpec a, IsPred p) ⇒ Term a → (Term a → p) → Pred

reify ∷ (HasSpec a, HasSpec b, IsPred p) ⇒ Term a → (a → b) → (Term b → p) → Pred

assertReified ∷ (HasSpec Bool, HasSpec a) ⇒ Term a → (a → Bool) → Pred

requires (HasSpec Bool)

explanation ∷ NonEmpty String → Pred → Pred

Wrap an Explain around a Pred, unless there is a simpler form.

monitor ∷ ((∀ a. Term a → a) → Property → Property) → Pred

Add QuickCheck monitoring (e.g. collect or counterexample) to a predicate. To use the monitoring in a property call monitorSpec on the Specification containing the monitoring and a value generated from the specification.

reifies ∷ (HasSpec a, HasSpec b) ⇒ Term b → Term a → (a → b) → Pred

dependsOn ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Pred

lit ∷ HasSpec a ⇒ a → Term a

genHint ∷ ∀ t. HasGenHint t ⇒ Hint t → Term t → Pred

giveHint ∷ HasGenHint a ⇒ Hint a → Specification a

(<.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(<=.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(>=.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(>.) ∷ ∀ a. OrdLike a ⇒ Term a → Term a → Term Bool infixr 4

(==.) ∷ HasSpec a ⇒ Term a → Term a → Term Bool infix 4

(/=.) ∷ HasSpec a ⇒ Term a → Term a → Term Bool infix 4

not_ ∷ Term Bool → Term Bool

or_ ∷ Term Bool → Term Bool → Term Bool

(||.) ∷ Term Bool → Term Bool → Term Bool infixr 2

toGeneric_ ∷ ∀ a. GenericRequires a ⇒ Term a → Term (SimpleRep a)

fromGeneric_ ∷ ∀ a. (GenericRequires a, AppRequires BaseW '[SimpleRep a] a) ⇒ Term (SimpleRep a) → Term a

(+.) ∷ NumLike a ⇒ Term a → Term a → Term a infix 4

(-.) ∷ Numeric n ⇒ Term n → Term n → Term n infix 4

negate_ ∷ NumLike a ⇒ Term a → Term a

addFn ∷ ∀ a. NumLike a ⇒ Term a → Term a → Term a

negateFn ∷ ∀ a. NumLike a ⇒ Term a → Term a

type Numeric a = (HasSpec a, Ord a, Num a, TypeSpec a ~ NumSpec a, MaybeBounded a)

pair_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term a → Term b → Term (a, b)

fst_ ∷ (HasSpec x, HasSpec y) ⇒ Term (x, y) → Term x

snd_ ∷ (HasSpec x, HasSpec y) ⇒ Term (x, y) → Term y

prodSnd_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term b

prodFst_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term a

prod_ ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Term (Prod a b)

type IsNormalType a = (Cases a ~ '[a], Args a ~ '[a], IsProd a, CountCases a ~ 1)

injLeft_ ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term a → Term (Sum a b)

injRight_ ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term b → Term (Sum a b)

left_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term a → Term (Either a b)

right_ ∷ (HasSpec a, HasSpec b, IsNormalType a, IsNormalType b) ⇒ Term b → Term (Either a b)

cJust_ ∷ (HasSpec a, IsNormalType a) ⇒ Term a → Term (Maybe a)

cNothing_ ∷ (HasSpec a, IsNormalType a) ⇒ Term (Maybe a)

caseOn ∷ ∀ a. (HasSpec a, HasSpec (SimpleRep a), HasSimpleRep a, TypeSpec a ~ TypeSpec (SimpleRep a), SimpleRep a ~ SumOver (Cases (SimpleRep a)), TypeList (Cases (SimpleRep a))) ⇒ Term a → FunTy (MapList (Weighted Binder) (Cases (SimpleRep a))) Pred

branch ∷ ∀ p a. (HasSpec a, All HasSpec (Args a), IsPred p, IsProd a) ⇒ FunTy (MapList Term (Args a)) p → Weighted Binder a

branchW ∷ ∀ p a. (HasSpec a, All HasSpec (Args a), IsPred p, IsProd a) ⇒ Int → FunTy (MapList Term (Args a)) p → Weighted Binder a

forAll' ∷ ∀ t a p. (Forallable t a, Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec a ~ TypeSpec (SimpleRep a), HasSpec t, HasSpec (SimpleRep a), HasSimpleRep a, All HasSpec (Args (SimpleRep a)), IsPred p, IsProd (SimpleRep a), HasSpec a) ⇒ Term t → FunTy (MapList Term (Args (SimpleRep a))) p → Pred

Like forAll but pattern matches on the `Term a`

constrained' ∷ ∀ a p. (Cases (SimpleRep a) ~ '[SimpleRep a], TypeSpec a ~ TypeSpec (SimpleRep a), HasSpec (SimpleRep a), HasSimpleRep a, All HasSpec (Args (SimpleRep a)), IsProd (SimpleRep a), HasSpec a, IsPred p) ⇒ FunTy (MapList Term (Args (SimpleRep a))) p → Specification a

Like constrained but pattern matches on the bound `Term a`

reify' ∷ ∀ a b p. (Cases (SimpleRep b) ~ '[SimpleRep b], TypeSpec b ~ TypeSpec (SimpleRep b), HasSpec (SimpleRep b), HasSimpleRep b, All HasSpec (Args (SimpleRep b)), IsProd (SimpleRep b), HasSpec a, HasSpec b, IsPred p) ⇒ Term a → (a → b) → FunTy (MapList Term (Args (SimpleRep b))) p → Pred

Like reify but pattern matches on the bound `Term b`

con ∷ ∀ c a r. (SimpleRep a ~ SOP (TheSop a), TypeSpec a ~ TypeSpec (SOP (TheSop a)), TypeList (ConstrOf c (TheSop a)), HasSpec a, HasSimpleRep a, r ~ FunTy (MapList Term (ConstrOf c (TheSop a))) (Term a), ResultType r ~ Term a, SOPTerm c (TheSop a), ConstrTerm (ConstrOf c (TheSop a))) ⇒ r

onCon ∷ ∀ c a p. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, HasSpec a, HasSpec (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All HasSpec (Cases (SOP (TheSop a))), HasSpec (ProdOver (ConstrOf c (TheSop a))), IsPred p, Args (ProdOver (ConstrOf c (TheSop a))) ~ ConstrOf c (TheSop a), All HasSpec (ConstrOf c (TheSop a)), IsProd (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term a → FunTy (MapList Term (ConstrOf c (TheSop a))) p → Pred

isCon ∷ ∀ c a. (IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a), TypeSpec a ~ TypeSpec (SimpleRep a), HasSimpleRep a, HasSpec a, HasSpec (SimpleRep a), SumOver (Cases (SOP (TheSop a))) ~ SimpleRep a, All HasSpec (Cases (SOP (TheSop a))), HasSpec (ProdOver (ConstrOf c (TheSop a)))) ⇒ Term a → Pred

sel ∷ ∀ n a c as. (SimpleRep a ~ ProdOver as, TheSop a ~ '[c ::: as], TypeSpec a ~ TypeSpec (ProdOver as), Select n as, HasSpec a, HasSpec (ProdOver as), HasSimpleRep a) ⇒ Term a → Term (At n as)

match ∷ ∀ p a. (HasSpec a, IsProductType a, IsPred p) ⇒ Term a → FunTy (MapList Term (ProductAsList a)) p → Pred

onJust ∷ ∀ a p. (HasSpec a, IsNormalType a, IsPred p) ⇒ Term (Maybe a) → (Term a → p) → Pred

isJust ∷ ∀ a. (HasSpec a, IsNormalType a) ⇒ Term (Maybe a) → Pred

chooseSpec ∷ HasSpec a ⇒ (Int, Specification a) → (Int, Specification a) → Specification a

ChooseSpec is one of the ways we can Or two Specs together This works for any kind of type that has a HasSpec instance. If your type is a Sum type. One can use CaseOn which is much easier.

equalSpec ∷ a → Specification a

notEqualSpec ∷ ∀ a. HasSpec a ⇒ a → Specification a

notMemberSpec ∷ ∀ a f. (HasSpec a, Foldable f) ⇒ f a → Specification a

id_ ∷ ∀ a. HasSpec a ⇒ Term a → Term a

flip_ ∷ ∀ (t ∷ [Type] → Type → Type) a b r. (HasSpec b, HasSpec a, AppRequires t '[a, b] r) ⇒ t '[a, b] r → Term b → Term a → Term r

compose_ ∷ ∀ b t1 t2 a r. (AppRequires t1 '[b] r, AppRequires t2 '[a] b) ⇒ t1 '[b] r → t2 '[a] b → Term a → Term r

foldMap_ ∷ ∀ a b. (Foldy b, HasSpec a) ⇒ (Term a → Term b) → Term [a] → Term b

sum_ ∷ Foldy a ⇒ Term [a] → Term a

elem_ ∷ (Sized [a], HasSpec a) ⇒ Term a → Term [a] → Term Bool infix 4

singletonList_ ∷ (Sized [a], HasSpec a) ⇒ Term a → Term [a]

append_ ∷ (Sized [a], HasSpec a) ⇒ Term [a] → Term [a] → Term [a]

(++.) ∷ HasSpec a ⇒ Term [a] → Term [a] → Term [a] infixr 5

sizeOf ∷ Sized t ⇒ t → Integer

sizeOf_ ∷ (HasSpec a, Sized a) ⇒ Term a → Term Integer

null_ ∷ (HasSpec a, Sized a) ⇒ Term a → Term Bool

hasSize ∷ (HasSpec t, Sized t) ⇒ SizeSpec → Specification t

rangeSize ∷ Integer → Integer → SizeSpec

length_ ∷ HasSpec a ⇒ Term [a] → Term Integer

genFromSizeSpec ∷ MonadGenError m ⇒ Specification Integer → GenT m Integer

Because Sizes should always be >= 0, We provide this alternate generator that can be used to replace (genFromSpecT @Integer), to ensure this important property

between ∷ (HasSpec a, TypeSpec a ~ NumSpec a) ⇒ a → a → Specification a

maxSpec ∷ Specification Integer → Specification Integer

The widest interval whose largest element is admitted by the original spec

data SetW (d ∷ [Type]) (r ∷ Type) where

Constructors

SingletonW ∷ (HasSpec a, Ord a) ⇒ SetW '[a] (Set a)

UnionW ∷ (HasSpec a, Ord a) ⇒ SetW '[Set a, Set a] (Set a)

SubsetW ∷ (HasSpec a, Ord a, HasSpec a) ⇒ SetW '[Set a, Set a] Bool

MemberW ∷ (HasSpec a, Ord a) ⇒ SetW '[a, Set a] Bool

DisjointW ∷ (HasSpec a, Ord a) ⇒ SetW '[Set a, Set a] Bool

FromListW ∷ (HasSpec a, Ord a) ⇒ SetW '[[a]] (Set a)

data SetSpec a

Constructors

SetSpec (Set a) (Specification a) (Specification Integer)

singleton_ ∷ (Ord a, HasSpec a) ⇒ Term a → Term (Set a)

member_ ∷ (Ord a, HasSpec a) ⇒ Term a → Term (Set a) → Term Bool

union_ ∷ (Ord a, HasSpec a) ⇒ Term (Set a) → Term (Set a) → Term (Set a)

subset_ ∷ (Ord a, HasSpec a) ⇒ Term (Set a) → Term (Set a) → Term Bool

disjoint_ ∷ (Ord a, HasSpec a) ⇒ Term (Set a) → Term (Set a) → Term Bool

fromList_ ∷ ∀ a. (Ord a, HasSpec a) ⇒ Term [a] → Term (Set a)

pattern Elem ∷ ∀ b. () ⇒ ∀ a. (b ~ Bool, Eq a, HasSpec a) ⇒ Term a → Term [a] → Term b

pattern ToGeneric ∷ ∀ rng. () ⇒ ∀ a. (rng ~ SimpleRep a, GenericRequires a, HasSpec a, AppRequires BaseW '[a] rng) ⇒ Term a → Term rng

pattern FromGeneric ∷ ∀ rng. () ⇒ ∀ a. (rng ~ a, GenericRequires a, HasSpec a, AppRequires BaseW '[SimpleRep a] rng) ⇒ Term (SimpleRep a) → Term rng

pattern Unary ∷ HOLE a' a → ListCtx f '[a] (HOLE a')

A Convenient pattern for singleton contexts

pattern (:<:) ∷ (Typeable b, Show b) ⇒ HOLE c a → b → ListCtx Value '[a, b] (HOLE c)

Convenient patterns for binary contexts (the arrow :<: points towards the hole)

pattern (:>:) ∷ (Typeable a, Show a) ⇒ a → HOLE c b → ListCtx Value '[a, b] (HOLE c)

Convenient patterns for binary contexts (the arrow :>: points towards the hole)

printPlan ∷ HasSpec a ⇒ Specification a → IO ()

class (Num a, HasSpec a) ⇒ NumLike a

data PairSpec a b

Constructors

Cartesian (Specification a) (Specification b)

data MapSpec k v

Constructors

MapSpec

Fields

mapSpecHint ∷ Maybe Integer

mapSpecMustKeys ∷ Set k

mapSpecMustValues ∷ [v]

mapSpecSize ∷ Specification Integer

mapSpecElem ∷ Specification (k, v)

mapSpecFold ∷ FoldSpec v

dom_ ∷ (HasSpec (Map k v), HasSpec v, HasSpec k, Ord k, IsNormalType k, IsNormalType v) ⇒ Term (Map k v) → Term (Set k)

rng_ ∷ (HasSpec k, HasSpec v, Ord k, IsNormalType k, IsNormalType v) ⇒ Term (Map k v) → Term [v]

lookup_ ∷ (HasSpec k, HasSpec v, Ord k, IsNormalType k, IsNormalType v) ⇒ Term k → Term (Map k v) → Term (Maybe v)

fstSpec ∷ ∀ k v. (HasSpec k, HasSpec v) ⇒ Specification (k, v) → Specification k

sndSpec ∷ ∀ k v. (HasSpec k, HasSpec v) ⇒ Specification (k, v) → Specification v

var ∷ QuasiQuoter

data Prod a b

Constructors

Prod

Fields

prodFst ∷ a

prodSnd ∷ b