Constrained.Base

Description

This module contains most of the implementation of the framework.

NOTE: This is a very big module. Splitting it up would be a nice thing to do but it's not very easy. The problem is that a lot of the things in here depend on each other via a cycle like Pred depends on Term which depend on HasSpec which depends on Specification and Generic and Specification depends in turn on Pred and so on.

Documentation

data Term (fn ∷ [Type] → Type → Type) a where

Typed first order terms with function symbols from fn.

Constructors

App ∷ (Typeable as, TypeList as, All (HasSpec fn) as, HasSpec fn b, BaseUniverse fn) ⇒ fn as b → List (Term fn) as → Term fn b

Lit ∷ Show a ⇒ a → Term fn a

V ∷ HasSpec fn a ⇒ Var a → Term fn a

type HasSpecImpliesEq fn a f = HasSpec fn a ⇒ Eq (f a) ∷ Constraint

data Binder fn a where

Constructors

(:->) ∷ HasSpec fn a ⇒ Var a → Pred fn → Binder fn a

data Pred (fn ∷ [Type] → Type → Type) where

Constructors

Monitor ∷ ((∀ a. Term fn a → a) → Property → Property) → Pred fn

Block ∷ [Pred fn] → Pred fn

Exists

Fields

∷ ((∀ b. Term fn b → b) → GE a)

Constructive recovery function for checking existential quantification

→ Binder fn a

→ Pred fn

Subst ∷ HasSpec fn a ⇒ Var a → Term fn a → Pred fn → Pred fn

Let ∷ Term fn a → Binder fn a → Pred fn

Assert ∷ BaseUniverse fn ⇒ [String] → Term fn Bool → Pred fn

Reifies

Fields

∷ (HasSpec fn a, HasSpec fn b)

⇒ Term fn b

This depends on the a term

→ Term fn a

→ (a → b)

Recover a useable value from the a term.

→ Pred fn

DependsOn ∷ (HasSpec fn a, HasSpec fn b) ⇒ Term fn a → Term fn b → Pred fn

ForAll ∷ (Forallable t a, HasSpec fn t, HasSpec fn a) ⇒ Term fn t → Binder fn a → Pred fn

Case

Fields

∷ HasSpec fn (SumOver as)

⇒ Term fn (SumOver as)

→ List (Weighted (Binder fn)) as

Each branch of the type is bound with only one variable because as are types. Constructors with multiple arguments are encoded with ProdOver (c.f. Univ).

→ Pred fn

When ∷ HasSpec fn Bool ⇒ Term fn Bool → Pred fn → Pred fn

GenHint ∷ HasGenHint fn a ⇒ Hint a → Term fn a → Pred fn

TruePred ∷ Pred fn

FalsePred ∷ [String] → Pred fn

Explain ∷ [String] → Pred fn → Pred fn

data Weighted f a

Constructors

Weighted

Fields

weight ∷ Maybe Int

thing ∷ f a

mapWeighted ∷ (f a → g b) → Weighted f a → Weighted g b

traverseWeighted ∷ Applicative m ⇒ (f a → m (g a)) → Weighted f a → m (Weighted g a)

data Ctx (fn ∷ [Type] → Type → Type) v a where

Contexts for Terms, basically a term with a _single_ HOLE instead of a variable. This is used to traverse the defining constraints for a variable and turn them into a spec. Each subterm `f vs Ctx vs'` for lists of values vs and vs` gets given to the propagateSpecFun for f as `f vs HOLE vs'`.

Constructors

CtxHOLE ∷ HasSpec fn v ⇒ Ctx fn v v
A single hole of type v

CtxApp ∷ (HasSpec fn b, TypeList as, Typeable as, All (HasSpec fn) as) ⇒ fn as b → ListCtx Value as (Ctx fn v) → Ctx fn v b
The application `f vs Ctx vs'`

data HOLE a b where

This is used together with ListCtx to form just the arguments to `f vs Ctx vs'` - replacing Ctx with HOLE - to provide to propagateSpecFun.

Constructors

HOLE ∷ HOLE a a

toCtx ∷ ∀ m fn v a. (BaseUniverse fn, Typeable v, MonadGenError m, HasCallStack) ⇒ Var v → Term fn a → m (Ctx fn v a)

toCtxList ∷ ∀ m fn v as. (BaseUniverse fn, Typeable v, MonadGenError m, HasCallStack) ⇒ Var v → List (Term fn) as → m (ListCtx Value as (Ctx fn v))

runTerm ∷ MonadGenError m ⇒ Env → Term fn a → m a

monitorSpec ∷ (FunctionLike fn, Testable p) ⇒ Specification fn a → a → p → Property

Collect the monitor calls from a specification instantiated to the given value. Typically,

>>> quickCheck $ forAll (genFromSpec spec) $ \ x -> monitorSpec spec x $ ...

monitorPred ∷ ∀ fn m. (FunctionLike fn, MonadGenError m) ⇒ Env → Pred fn → m (Property → Property)

checkPred ∷ ∀ fn m. (FunctionLike fn, MonadGenError m) ⇒ Env → Pred fn → m Bool

checkPreds ∷ (MonadGenError m, Traversable t, FunctionLike fn) ⇒ Env → t (Pred fn) → m Bool

checkPredPure ∷ FunctionLike fn ⇒ Env → Pred fn → Bool

runCaseOn ∷ SumOver as → List (Binder fn) as → (∀ a. HasSpec fn a ⇒ Var a → a → Pred fn → r) → r

type OrdSet a = [a]

data Specification fn a where

A `Specification fn a` denotes a set of as

Constructors

MemberSpec
Elements of a known set

Fields

∷ OrdSet a

It must be an element of this set

→ Specification fn a

ErrorSpec ∷ [String] → Specification fn a
The empty set

SuspendedSpec
The set described by some predicates over the bound variable.

TODO: possibly we want to use a Binder here

Fields

∷ HasSpec fn a

⇒ Var a

This variable ranges over values denoted by the spec

→ Pred fn

And the variable is subject to these constraints

→ Specification fn a

TypeSpec
A type-specific spec

Fields

∷ HasSpec fn a

⇒ TypeSpec fn a

→ OrdSet a

It can't be any of the elements of this set

→ Specification fn a

TrueSpec ∷ Specification fn a
Anything

shrinkPred ∷ Pred fn → [Pred fn]

isSuspendedSpec ∷ Specification fn a → Bool

equalSpec ∷ a → Specification fn a

notEqualSpec ∷ ∀ fn a. HasSpec fn a ⇒ a → Specification fn a

notMemberSpec ∷ ∀ fn a f. (HasSpec fn a, Foldable f) ⇒ f a → Specification fn a

typeSpec ∷ HasSpec fn a ⇒ TypeSpec fn a → Specification fn a

class (Typeable a, Eq a, Show a, Show (TypeSpec fn a), BaseUniverse fn) ⇒ HasSpec fn a where

This class provides the interface that allows you to extend the language to handle a new type. In the case of types that have a generic representation (via HasSimpleRep) that already has an instance of HasSpec it is sufficient to provide an empty instance. However, for types that are used together with specific functions in the function universe fn it may be necessary to provide a specific implementation of HasSpec. This is typically necessary when the Specification for the generic representation does not permit an implementation of propagateSpecFun for some function.

The basic types provided in the language, Set, `[]`, Map, Int, Word64, (,), Either, etc. have instances of this class (technically (,) and Either have instances derived from the instances for their generic Prod and Sum implementations).

Minimal complete definition

Nothing

Associated Types

type TypeSpec (fn ∷ [Type] → Type → Type) a

The `TypeSpec fn a` is the type-specific `Specification fn a`.

type TypeSpec fn a = TypeSpec fn (SimpleRep a)

type Prerequisites fn a ∷ Constraint

Prerequisites for the instance that are sometimes necessary when working with e.g. Specifications or functions in the universe.

type Prerequisites fn a = ()

Methods

emptySpec ∷ TypeSpec fn a

default emptySpec ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ TypeSpec fn a

combineSpec ∷ TypeSpec fn a → TypeSpec fn a → Specification fn a

default combineSpec ∷ (HasSimpleRep a, HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ TypeSpec fn a → TypeSpec fn a → Specification fn a

genFromTypeSpec ∷ (HasCallStack, MonadGenError m) ⇒ TypeSpec fn a → GenT m a

Generate a value that satisfies the Specification. The key property for this generator is soundness: ∀ a ∈ genFromTypeSpec spec. a conformsTo spec

default genFromTypeSpec ∷ (HasSimpleRep a, HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ (HasCallStack, MonadGenError m) ⇒ TypeSpec fn a → GenT m a

conformsTo ∷ HasCallStack ⇒ a → TypeSpec fn a → Bool

Check conformance to the spec.

default conformsTo ∷ (HasSimpleRep a, HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ HasCallStack ⇒ a → TypeSpec fn a → Bool

shrinkWithTypeSpec ∷ TypeSpec fn a → a → [a]

Shrink an a with the aide of a Specification

default shrinkWithTypeSpec ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ TypeSpec fn a → a → [a]

toPreds ∷ Term fn a → TypeSpec fn a → Pred fn

Convert a spec to predicates: The key property here is: ∀ a. a conformsTo spec == a conformsTo constrained (t -> toPreds t spec)

default toPreds ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ Term fn a → TypeSpec fn a → Pred fn

cardinalTypeSpec ∷ TypeSpec fn a → Specification fn Integer

Compute an upper and lower bound on the number of solutions genFromTypeSpec might return

default cardinalTypeSpec ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ TypeSpec fn a → Specification fn Integer

cardinalTrueSpec ∷ Specification fn Integer

A bound on the number of solutions `genFromTypeSpec TrueSpec` can produce. For a type with finite elements, we can get a much more accurate answer than TrueSpec

typeSpecOpt ∷ TypeSpec fn a → [a] → Specification fn a

For some types (especially finite ones) there may be much better ways to construct a Specification than the default method of just adding a large bad list to TypSpec. This function allows each HasSpec instance to decide.

prerequisites ∷ Evidence (Prerequisites fn a)

Materialize the Prerequisites dictionary. It should not be necessary to implement this function manually.

default prerequisites ∷ Prerequisites fn a ⇒ Evidence (Prerequisites fn a)

data WithHasSpec fn f a where

Constructors

WithHasSpec ∷ HasSpec fn a ⇒ f a → WithHasSpec fn f a

class Forallable t e | t → e where

Minimal complete definition

Nothing

Methods

fromForAllSpec ∷ (HasSpec fn t, HasSpec fn e, BaseUniverse fn) ⇒ Specification fn e → Specification fn t

default fromForAllSpec ∷ (HasSpec fn t, HasSpec fn e, HasSimpleRep t, TypeSpec fn t ~ TypeSpec fn (SimpleRep t), Forallable (SimpleRep t) e, HasSpec fn (SimpleRep t)) ⇒ Specification fn e → Specification fn t

forAllToList ∷ t → [e]

default forAllToList ∷ (HasSimpleRep t, Forallable (SimpleRep t) e) ⇒ t → [e]

class (HasSpec fn a, Show (Hint a)) ⇒ HasGenHint fn a where

Hints are things that only affect generation, and not validation. For instance, parameters to control distribution of generated values.

Associated Types

type Hint a

Methods

giveHint ∷ Hint a → Specification fn a

conformsToSpecM ∷ ∀ fn a m. (HasSpec fn a, MonadGenError m, Alternative m) ⇒ a → Specification fn a → m ()

conformsToSpecProp ∷ ∀ fn a. HasSpec fn a ⇒ a → Specification fn a → Property

conformsToSpec ∷ ∀ fn a. HasSpec fn a ⇒ a → Specification fn a → Bool

satisfies ∷ ∀ fn a. HasSpec fn a ⇒ Term fn a → Specification fn a → Pred fn

genFromSpecT ∷ ∀ fn a m. (HasCallStack, HasSpec fn a, MonadGenError m) ⇒ Specification fn a → GenT m a

Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

shrinkWithSpec ∷ ∀ fn a. HasSpec fn a ⇒ Specification fn a → a → [a]

genFromSpec ∷ ∀ fn a. (HasCallStack, HasSpec fn a) ⇒ Specification fn a → Gen a

A version of genFromSpecT that simply errors if the generator fails

genFromSpecWithSeed ∷ ∀ fn a. (HasCallStack, HasSpec fn a) ⇒ Int → Int → Specification fn a → a

A version of genFromSpecT that takes a seed and a size and gives you a result

genInverse ∷ (MonadGenError m, HasSpec fn a, Show b, Functions fn fn, HasSpec fn b) ⇒ fn '[a] b → Specification fn a → b → GenT m a

flattenPred ∷ ∀ fn. BaseUniverse fn ⇒ Pred fn → [Pred fn]

Flatten nested Let, Exists, and Block in a `Pred fn`. Let and Exists bound variables become free in the result.

computeDependencies ∷ Pred fn → DependGraph fn

data SolverStage fn where

Constructors

SolverStage

Fields

∷ HasSpec fn a

⇒ { stageVar ∷ Var a

, stagePreds ∷ [Pred fn]

, stageSpec ∷ Specification fn a

} → SolverStage fn

data SolverPlan fn

Constructors

SolverPlan

Fields

solverPlan ∷ [SolverStage fn]

solverDependencies ∷ Graph (Name fn)

isTrueSpec ∷ Specification fn a → Bool

prettyLinear ∷ [SolverStage fn] → Doc ann

prepareLinearization ∷ ∀ fn. BaseUniverse fn ⇒ Pred fn → GE (SolverPlan fn)

Linearize a predicate, turning it into a list of variables to solve and their defining constraints such that each variable can be solved independently.

mergeSolverStage ∷ SolverStage fn → [SolverStage fn] → [SolverStage fn]

Does nothing if the variable is not in the plan already.

backPropagation ∷ ∀ fn. SolverPlan fn → SolverPlan fn

Push as much information we can backwards through the plan.

pattern Eql ∷ ∀ fn. () ⇒ ∀ a. HasSpec fn a ⇒ Term fn a → Term fn a → Term fn Bool

prettyPlan ∷ HasSpec fn a ⇒ Specification fn a → Doc ann

printPlan ∷ HasSpec fn a ⇒ Specification fn a → IO ()

isEmptyPlan ∷ SolverPlan fn → Bool

stepPlan ∷ MonadGenError m ⇒ SolverPlan fn → GenT m (Env, SolverPlan fn)

substStage ∷ Env → SolverStage fn → SolverStage fn

genFromPreds ∷ (MonadGenError m, BaseUniverse fn) ⇒ Pred fn → GenT m Env

Generate a satisfying Env for a `p : Pred fn`. The Env contains values for all the free variables in `flattenPred p`.

computeHints ∷ ∀ fn. [Pred fn] → Hints fn

computeBinderDependencies ∷ Binder fn a → DependGraph fn

computeTermDependencies ∷ Term fn a → DependGraph fn

computeTermDependencies' ∷ Term fn a → (DependGraph fn, Set (Name fn))

linearize ∷ (MonadGenError m, BaseUniverse fn) ⇒ [Pred fn] → DependGraph fn → m [SolverStage fn]

normalizeSolverStage ∷ SolverStage fn → SolverStage fn

fromGESpec ∷ HasCallStack ⇒ GE (Specification fn a) → Specification fn a

explainSpec ∷ [String] → Specification fn a → Specification fn a

regularize ∷ HasVariables fn t ⇒ Var a → t → Var a

regularizeBinder ∷ Binder fn a → Binder fn a

regularizeNamesPred ∷ Pred fn → Pred fn

regularizeNames ∷ Specification fn a → Specification fn a

simplifySpec ∷ HasSpec fn a ⇒ Specification fn a → Specification fn a

computeSpecSimplified ∷ ∀ fn a. (HasSpec fn a, HasCallStack) ⇒ Var a → Pred fn → GE (Specification fn a)

Precondition: the `Pred fn` defines the `Var a`

Runs in GE in order for us to have detailed context on failure.

computeSpec ∷ ∀ fn a. (HasSpec fn a, HasCallStack) ⇒ Var a → Pred fn → GE (Specification fn a)

Precondition: the `Pred fn` defines the `Var a`.

Runs in GE in order for us to have detailed context on failure.

computeSpecBinder ∷ Binder fn a → GE (Specification fn a)

computeSpecBinderSimplified ∷ Binder fn a → GE (Specification fn a)

caseSpec ∷ ∀ fn as. HasSpec fn (SumOver as) ⇒ List (Weighted (Specification fn)) as → Specification fn (SumOver as)

totalWeight ∷ List (Weighted f) as → Maybe Int

propagateSpec ∷ ∀ fn v a. HasSpec fn v ⇒ Specification fn a → Ctx fn v a → Specification fn v

ctxHasSpec ∷ Ctx fn v a → Evidence (HasSpec fn a)

class (∀ as b. Show (f as b), ∀ as b. Eq (f as b), Typeable f, FunctionLike f) ⇒ Functions f fn where

Minimal complete definition

propagateSpecFun, mapTypeSpec

Methods

propagateSpecFun ∷ (TypeList as, Typeable as, HasSpec fn a, HasSpec fn b, All (HasSpec fn) as) ⇒ f as b → ListCtx Value as (HOLE a) → Specification fn b → Specification fn a

rewriteRules ∷ (TypeList as, Typeable as, HasSpec fn b, All (HasSpec fn) as) ⇒ f as b → List (Term fn) as → Maybe (Term fn b)

mapTypeSpec ∷ (HasSpec fn a, HasSpec fn b) ⇒ f '[a] b → TypeSpec fn a → Specification fn b

mapSpec ∷ ∀ fn a b. (HasSpec fn a, HasSpec fn b) ⇒ fn '[a] b → Specification fn a → Specification fn b

caseBoolSpec ∷ HasSpec fn a ⇒ Specification fn Bool → (Bool → Specification fn a) → Specification fn a

If the `Specification fn Bool` doesn't constrain the boolean you will get a TrueSpec out.

isErrorLike ∷ Specification fn a → Bool

type DependGraph fn = Graph (Name fn)

dependency ∷ HasVariables fn t ⇒ Name fn → t → DependGraph fn

irreflexiveDependencyOn ∷ ∀ fn t t'. (HasVariables fn t, HasVariables fn t') ⇒ t → t' → DependGraph fn

noDependencies ∷ HasVariables fn t ⇒ t → DependGraph fn

type Hints fn = DependGraph fn

respecting ∷ Hints f → DependGraph f → DependGraph f

solvableFrom ∷ Name f → Set (Name f) → DependGraph f → Bool

freeVarNames ∷ ∀ fn t. HasVariables fn t ⇒ t → Set Int

data Name fn where

Constructors

Name ∷ HasSpec fn a ⇒ Var a → Name fn

newtype FreeVars fn

Constructors

FreeVars

Fields

unFreeVars ∷ Map (Name fn) Int

restrictedTo ∷ FreeVars fn → Set (Name fn) → FreeVars fn

memberOf ∷ Name fn → FreeVars fn → Bool

count ∷ Name fn → FreeVars fn → Int

freeVar ∷ Name fn → FreeVars fn

singleton ∷ Name fn → Int → FreeVars fn

without ∷ Foldable t ⇒ FreeVars fn → t (Name fn) → FreeVars fn

class HasVariables fn a | a → fn where

Minimal complete definition

freeVars

Methods

freeVars ∷ a → FreeVars fn

freeVarSet ∷ a → Set (Name fn)

countOf ∷ Name fn → a → Int

appearsIn ∷ Name fn → a → Bool

type Subst fn = [SubstEntry fn]

data SubstEntry fn where

Constructors

(:=) ∷ HasSpec fn a ⇒ Var a → Term fn a → SubstEntry fn

backwardsSubstitution ∷ ∀ fn a. HasSpec fn a ⇒ Subst fn → Term fn a → Term fn a

fastInequality ∷ Term fn a → Term fn b → Bool

Sound but not complete inequality on terms

substituteTerm ∷ ∀ fn a. Subst fn → Term fn a → Term fn a

substituteTerm' ∷ ∀ fn a. Subst fn → Term fn a → Writer Any (Term fn a)

substituteBinder ∷ HasSpec fn a ⇒ Var a → Term fn a → Binder fn b → Binder fn b

substitutePred ∷ HasSpec fn a ⇒ Var a → Term fn a → Pred fn → Pred fn

substTerm ∷ Env → Term fn a → Term fn a

substBinder ∷ Env → Binder fn a → Binder fn a

substPred ∷ BaseUniverse fn ⇒ Env → Pred fn → Pred fn

unBind ∷ a → Binder fn a → Pred fn

simplifyTerm ∷ ∀ fn a. BaseUniverse fn ⇒ Term fn a → Term fn a

fromLits ∷ List (Term fn) as → Maybe (List Value as)

fromLit ∷ Term fn a → Maybe (Value a)

isLit ∷ Term fn a → Bool

simplifyPred ∷ ∀ fn. BaseUniverse fn ⇒ Pred fn → Pred fn

simplifyPreds ∷ BaseUniverse fn ⇒ [Pred fn] → [Pred fn]

simplifyBinder ∷ Binder fn a → Binder fn a

pinnedBy ∷ ∀ fn a. (BaseUniverse fn, Typeable a) ⇒ Var a → Pred fn → Maybe (Term fn a)

Is the variable x pinned to some free term in p? (free term meaning that all the variables in the term are free in p).

TODO: complete this with more cases!

optimisePred ∷ BaseUniverse fn ⇒ Pred fn → Pred fn

letSubexpressionElimination ∷ BaseUniverse fn ⇒ Pred fn → Pred fn

letFloating ∷ BaseUniverse fn ⇒ Pred fn → Pred fn

aggressiveInlining ∷ BaseUniverse fn ⇒ Pred fn → Pred fn

substituteAndSimplifyTerm ∷ BaseUniverse fn ⇒ Subst fn → Term fn a → Term fn a

Apply a substitution and simplify the resulting term if the substitution changed the term.

class HasSimpleRep a where

Minimal complete definition

Nothing

Associated Types

type SimpleRep a

type SimpleRep a = SOP (TheSop a)

type TheSop a ∷ [Type]

type TheSop a = SOPOf (Rep a)

Methods

toSimpleRep ∷ a → SimpleRep a

default toSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ a → SimpleRep a

fromSimpleRep ∷ SimpleRep a → a

default fromSimpleRep ∷ (Generic a, SimpleGeneric (Rep a), SimpleRep a ~ SimplifyRep (Rep a)) ⇒ SimpleRep a → a

type family SimplifyRep f where ...

Equations

SimplifyRep f = SOP (SOPOf f)

toGenericFn ∷ ∀ fn a. (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ fn '[a] (SimpleRep a)

fromGenericFn ∷ ∀ fn a. (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ fn '[SimpleRep a] a

data GenericsFn fn args res where

Constructors

ToGeneric ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ GenericsFn fn '[a] (SimpleRep a)

FromGeneric ∷ (HasSpec fn (SimpleRep a), TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a) ⇒ GenericsFn fn '[SimpleRep a] a

fromSimpleRepSpec ∷ ∀ a fn. (HasSpec fn a, HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Specification fn (SimpleRep a) → Specification fn a

toSimpleRepSpec ∷ ∀ a fn. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Specification fn a → Specification fn (SimpleRep a)

data (c ∷ Symbol) ::: (ts ∷ [Type])

A constructor name with the types of its arguments

type family SOPOf f where ...

Turn a Rep into a list that flattens the sum structre and gives the constructors names: Maybe Int -> '[Nothing ::: '[()], Just ::: '[Int]] Either Int Bool -> '[Left ::: '[Int], Right ::: '[Bool]] data Foo = Foo Int Bool | Bar Double -> '[Foo ::: '[Int, Bool], Bar ::: '[Double]]

Equations

SOPOf (D1 _ f) = SOPOf f

SOPOf (f :+: g) = Append (SOPOf f) (SOPOf g)

SOPOf (C1 ('MetaCons constr _ _) f) = '[constr ::: Constr f]

type family Constr f where ...

Flatten a single constructor

Equations

Constr (S1 _ f) = Constr f

Constr (K1 _ k) = '[k]

Constr U1 = '[()]

Constr (f :*: g) = Append (Constr f) (Constr g)

type family SOP constrs where ...

Turn a list from SOPOf into a Sum over Prod representation.

Equations

SOP '[c ::: prod] = ProdOver prod

SOP ((c ::: prod) ': constrs) = Sum (ProdOver prod) (SOP constrs)

type family ConstrOf c sop where ...

Equations

ConstrOf c ((c ::: constr) ': sop) = constr

ConstrOf c (_ ': sop) = ConstrOf c sop

class Inject c constrs r where

Methods

inject' ∷ (SOP constrs → r) → FunTy (ConstrOf c constrs) r

inject ∷ ∀ c constrs. Inject c constrs (SOP constrs) ⇒ FunTy (ConstrOf c constrs) (SOP constrs)

type family ALG constrs r where ...

An `ALG constrs r` is a function that takes a way to turn every constructor into an r: ``` ALG (SOPOf (Rep (Either Int Bool))) r = (Int -> r) -> (Bool -> r) -> r ```

Equations

ALG '[c ::: prod] r = FunTy prod r → r

ALG ((c ::: prod) ': constrs) r = FunTy prod r → ALG constrs r

class SOPLike constrs r where

Methods

algebra ∷ SOP constrs → ALG constrs r

Run a SOP

consts ∷ r → ALG constrs r

Ignore everything in the SOP

class SimpleConstructor rep where

Methods

toSimpleCon' ∷ rep p → ProdOver (Constr rep)

fromSimpleCon' ∷ ProdOver (Constr rep) → rep p

class SopList xs ys where

Construct and deconstruct cases in a SOP

Methods

injectSOPLeft ∷ SOP xs → SOP (Append xs ys)

injectSOPRight ∷ SOP ys → SOP (Append xs ys)

caseSOP ∷ SOP (Append xs ys) → Sum (SOP xs) (SOP ys)

class SimpleGeneric rep where

Methods

toSimpleRep' ∷ rep p → SimplifyRep rep

fromSimpleRep' ∷ SimplifyRep rep → rep p

class HasSpec fn a ⇒ Foldy fn a where

Methods

genList ∷ (BaseUniverse fn, MonadGenError m) ⇒ Specification fn a → Specification fn a → GenT m [a]

theAddFn ∷ fn '[a, a] a

theZero ∷ a

adds ∷ ∀ fn a. Foldy fn a ⇒ [a] → a

data FoldSpec (fn ∷ [Type] → Type → Type) a where

Constructors

NoFold ∷ FoldSpec fn a

FoldSpec ∷ ∀ b fn a. (HasSpec fn a, HasSpec fn b, Foldy fn b, Member (ListFn fn) fn, BaseUniverse fn) ⇒ fn '[a] b → Specification fn b → FoldSpec fn a

preMapFoldSpec ∷ HasSpec fn a ⇒ fn '[a] b → FoldSpec fn b → FoldSpec fn a

combineFoldSpec ∷ MonadGenError m ⇒ FoldSpec fn a → FoldSpec fn a → m (FoldSpec fn a)

conformsToFoldSpec ∷ ∀ fn a. [a] → FoldSpec fn a → Bool

toPredsFoldSpec ∷ ∀ fn a. BaseUniverse fn ⇒ Term fn [a] → FoldSpec fn a → Pred fn

enumerateInterval ∷ (Enum a, Num a, Ord a, MaybeBounded a) ⇒ NumSpec a → [a]

Note: potentially infinite list

isEmptyNumSpec ∷ (TypeSpec fn a ~ NumSpec a, Ord a, Enum a, Num a, MaybeBounded a) ⇒ Specification fn a → Bool

knownUpperBound ∷ (TypeSpec fn a ~ NumSpec a, Ord a, Enum a, Num a, MaybeBounded a) ⇒ Specification fn a → Maybe a

knownLowerBound ∷ (TypeSpec fn a ~ NumSpec a, Ord a, Enum a, Num a, MaybeBounded a) ⇒ Specification fn a → Maybe a

narrowByFuelAndSize

Arguments

∷ ∀ a fn. (BaseUniverse fn, TypeSpec fn a ~ NumSpec a, HasSpec fn a, Arbitrary a, Integral a, Ord a, Random a, MaybeBounded a)

⇒ a
Fuel

→ Int
Integer

→ (Specification fn a, Specification fn a)

→ (Specification fn a, Specification fn a)

narrowFoldSpecs ∷ ∀ a fn. (BaseUniverse fn, TypeSpec fn a ~ NumSpec a, HasSpec fn a, Arbitrary a, Integral a, Ord a, Random a, MaybeBounded a) ⇒ (Specification fn a, Specification fn a) → (Specification fn a, Specification fn a)

genNumList ∷ ∀ a fn m. (BaseUniverse fn, MonadGenError m, TypeSpec fn a ~ NumSpec a, HasSpec fn a, Arbitrary a, Integral a, Ord a, Random a, MaybeBounded a, Foldy fn a) ⇒ Specification fn a → Specification fn a → GenT m [a]

genFromFold ∷ ∀ m fn a b. (MonadGenError m, Foldy fn b, HasSpec fn a) ⇒ [a] → Specification fn Integer → Specification fn a → fn '[a] b → Specification fn b → GenT m [a]

guardSumSpec ∷ (HasSpec fn a, HasSpec fn b, KnownNat (CountCases b)) ⇒ SumSpec fn a b → Specification fn (Sum a b)

data SumSpec fn a b

Constructors

SumSpec (Maybe (Int, Int)) (Specification fn a) (Specification fn b)

type family CountCases a where ...

Equations

CountCases (Sum a b) = 1 + CountCases b

CountCases _ = 1

countCases ∷ ∀ a. KnownNat (CountCases a) ⇒ Int

data SetSpec fn a

Constructors

SetSpec (Set a) (Specification fn a) (Specification fn Integer)

guardSetSpec ∷ (HasSpec fn a, Ord a) ⇒ SetSpec fn a → Specification fn (Set a)

data ListSpec fn a

Constructors

ListSpec

Fields

listSpecHint ∷ Maybe Integer

listSpecMust ∷ [a]

listSpecSize ∷ Specification fn Integer

listSpecElem ∷ Specification fn a

listSpecFold ∷ FoldSpec fn a

class MaybeBounded a where

Minimal complete definition

Nothing

Methods

lowerBound ∷ Maybe a

default lowerBound ∷ Bounded a ⇒ Maybe a

upperBound ∷ Maybe a

default upperBound ∷ Bounded a ⇒ Maybe a

data NumSpec n

Constructors

NumSpecInterval (Maybe n) (Maybe n)

emptyNumSpec ∷ Ord a ⇒ NumSpec a

combineNumSpec ∷ (HasSpec fn n, Ord n, TypeSpec fn n ~ NumSpec n) ⇒ NumSpec n → NumSpec n → Specification fn n

genFromNumSpec ∷ (MonadGenError m, Show n, Random n, Ord n, Num n, MaybeBounded n) ⇒ NumSpec n → GenT m n

shrinkWithNumSpec ∷ Arbitrary n ⇒ NumSpec n → n → [n]

constrainInterval ∷ (MonadGenError m, Ord a, Num a, Show a) ⇒ Maybe a → Maybe a → Integer → m (a, a)

conformsToNumSpec ∷ Ord n ⇒ n → NumSpec n → Bool

toPredsNumSpec ∷ (Ord n, OrdLike fn n) ⇒ Term fn n → NumSpec n → Pred fn

type BaseFns = '[EqFn, SetFn, BoolFn, PairFn, IntFn, OrdFn, GenericsFn, ListFn, SumFn, MapFn, FunFn, SizeFn]

type BaseFn = Fix (OneofL BaseFns)

type BaseUniverse fn = (Functions fn fn, Member (EqFn fn) fn, Member (SetFn fn) fn, Member (BoolFn fn) fn, Member (PairFn fn) fn, Member (IntFn fn) fn, Member (OrdFn fn) fn, Member (GenericsFn fn) fn, Member (ListFn fn) fn, Member (SumFn fn) fn, Member (MapFn fn) fn, Member (FunFn fn) fn, Member (SizeFn fn) fn)

A minimal Universe contains functions for a bunch of different things.

idFn ∷ ∀ fn a. Member (FunFn fn) fn ⇒ fn '[a] a

composeFn ∷ (Member (FunFn fn) fn, HasSpec fn b, Show (fn '[a] b), Show (fn '[b] c), Eq (fn '[a] b), Eq (fn '[b] c)) ⇒ fn '[b] c → fn '[a] b → fn '[a] c

flip_ ∷ ∀ fn a b c. (Member (FunFn fn) fn, Typeable a, Typeable b, HasSpec fn a, HasSpec fn b, HasSpec fn c) ⇒ (Term fn a → Term fn b → Term fn c) → Term fn b → Term fn a → Term fn c

data FunFn fn args res where

Constructors

Id ∷ FunFn fn '[a] a

Compose ∷ (Typeable b, HasSpec fn b, Show (fn '[a] b), Show (fn '[b] c), Eq (fn '[a] b), Eq (fn '[b] c)) ⇒ fn '[b] c → fn '[a] b → FunFn fn '[a] c

Flip ∷ (Show (fn '[a, b] c), Eq (fn '[a, b] c), HasSpec fn a, HasSpec fn b) ⇒ fn '[a, b] c → FunFn fn '[b, a] c

lessOrEqualFn ∷ ∀ fn a. (Ord a, OrdLike fn a) ⇒ fn '[a, a] Bool

lessFn ∷ ∀ fn a. (Ord a, OrdLike fn a) ⇒ fn '[a, a] Bool

data OrdFn (fn ∷ [Type] → Type → Type) as b where

Constructors

LessOrEqual ∷ (Ord a, OrdLike fn a) ⇒ OrdFn fn '[a, a] Bool

Less ∷ (Ord a, OrdLike fn a) ⇒ OrdFn fn '[a, a] Bool

class HasSpec fn a ⇒ OrdLike fn a where

Minimal complete definition

Nothing

Methods

leqSpec ∷ a → Specification fn a

default leqSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Specification fn a

ltSpec ∷ a → Specification fn a

default ltSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Specification fn a

geqSpec ∷ a → Specification fn a

default geqSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Specification fn a

gtSpec ∷ a → Specification fn a

default gtSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, OrdLike fn (SimpleRep a)) ⇒ a → Specification fn a

data ListFn fn args res where

Constructors

FoldMap ∷ (HasSpec fn a, Foldy fn b, Show (fn '[a] b), Eq (fn '[a] b)) ⇒ fn '[a] b → ListFn fn '[[a]] b

foldMapFn ∷ ∀ fn a b. (HasSpec fn a, Foldy fn b, Show (fn '[a] b), Eq (fn '[a] b)) ⇒ fn '[a] b → fn '[[a]] b

addFn ∷ ∀ fn a. NumLike fn a ⇒ fn '[a, a] a

negateFn ∷ ∀ fn a. NumLike fn a ⇒ fn '[a] a

data IntFn (fn ∷ [Type] → Type → Type) as b where

Constructors

Add ∷ NumLike fn a ⇒ IntFn fn '[a, a] a

Negate ∷ NumLike fn a ⇒ IntFn fn '[a] a

class (Num a, HasSpec fn a) ⇒ NumLike fn a where

Minimal complete definition

Nothing

Methods

subtractSpec ∷ a → TypeSpec fn a → Specification fn a

default subtractSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, NumLike fn (SimpleRep a)) ⇒ a → TypeSpec fn a → Specification fn a

negateSpec ∷ TypeSpec fn a → Specification fn a

default negateSpec ∷ (TypeSpec fn a ~ TypeSpec fn (SimpleRep a), HasSimpleRep a, NumLike fn (SimpleRep a)) ⇒ TypeSpec fn a → Specification fn a

safeSubtract ∷ a → a → Maybe a

default safeSubtract ∷ (HasSimpleRep a, NumLike fn (SimpleRep a)) ⇒ a → a → Maybe a

toGeneric_ ∷ ∀ a fn. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Term fn a → Term fn (SimpleRep a)

fromGeneric_ ∷ ∀ a fn. (HasSpec fn a, HasSpec fn (SimpleRep a), HasSimpleRep a, TypeSpec fn a ~ TypeSpec fn (SimpleRep a)) ⇒ Term fn (SimpleRep a) → Term fn a

not_ ∷ BaseUniverse fn ⇒ Term fn Bool → Term fn Bool

(||.) ∷ BaseUniverse fn ⇒ Term fn Bool → Term fn Bool → Term fn Bool infixr 2

elem_ ∷ ∀ a fn. HasSpec fn a ⇒ Term fn a → Term fn [a] → Term fn Bool

member_ ∷ ∀ a fn. (HasSpec fn a, Ord a) ⇒ Term fn a → Term fn (Set a) → Term fn Bool

subset_ ∷ (HasSpec fn a, Ord a) ⇒ Term fn (Set a) → Term fn (Set a) → Term fn Bool

disjoint_ ∷ (HasSpec fn a, Ord a) ⇒ Term fn (Set a) → Term fn (Set a) → Term fn Bool

singleton_ ∷ (HasSpec fn a, Ord a) ⇒ Term fn a → Term fn (Set a)

union_ ∷ ∀ a fn. (HasSpec fn a, Ord a) ⇒ Term fn (Set a) → Term fn (Set a) → Term fn (Set a)

fromList_ ∷ ∀ a fn. (HasSpec fn a, Ord a) ⇒ Term fn [a] → Term fn (Set a)

sizeOf_ ∷ ∀ a fn. (HasSpec fn a, Sized a) ⇒ Term fn a → Term fn Integer

size_ ∷ ∀ a fn. (HasSpec fn (Set a), Ord a) ⇒ Term fn (Set a) → Term fn Integer

special instance of sizeOf (for Sets) for backward compatibility

length_ ∷ ∀ a fn. HasSpec fn [a] ⇒ Term fn [a] → Term fn Integer

special instance of sizeOf (for Lists) for backward compatibility

null_ ∷ (HasSpec fn a, Sized a) ⇒ Term fn a → Term fn Bool

(<=.) ∷ (Ord a, OrdLike fn a) ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(>=.) ∷ (Ord a, OrdLike fn a) ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(<.) ∷ (Ord a, OrdLike fn a) ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(>.) ∷ (Ord a, OrdLike fn a) ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(==.) ∷ HasSpec fn a ⇒ Term fn a → Term fn a → Term fn Bool infix 4

(/=.) ∷ HasSpec fn a ⇒ Term fn a → Term fn a → Term fn Bool infix 4

sum_ ∷ (BaseUniverse fn, Member (FunFn fn) fn, Foldy fn a) ⇒ Term fn [a] → Term fn a

foldMap_ ∷ ∀ fn a b. (BaseUniverse fn, Foldy fn b, HasSpec fn a) ⇒ (Term fn a → Term fn b) → Term fn [a] → Term fn b

constrained ∷ ∀ a fn p. (IsPred p fn, HasSpec fn a) ⇒ (Term fn a → p) → Specification fn a

assertExplain ∷ (BaseUniverse fn, IsPred p fn) ⇒ [String] → p → Pred fn

assert ∷ (BaseUniverse fn, IsPred p fn) ⇒ p → Pred fn

forAll ∷ (Forallable t a, HasSpec fn t, HasSpec fn a, IsPred p fn) ⇒ Term fn t → (Term fn a → p) → Pred fn

mkForAll ∷ (Forallable t a, HasSpec fn t, HasSpec fn a) ⇒ Term fn t → Binder fn a → Pred fn

exists ∷ ∀ a p fn. (HasSpec fn a, IsPred p fn) ⇒ ((∀ b. Term fn b → b) → GE a) → (Term fn a → p) → Pred fn

unsafeExists ∷ ∀ a p fn. (HasSpec fn a, IsPred p fn) ⇒ (Term fn a → p) → Pred fn

letBind ∷ (HasSpec fn a, IsPred p fn) ⇒ Term fn a → (Term fn a → p) → Pred fn

reify ∷ (HasSpec fn a, HasSpec fn b, IsPred p fn) ⇒ Term fn a → (a → b) → (Term fn b → p) → Pred fn

explanation ∷ [String] → Pred fn → Pred fn

monitor ∷ ((∀ a. Term fn a → a) → Property → Property) → Pred fn

Add QuickCheck monitoring (e.g. collect or counterexample) to a predicate. To use the monitoring in a property call monitorSpec on the Specification containing the monitoring and a value generated from the specification.

assertReified ∷ HasSpec fn a ⇒ Term fn a → (a → Bool) → Pred fn

reifies ∷ (HasSpec fn a, HasSpec fn b) ⇒ Term fn b → Term fn a → (a → b) → Pred fn

dependsOn ∷ (HasSpec fn a, HasSpec fn b) ⇒ Term fn a → Term fn b → Pred fn

lit ∷ Show a ⇒ a → Term fn a

ifElse ∷ (BaseUniverse fn, IsPred p fn, IsPred q fn) ⇒ Term fn Bool → p → q → Pred fn

whenTrue ∷ ∀ fn p. (BaseUniverse fn, IsPred p fn) ⇒ Term fn Bool → p → Pred fn

genHint ∷ ∀ fn t. HasGenHint fn t ⇒ Hint t → Term fn t → Pred fn

app ∷ (HasSpec fn b, Typeable as, TypeList as, All (HasSpec fn) as) ⇒ fn as b → FunTy (MapList (Term fn) as) (Term fn b)

name ∷ String → Term fn a → Term fn a

bind ∷ (HasSpec fn a, IsPred p fn) ⇒ (Term fn a → p) → Binder fn a

mkCase ∷ HasSpec fn (SumOver as) ⇒ Term fn (SumOver as) → List (Weighted (Binder fn)) as → Pred fn

type IsPred p fn = (PredLike p, UnivConstr p fn)

class PredLike p where

Associated Types

type UnivConstr p (fn ∷ [Type] → Type → Type) ∷ Constraint

Methods

toPredExplain ∷ (BaseUniverse fn, UnivConstr p fn) ⇒ [String] → p → Pred fn

toPred ∷ (BaseUniverse fn, IsPred p fn) ⇒ p → Pred fn

data WithPrec a

Constructors

WithPrec Int a

parensIf ∷ Bool → Doc ann → Doc ann

prettyPrec ∷ Pretty (WithPrec a) ⇒ Int → a → Doc ann

ppList ∷ ∀ fn f as ann. All (HasSpec fn) as ⇒ (∀ a. HasSpec fn a ⇒ f a → Doc ann) → List f as → [Doc ann]

ppList_ ∷ ∀ f as ann. (∀ a. f a → Doc ann) → List f as → [Doc ann]

vsep' ∷ [Doc ann] → Doc ann

(/>) ∷ Doc ann → Doc ann → Doc ann infixl 5

genFromSizeSpec ∷ (BaseUniverse fn, MonadGenError m) ⇒ Specification fn Integer → GenT m Integer

Because Sizes should always be >= 0, We provide this alternate generator that can be used to replace (genFromSpecT @Integer), to ensure this important property

data SizeFn (fn ∷ [Type] → Type → Type) as b where

Constructors

SizeOf ∷ ∀ fn a. (Sized a, HasSpec fn a) ⇒ SizeFn fn '[a] Integer

sizeOfFn ∷ ∀ fn a. (HasSpec fn a, Member (SizeFn fn) fn, Sized a) ⇒ fn '[a] Integer

mapTypeSpecSize ∷ ∀ fn a b f. f ~ SizeFn fn ⇒ f '[a] b → TypeSpec fn a → Specification fn b

type SizeSpec = NumSpec Integer

rangeSize ∷ Integer → Integer → SizeSpec

maxSpec ∷ BaseUniverse fn ⇒ Specification fn Integer → Specification fn Integer

The widest interval whose largest element is admitted by the original spec

class Sized t where

Methods

sizeOf ∷ t → Integer

liftSizeSpec ∷ HasSpec fn t ⇒ SizeSpec → [Integer] → Specification fn t

liftMemberSpec ∷ HasSpec fn t ⇒ OrdSet Integer → Specification fn t

sizeOfTypeSpec ∷ HasSpec fn t ⇒ TypeSpec fn t → Specification fn Integer

hasSize ∷ (HasSpec fn t, Sized t) ⇒ SizeSpec → Specification fn t

guardEmpty ∷ Maybe Integer → Maybe Integer → NumSpec Integer → NumSpec Integer

addNumSpec ∷ NumSpec Integer → NumSpec Integer → NumSpec Integer

subNumSpec ∷ NumSpec Integer → NumSpec Integer → NumSpec Integer

multNumSpec ∷ NumSpec Integer → NumSpec Integer → NumSpec Integer

negNumSpec ∷ NumSpec Integer → NumSpec Integer

addSpecInt ∷ BaseUniverse fn ⇒ Specification fn Integer → Specification fn Integer → Specification fn Integer

subSpecInt ∷ BaseUniverse fn ⇒ Specification fn Integer → Specification fn Integer → Specification fn Integer

multSpecInt ∷ BaseUniverse fn ⇒ Specification fn Integer → Specification fn Integer → Specification fn Integer

operateSpec ∷ (TypeSpec fn n ~ NumSpec n, Enum n, Ord n) ⇒ (n → n → n) → (TypeSpec fn n → TypeSpec fn n → TypeSpec fn n) → Specification fn n → Specification fn n → Specification fn n

let n be some numeric type, and f and ft be operations on n and (TypeSpec fn n) Then lift these operations from (TypeSpec fn n) to (Specification fn n) Normally f will be a (Num n) instance method (+,-,*) on n, and ft will be a a (Num (TypeSpec fn n)) instance method (+,-,*) on (TypeSpec fn n) But this will work for any operations f and ft with the right types

cardinality ∷ ∀ fn a. (Eq a, BaseUniverse fn, HasSpec fn a) ⇒ Specification fn a → Specification fn Integer

Put some (admittedly loose bounds) on the number of solutions that genFromTypeSpec might return. For lots of types, there is no way to be very accurate. Here we lift the HasSpec methods cardinalTrueSpec and cardinalTypeSpec from (TypeSpec fn Integer) to (Specification fn Integer)

cardinalNumSpec ∷ ∀ n fn. (Integral n, Num n, MaybeBounded n) ⇒ NumSpec n → Specification fn Integer

A generic function to use as an instance for the HasSpec method cardinalTypeSpec :: HasSpec fn a => TypeSpec fn a -> Specification fn Integer for types n such that (TypeSpec n ~ NumSpec n)

lowBound ∷ Bounded n ⇒ Maybe n → n

highBound ∷ Bounded n ⇒ Maybe n → n

countSpec ∷ ∀ n. (Bounded n, Integral n) ⇒ NumSpec n → Integer

The exact count of the number elements in a Bounded NumSpec

finiteSize ∷ ∀ n. (Integral n, Bounded n) ⇒ Integer

The exact number of elements in a Bounded Integral type.

notInNumSpec ∷ ∀ fn n. (Functions fn fn, BaseUniverse fn, HasSpec fn n, TypeSpec fn n ~ NumSpec n, Bounded n, Integral n) ⇒ NumSpec n → [n] → Specification fn n

This is an optimizing version of TypeSpec :: TypeSpec fn n -> [n] -> Specification fn n for Bounded NumSpecs. notInNumSpec :: Bounded n => TypeSpec fn n -> [n] -> Specification fn n We use this function to specialize the (HasSpec fn t) method typeSpecOpt for Bounded n. So given (TypeSpec interval badlist) we might want to transform it to (MemberSpec goodlist) There are 2 opportunities where this can payoff big time. 1) Suppose the total count of the elements in the interval is < length badlist we can then return (MemberSpec (filter elements (notElem badlist))) this must be smaller than (TypeSpec interval badlist) because the filtered list must be smaller than badlist 2) Suppose the type t is finite with size N. If the length of the badlist > (N/2), then the number of possible good things must be smaller than (length badlist), because (possible good + bad == N), so regardless of the count of the interval (MemberSpec (filter elements (notElem badlist))) is better. Sometimes much better. Example, let n be the finite set {0,1,2,3,4,5,6,7,8,9} and the bad list be [0,1,3,4,5,6,8,9] (TypeSpec [0..9] [0,1,3,4,5,6,8,9]) = filter {0,1,2,3,4,5,6,7,8,9} (notElem [0,1,3,4,5,6,8,9]) = [2,7] So (MemberSpec [2,7]) is better than (TypeSpec [0..9] [0,1,3,4,5,6,8,9]). This works no matter what the count of interval is. We only need the (length badlist > (N/2)).

data T x

T is a sort of special version of Maybe, with two Nothings. Given:: NumSpecInterval (Maybe n) (Maybe n) -> Numspec We can't distinguish between the two Nothings in (NumSpecInterval Nothing Nothing) But using (NumSpecInterval NegInf PosInf) we can, In fact we can make a total ordering on T So an ascending Sorted [T x] would all the NegInf on the left and all the PosInf on the right, with the Ok's sorted in between. I.e. [NegInf, NegInf, Ok 3, Ok 6, Ok 12, Pos Inf]

Constructors

NegInf

Ok x

PosInf

unT ∷ T x → Maybe x

Conversion between (T x) and (Maybe x)

neg ∷ Maybe x → T x

Use this on the lower bound. I.e. lo from pair (lo,hi)

pos ∷ Maybe x → T x

Use this on the upper bound. I.e. hi from pair (lo,hi)

multT ∷ Num x ⇒ T x → T x → T x

multiply two (T x), correctly handling the infinities NegInf and PosInf

sizeOfSpec ∷ ∀ fn t. (BaseUniverse fn, Sized t) ⇒ Specification fn t → Specification fn Integer

sizeOfSpec generalizes the method sizeOfTypeSpec From (sizeOfTypeSpec :: TypeSpec fn t -> Specification fn Integer) To (sizeOfSpec :: Specification fn t -> Specification fn Integer) It is not unusual for instances (HasSpec fn t) to define sizeOfTypeSpec with calls to sizeOfSpec, Because many (TypeSpec fn t)'s contain (Specification fn s), for types s different from t

checkForNegativeSize ∷ Specification fn Integer → Specification fn Integer

Turn a Size spec into an ErrorSpec if it has negative numbers.

Orphan instances

Arbitrary Natural

Instance details

Methods

arbitrary ∷ Gen Natural

shrink ∷ Natural → [Natural]

Random Natural

Instance details

Methods

randomR ∷ RandomGen g ⇒ (Natural, Natural) → g → (Natural, g)

random ∷ RandomGen g ⇒ g → (Natural, g)

randomRs ∷ RandomGen g ⇒ (Natural, Natural) → g → [Natural]

randoms ∷ RandomGen g ⇒ g → [Natural]

Uniform Natural

Instance details

Methods

uniformM ∷ StatefulGen g m ⇒ g → m Natural

Pretty (Var a)

Instance details

Methods

pretty ∷ Var a → Doc ann

prettyList ∷ [Var a] → Doc ann

Random (Ratio Integer)

Instance details

Methods

randomR ∷ RandomGen g ⇒ (Ratio Integer, Ratio Integer) → g → (Ratio Integer, g)

random ∷ RandomGen g ⇒ g → (Ratio Integer, g)

randomRs ∷ RandomGen g ⇒ (Ratio Integer, Ratio Integer) → g → [Ratio Integer]

randoms ∷ RandomGen g ⇒ g → [Ratio Integer]