Constrained.TheKnot

Description

All the things that are mutually recursive.

Documentation

data SumSpec a b

The Specification for Sums.

Constructors

SumSpecRaw (Maybe String) (Maybe (Int, Int)) (Specification a) (Specification b)

pattern SumSpec ∷ Maybe (Int, Int) → Specification a → Specification b → SumSpec a b

guardSumSpec ∷ ∀ a b. (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ [String] → SumSpec a b → Specification (Sum a b)

combTypeName ∷ Maybe String → Maybe String → Maybe String

type family CountCases a where ...

Equations

CountCases (Sum a b) = 1 + CountCases b

CountCases _ = 1

countCases ∷ ∀ a. KnownNat (CountCases a) ⇒ Int

totalWeight ∷ List (Weighted f) as → Maybe Int

sumType ∷ Maybe String → String

data SumW dom rng where

Constructors

InjLeftW ∷ SumW '[a] (Sum a b)

InjRightW ∷ SumW '[b] (Sum a b)

injLeft_ ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term a → Term (Sum a b)

injRight_ ∷ (HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ Term b → Term (Sum a b)

pattern InjRight ∷ ∀ c. () ⇒ ∀ a b. (c ~ Sum a b, AppRequires SumW '[b] c) ⇒ Term b → Term c

pattern InjLeft ∷ ∀ c. () ⇒ ∀ a b. (c ~ Sum a b, AppRequires SumW '[a] c) ⇒ Term a → Term c

data BoolW (dom ∷ [Type]) (rng ∷ Type) where

Operations on Bool

Constructors

NotW ∷ BoolW '[Bool] Bool

OrW ∷ BoolW '[Bool, Bool] Bool

boolSem ∷ BoolW dom rng → FunTy dom rng

not_ ∷ Term Bool → Term Bool

okOr ∷ Bool → Bool → Specification Bool

We have something like (constant ||. HOLE) must evaluate to need. Return a (Specification Bool) for HOLE, that makes that True.

or_ ∷ Term Bool → Term Bool → Term Bool

data EqW ∷ [Type] → Type → Type where

Constructors

EqualW ∷ (Eq a, HasSpec a) ⇒ EqW '[a, a] Bool

(==.) ∷ HasSpec a ⇒ Term a → Term a → Term Bool infix 4

pattern Equal ∷ ∀ b. () ⇒ ∀ a. (b ~ Bool, Eq a, HasSpec a) ⇒ Term a → Term a → Term b

simplifySpec ∷ HasSpec a ⇒ Specification a → Specification a

ifElse ∷ (IsPred p, IsPred q) ⇒ Term Bool → p → q → Pred

If the `Specification Bool` doesn't constrain the boolean you will get a TrueSpec out.

whenTrue ∷ ∀ p. IsPred p ⇒ Term Bool → p → Pred

pinnedBy ∷ ∀ a. HasSpec a ⇒ Var a → Pred → Maybe (Term a)

Is the variable x pinned to some free term in p? (free term meaning that all the variables in the term are free in p).

TODO: complete this with more cases!

optimisePred ∷ Pred → Pred

aggressiveInlining ∷ Pred → Pred

substituteAndSimplifyTerm ∷ Subst → Term a → Term a

Apply a substitution and simplify the resulting term if the substitution changed the term.

simplifyTerm ∷ ∀ a. Term a → Term a

Simplify a Term, if the Term is an App, apply the rewrite rules chosen by the (Logic sym t bs a) instance attached to the function witness f

simplifyPred ∷ Pred → Pred

simplifyPreds ∷ [Pred] → [Pred]

simplifyBinder ∷ Binder a → Binder a

computeSpecSimplified ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Specification a)

Precondition: the Pred defines the `Var a` Runs in GE in order for us to have detailed context on failure.

computeSpec ∷ ∀ a. (HasSpec a, HasCallStack) ⇒ Var a → Pred → GE (Specification a)

Precondition: the `Pred fn` defines the `Var a`. Runs in GE in order for us to have detailed context on failure.

computeSpecBinder ∷ Binder a → GE (Specification a)

computeSpecBinderSimplified ∷ Binder a → GE (Specification a)

sumWeightL ∷ Maybe (Int, Int) → Doc a

sumWeightR ∷ Maybe (Int, Int) → Doc a

caseSpec ∷ ∀ as. HasSpec (SumOver as) ⇒ Maybe String → List (Weighted Specification) as → Specification (SumOver as)

Turn a list of branches into a SumSpec. If all the branches fail return an ErrorSpec. Note the requirement of HasSpec(SumOver).

genFromSpecT ∷ ∀ a m. (HasCallStack, HasSpec a, MonadGenError m) ⇒ Specification a → GenT m a

Generate a value that satisfies the spec. This function can fail if the spec is inconsistent, there is a dependency error, or if the underlying generators are not flexible enough.

genFromSpec ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Specification a → Gen a

A version of genFromSpecT that simply errors if the generator fails

genFromSpecWithSeed ∷ ∀ a. (HasCallStack, HasSpec a) ⇒ Int → Int → Specification a → a

A version of genFromSpecT that takes a seed and a size and gives you a result

debugSpec ∷ ∀ a. HasSpec a ⇒ Specification a → IO ()

A version of genFromSpecT that runs in the IO monad. Good for debugging.

shrinkWithSpec ∷ ∀ a. HasSpec a ⇒ Specification a → a → [a]

shrinkFromPreds ∷ HasSpec a ⇒ Pred → Var a → a → [a]

shrinkEnvFromPlan ∷ Env → SolverPlan → [Env]

substStage ∷ Env → SolverStage → SolverStage

normalizeSolverStage ∷ SolverStage → SolverStage

fixupWithSpec ∷ ∀ a. HasSpec a ⇒ Specification a → a → Maybe a

computeHints ∷ [Pred] → Hints

prepareLinearization ∷ Pred → GE SolverPlan

Linearize a predicate, turning it into a list of variables to solve and their defining constraints such that each variable can be solved independently.

flattenPred ∷ Pred → [Pred]

Flatten nested Let, Exists, and And in a `Pred fn`. Let and Exists bound variables become free in the result.

linearize ∷ MonadGenError m ⇒ [Pred] → DependGraph → m [SolverStage]

mergeSolverStage ∷ SolverStage → [SolverStage] → [SolverStage]

Does nothing if the variable is not in the plan already.

prettyPlan ∷ HasSpec a ⇒ Specification a → Doc ann

printPlan ∷ HasSpec a ⇒ Specification a → IO ()

isEmptyPlan ∷ SolverPlan → Bool

stepPlan ∷ MonadGenError m ⇒ Env → SolverPlan → GenT m (Env, SolverPlan)

genFromPreds ∷ ∀ m. MonadGenError m ⇒ Env → Pred → GenT m Env

Generate a satisfying Env for a `p : Pred fn`. The Env contains values for all the free variables in `flattenPred p`.

backPropagation ∷ SolverPlan → SolverPlan

Push as much information we can backwards through the plan.

mapSpec ∷ ∀ t a b. AppRequires t '[a] b ⇒ t '[a] b → Specification a → Specification b

Functor like property for Specification, but instead of a Haskell function (a -> b), it takes a function symbol (t '[a] b) from a to b. Note, in this context, a function symbol is some constructor of a witnesstype. Eg. ProdFstW, InjRightW, SingletonW, etc. NOT the lifted versions like fst_ singleton_, which construct Terms. We had to wait until here to define this because it depends on Semigroup property of Specification, and Asserting equality

saturatePred ∷ Pred → [Pred]

pairView ∷ Term (Prod a b) → Maybe (Term a, Term b)

cartesian ∷ ∀ a b. (HasSpec a, HasSpec b) ⇒ Specification a → Specification b → Specification (Prod a b)

data PairSpec a b

Constructors

Cartesian (Specification a) (Specification b)

data ProdW ∷ [Type] → Type → Type where

Constructors

ProdW ∷ (HasSpec a, HasSpec b) ⇒ ProdW '[a, b] (Prod a b)

ProdFstW ∷ (HasSpec a, HasSpec b) ⇒ ProdW '[Prod a b] a

ProdSndW ∷ (HasSpec a, HasSpec b) ⇒ ProdW '[Prod a b] b

prodFst_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term a

prodSnd_ ∷ (HasSpec a, HasSpec b) ⇒ Term (Prod a b) → Term b

sameFst ∷ Eq a1 ⇒ a1 → [Prod a1 a2] → [a2]

sameSnd ∷ Eq a1 ⇒ a1 → [Prod a2 a1] → [a2]

prod_ ∷ (HasSpec a, HasSpec b) ⇒ Term a → Term b → Term (Prod a b)

pattern Product ∷ ∀ c. () ⇒ ∀ a b. (c ~ Prod a b, AppRequires ProdW '[a, b] (Prod a b)) ⇒ Term a → Term b → Term c

data ElemW ∷ [Type] → Type → Type where

Constructors

ElemW ∷ HasSpec a ⇒ ElemW '[a, [a]] Bool

elem_ ∷ (Sized [a], HasSpec a) ⇒ Term a → Term [a] → Term Bool infix 4

elemFn ∷ HasSpec a ⇒ Fun '[a, [a]] Bool

pattern Elem ∷ ∀ b. () ⇒ ∀ a. (b ~ Bool, Eq a, HasSpec a) ⇒ Term a → Term [a] → Term b

data ListSpec a

Constructors

ListSpec

Fields

listSpecHint ∷ Maybe Integer

listSpecMust ∷ [a]

listSpecSize ∷ Specification Integer

listSpecElem ∷ Specification a

listSpecFold ∷ FoldSpec a

guardListSpec ∷ HasSpec a ⇒ [String] → ListSpec a → Specification [a]

sizeOf_ ∷ (HasSpec a, Sized a) ⇒ Term a → Term Integer

genFromSizeSpec ∷ MonadGenError m ⇒ Specification Integer → GenT m Integer

Because Sizes should always be >= 0, We provide this alternate generator that can be used to replace (genFromSpecT @Integer), to ensure this important property

data ListW (args ∷ [Type]) (res ∷ Type) where

Constructors

FoldMapW ∷ ∀ a b. (Foldy b, HasSpec a) ⇒ Fun '[a] b → ListW '[[a]] b

SingletonListW ∷ HasSpec a ⇒ ListW '[a] [a]

AppendW ∷ (HasSpec a, Typeable a, Show a) ⇒ ListW '[[a], [a]] [a]

listSem ∷ ListW dom rng → FunTy dom rng

foldMap_ ∷ ∀ a b. (Foldy b, HasSpec a) ⇒ (Term a → Term b) → Term [a] → Term b

sum_ ∷ Foldy a ⇒ Term [a] → Term a

singletonList_ ∷ (Sized [a], HasSpec a) ⇒ Term a → Term [a]

append_ ∷ (Sized [a], HasSpec a) ⇒ Term [a] → Term [a] → Term [a]

appendFn ∷ ∀ a. (Sized [a], HasSpec a) ⇒ Fun '[[a], [a]] [a]

singletonListFn ∷ ∀ a. HasSpec a ⇒ Fun '[a] [a]

foldMapFn ∷ ∀ a b. (HasSpec a, Foldy b) ⇒ Fun '[a] b → Fun '[[a]] b

reverseFoldSpec ∷ FoldSpec a → Specification a

prefixedBy ∷ Eq a ⇒ [a] → [[a]] → [[a]]

suffixedBy ∷ Eq a ⇒ [a] → [[a]] → [[a]]

alreadyHave ∷ Eq a ⇒ [a] → ListSpec a → ListSpec a

alreadyHaveFold ∷ [a] → FoldSpec a → FoldSpec a

toPredsFoldSpec ∷ HasSpec a ⇒ Term [a] → FoldSpec a → Pred

Used in the HasSpec [a] instance

data FoldSpec a where

Constructors

NoFold ∷ FoldSpec a

FoldSpec ∷ ∀ b a. (HasSpec a, HasSpec b, Foldy b) ⇒ Fun '[a] b → Specification b → FoldSpec a

preMapFoldSpec ∷ HasSpec a ⇒ Fun '[a] b → FoldSpec b → FoldSpec a

composeFn ∷ (HasSpec b, HasSpec c) ⇒ Fun '[b] c → Fun '[a] b → Fun '[a] c

idFn ∷ HasSpec a ⇒ Fun '[a] a

combineFoldSpec ∷ FoldSpec a → FoldSpec a → Either [String] (FoldSpec a)

conformsToFoldSpec ∷ ∀ a. [a] → FoldSpec a → Bool

class (HasSpec a, NumLike a, Logic IntW) ⇒ Foldy a where

Minimal complete definition

genList, genSizedList, noNegativeValues

Methods

genList ∷ MonadGenError m ⇒ Specification a → Specification a → GenT m [a]

theAddFn ∷ IntW '[a, a] a

theZero ∷ a

genSizedList ∷ MonadGenError m ⇒ Specification Integer → Specification a → Specification a → GenT m [a]

noNegativeValues ∷ Bool

type SizeSpec = NumSpec Integer

class Sized t where

Minimal complete definition

Nothing

Methods

sizeOf ∷ t → Integer

default sizeOf ∷ (HasSimpleRep t, Sized (SimpleRep t)) ⇒ t → Integer

liftSizeSpec ∷ HasSpec t ⇒ SizeSpec → [Integer] → Specification t

default liftSizeSpec ∷ (Sized (SimpleRep t), GenericRequires t) ⇒ SizeSpec → [Integer] → Specification t

liftMemberSpec ∷ HasSpec t ⇒ [Integer] → Specification t

default liftMemberSpec ∷ (Sized (SimpleRep t), GenericRequires t) ⇒ [Integer] → Specification t

sizeOfTypeSpec ∷ HasSpec t ⇒ TypeSpec t → Specification Integer

default sizeOfTypeSpec ∷ (HasSpec (SimpleRep t), Sized (SimpleRep t), TypeSpec t ~ TypeSpec (SimpleRep t)) ⇒ TypeSpec t → Specification Integer

adds ∷ Foldy a ⇒ [a] → a

data FunW (dom ∷ [Type]) (rng ∷ Type) where

Constructors

IdW ∷ ∀ a. FunW '[a] a

ComposeW ∷ ∀ b t1 t2 a r. (AppRequires t1 '[b] r, AppRequires t2 '[a] b, HasSpec b) ⇒ t1 '[b] r → t2 '[a] b → FunW '[a] r

FlipW ∷ ∀ t a b r. AppRequires t '[a, b] r ⇒ t '[a, b] r → FunW '[b, a] r

funSem ∷ FunW dom rng → FunTy dom rng

compareWit ∷ ∀ t1 bs1 r1 t2 bs2 r2. (AppRequires t1 bs1 r1, AppRequires t2 bs2 r2) ⇒ t1 bs1 r1 → t2 bs2 r2 → Bool

id_ ∷ ∀ a. HasSpec a ⇒ Term a → Term a

flip_ ∷ ∀ (t ∷ [Type] → Type → Type) a b r. (HasSpec b, HasSpec a, AppRequires t '[a, b] r) ⇒ t '[a, b] r → Term b → Term a → Term r

compose_ ∷ ∀ b t1 t2 a r. (AppRequires t1 '[b] r, AppRequires t2 '[a] b) ⇒ t1 '[b] r → t2 '[a] b → Term a → Term r

genInverse ∷ (MonadGenError m, HasSpec a, HasSpec b) ⇒ Fun '[a] b → Specification a → b → GenT m a

genFromFold ∷ ∀ m a b. (MonadGenError m, Foldy b, HasSpec a) ⇒ [a] → Specification Integer → Specification a → Fun '[a] b → Specification b → GenT m [a]

addFun ∷ NumLike n ⇒ Fun '[n, n] n

data SizeW (dom ∷ [Type]) rng ∷ Type where

Constructors

SizeOfW ∷ (Sized n, HasSpec n) ⇒ SizeW '[n] Integer

sizeOfFn ∷ ∀ a. (HasSpec a, Sized a) ⇒ Fun '[a] Integer

rangeSize ∷ Integer → Integer → SizeSpec

between ∷ (HasSpec a, TypeSpec a ~ NumSpec a) ⇒ a → a → Specification a

maxSpec ∷ Specification Integer → Specification Integer

The widest interval whose largest element is admitted by the original spec

hasSize ∷ (HasSpec t, Sized t) ⇒ SizeSpec → Specification t

Orphan instances

Numeric a ⇒ Complete a

Instance details

Methods

simplifyA ∷ Specification a → Specification a

genFromSpecA ∷ ∀ (m ∷ Type → Type). (HasCallStack, HasSpec a, MonadGenError m) ⇒ Specification a → GenT m a

theAddA ∷ IntW '[a, a] a

(HasSpec a, Arbitrary (TypeSpec a)) ⇒ Arbitrary (Specification a)

Instance details

Methods

arbitrary ∷ Gen (Specification a)

shrink ∷ Specification a → [Specification a]

(Sized [a], HasSpec a) ⇒ HasGenHint [a]

Instance details

Associated Types

type Hint [a]

Methods

giveHint ∷ Hint [a] → Specification [a]

(Sized [a], HasSpec a) ⇒ HasSpec [a]

Instance details

Associated Types

type TypeSpec [a]

type Prerequisites [a]

Methods

emptySpec ∷ TypeSpec [a]

combineSpec ∷ TypeSpec [a] → TypeSpec [a] → Specification [a]

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec [a] → GenT m [a]

conformsTo ∷ [a] → TypeSpec [a] → Bool

shrinkWithTypeSpec ∷ TypeSpec [a] → [a] → [[a]]

toPreds ∷ Term [a] → TypeSpec [a] → Pred

cardinalTypeSpec ∷ TypeSpec [a] → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec [a] → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec [a] → BinaryShow

monadConformsTo ∷ [a] → TypeSpec [a] → Writer [String] Bool

typeSpecOpt ∷ TypeSpec [a] → [[a]] → Specification [a]

guardTypeSpec ∷ [String] → TypeSpec [a] → Specification [a]

prerequisites ∷ Evidence (Prerequisites [a])

Forallable [a] a

Instance details

Methods

fromForAllSpec ∷ Specification a → Specification [a]

forAllToList ∷ [a] → [a]

(HasSpec a, HasSpec b) ⇒ HasSpec (Prod a b)

Instance details

Associated Types

type TypeSpec (Prod a b)

type Prerequisites (Prod a b)

Methods

emptySpec ∷ TypeSpec (Prod a b)

combineSpec ∷ TypeSpec (Prod a b) → TypeSpec (Prod a b) → Specification (Prod a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Prod a b) → GenT m (Prod a b)

conformsTo ∷ Prod a b → TypeSpec (Prod a b) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Prod a b) → Prod a b → [Prod a b]

toPreds ∷ Term (Prod a b) → TypeSpec (Prod a b) → Pred

cardinalTypeSpec ∷ TypeSpec (Prod a b) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Prod a b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Prod a b) → BinaryShow

monadConformsTo ∷ Prod a b → TypeSpec (Prod a b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Prod a b) → [Prod a b] → Specification (Prod a b)

guardTypeSpec ∷ [String] → TypeSpec (Prod a b) → Specification (Prod a b)

prerequisites ∷ Evidence (Prerequisites (Prod a b))

(HasSpec a, HasSpec b, KnownNat (CountCases b)) ⇒ HasSpec (Sum a b)
The HasSpec Sum instance

Instance details

Associated Types

type TypeSpec (Sum a b)

type Prerequisites (Sum a b)

Methods

emptySpec ∷ TypeSpec (Sum a b)

combineSpec ∷ TypeSpec (Sum a b) → TypeSpec (Sum a b) → Specification (Sum a b)

genFromTypeSpec ∷ ∀ (m ∷ Type → Type). (HasCallStack, MonadGenError m) ⇒ TypeSpec (Sum a b) → GenT m (Sum a b)

conformsTo ∷ Sum a b → TypeSpec (Sum a b) → Bool

shrinkWithTypeSpec ∷ TypeSpec (Sum a b) → Sum a b → [Sum a b]

toPreds ∷ Term (Sum a b) → TypeSpec (Sum a b) → Pred

cardinalTypeSpec ∷ TypeSpec (Sum a b) → Specification Integer

cardinalTrueSpec ∷ Specification Integer

typeSpecHasError ∷ TypeSpec (Sum a b) → Maybe (NonEmpty String)

alternateShow ∷ TypeSpec (Sum a b) → BinaryShow

monadConformsTo ∷ Sum a b → TypeSpec (Sum a b) → Writer [String] Bool

typeSpecOpt ∷ TypeSpec (Sum a b) → [Sum a b] → Specification (Sum a b)

guardTypeSpec ∷ [String] → TypeSpec (Sum a b) → Specification (Sum a b)

prerequisites ∷ Evidence (Prerequisites (Sum a b))